% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{addBlockConnectivity}
\alias{addBlockConnectivity}
\title{addBlockConnectivity}
\usage{
addBlockConnectivity(object, connectivity, zlocation = NULL, sample)
}
\arguments{
\item{object}{a VoltRon object}

\item{connectivity}{a metadata of edges representing connected layers within a block}

\item{zlocation}{zlocation}

\item{sample}{sample name}
}
\description{
add connectivity information to the layers (vrLayer) of the same block (Block)
}
