% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{addFOVImage}
\alias{addFOVImage}
\title{}
\usage{
addFOVImage(seu, file, fov = "fov", overwrite = FALSE)
}
\arguments{
\item{seu}{Seurat Object with Xenium Data}

\item{file}{the morphology image file created by \code{GenerateXeniumDAPIImage}.}

\item{fov}{FOV name, preferably the name used in the Seurat Object}

\item{overwrite}{Overwrite the existing FOV image}
}
\description{
# check existing images in the spacerover object
  assay <- MainAssay(sr)
}
\details{

}
