% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(object, assay = NULL, type = NULL, value, label)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{type}{the assay type: ROI, spot or cell, or all for the entire metadata object}

\item{value}{the new values of the metadata column}

\item{label}{the label of the new column, either a new column or an existing one}
}
\description{
adding new columns or updating the values of the existing columns
}
