% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{addSpatialLayer}
\alias{addSpatialLayer}
\title{addSpatialLayer}
\usage{
addSpatialLayer(
  g,
  object,
  assay,
  group.by = "Sample",
  plot.segments = FALSE,
  group.ids = NULL,
  reg = FALSE,
  colors = NULL,
  alpha = 1,
  n.tile = NULL,
  pt.size = 2,
  cell.shape = 21,
  graph = NULL,
  graph.edge.color = "orange",
  spatial = NULL,
  combine.groups = FALSE
)
}
\arguments{
\item{g}{ggplot object}

\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{plot.segments}{plot segments from \link{vrSegments} instead of points}

\item{group.ids}{a subset of categories defined in metadata column from \code{group.by}}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainSpatial}) is requested}

\item{colors}{the color set for group.by. Should be of the same size of group.id (if specified) or unique elements in group.by}

\item{alpha}{alpha level of colors of visualized points and segments}

\item{n.tile}{The number of tiles on x-and y-axis for rasterizing points (see \link{geom_tile}). The rasterization is performed automatically for large number of points 
Only applicable to spots, cells and molecules. If \code{n.tile = 0} will turn of automated rasterization.}

\item{pt.size}{point size}

\item{cell.shape}{the shape of the points representing cells, see \link{geom_point}}

\item{graph}{if not NULL, the graph is added to the plot}

\item{graph.edge.color}{the color of graph edges, if \code{graph} is not NULL.}

\item{spatial}{the name of the spatial coordinate system}

\item{combine.groups}{if TRUE, tile colors will reflect relative abundance of either of two groups, strictly for visualizing two groups when assay is a molecule typed and tiled (see \code{n.tile}).}
}
\description{
adding additional layers of spatial plots to an existing \link{vrSpatialPlot}.
}
