% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{AlignmentTabPanel}
\alias{AlignmentTabPanel}
\title{RegisteredImageTabPanels}
\usage{
AlignmentTabPanel(len_images, centre, register_ind)
}
\arguments{
\item{len_images}{the number of query images}

\item{centre}{center image index}

\item{register_ind}{query image indices}
}
\value{
tabsetpanel
}
\description{
The UI for a set of query spatial slides
}
