% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{applyPerspectiveTransform}
\alias{applyPerspectiveTransform}
\title{applyPerspectiveTransform}
\usage{
applyPerspectiveTransform(
  object,
  mapping,
  reference_image,
  input,
  reg_mode,
  ref_extension,
  query_extension
)
}
\arguments{
\item{object}{A VoltRon objects}

\item{mapping}{a list of transformation matrices}

\item{reference_image}{the reference image}

\item{input}{input}

\item{reg_mode}{the registration mode, either "auto" or "manual"}

\item{ref_extension}{the shiny extension of reference image}

\item{query_extension}{the shiny extension of query image}
}
\description{
Applying a perspective transformation to the VoltRon object
}
