% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{applyTPSTransform}
\alias{applyTPSTransform}
\title{applyTPSTransform}
\usage{
applyTPSTransform(x, trafo)
}
\arguments{
\item{x}{matrix}

\item{trafo}{4x4 transformation matrix or an object of class "tpsCoeff"}
}
\description{
apply affine transformation to data, adapted from \code{Morpho} package.
}
