% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.AnnData}
\alias{as.AnnData}
\title{as.AnnData}
\usage{
as.AnnData(
  object,
  file,
  assay = NULL,
  flip_coordinates = FALSE,
  method = "anndataR",
  create.ometiff = FALSE,
  python.path = NULL,
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{file}{the name of the h5ad file.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium),
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{flip_coordinates}{if TRUE, the spatial coordinates
(including segments) will be flipped.}

\item{method}{the package to use for conversion: "anndataR" or "anndata".}

\item{create.ometiff}{should an ometiff file be generated of default
image of the object}

\item{python.path}{the path to the python binary, otherwise either
\code{basilisk} package is used or \code{getOption("voltron.python.path")}
should be not NULL.}

\item{...}{additional parameters passed to \link{vrImages}.}
}
\description{
Converting a VoltRon object into a AnnData (.h5ad) object
}
\details{
This function converts a VoltRon object into an AnnData object (.h5ad file).
It extracts assay data, spatial coordinates, and optionally flips
coordinates. Images associated with the assay can be included in the
resulting AnnData file, with additional customization parameters like
channel, scale.perc.
}
