% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.Giotto}
\alias{as.Giotto}
\alias{as.Giotto.VoltRon}
\title{as.Giotto}
\usage{
as.Giotto(object, ...)

\method{as.Giotto}{VoltRon}(object)
}
\arguments{
\item{object}{A VoltRon object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{as.Giotto}: The name of the default assay
}
\description{
Generic methods for conversion into a Giotto object
}
