% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.OmeTiff}
\alias{as.OmeTiff}
\title{as.OmeTiff}
\usage{
as.OmeTiff(object, out_path, image_id = "image_1")
}
\arguments{
\item{object}{a magick-image object}

\item{out_path}{output path to ome.tiff file}

\item{image_id}{image name}
}
\description{
Converting VoltRon (magick) images to ome.tiff
}
