% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.Seurat}
\alias{as.Seurat}
\alias{as.Seurat.VoltRon}
\title{as.Seurat}
\usage{
as.Seurat(object, ...)

\method{as.Seurat}{VoltRon}(
  object,
  cell.assay = NULL,
  molecule.assay = NULL,
  image_key = "fov",
  type = c("image", "spatial"),
  reg = FALSE
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{cell.assay}{the name(type) of the cell assay to be converted}

\item{molecule.assay}{the name(type) of the molecule assay to be added to the cell assay in Seurat object}

\item{image_key}{the name (or prefix) of the image(s)}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}

\item{reg}{if TRUE, registered coordinates will be used}
}
\description{
Generic methods for conversion into a Seurat object
}
