% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.SpatialData}
\alias{as.SpatialData}
\title{as.SpatialData}
\usage{
as.SpatialData(
  object,
  file,
  assay = NULL,
  type = c("image", "spatial"),
  flip_coordinates = FALSE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{file}{the name of the h5ad file}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}

\item{flip_coordinates}{if TRUE, the spatial coordinates (including segments) will be flipped}
}
\description{
Converting a VoltRon object into a SpatialData (.zarr) object
}
