% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.SpatialExperiment}
\alias{as.SpatialExperiment}
\title{as.SpatialExperiment}
\usage{
as.SpatialExperiment(object, assay = NULL, reg = FALSE)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{the name of the assay to be converted}

\item{reg}{if TRUE, registered coordinates will be used}
}
\description{
Converting a VoltRon object into a SpatialExperiment object
}
