% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{AssayNames.srAssay}
\alias{AssayNames.srAssay}
\alias{AssayNames<-.srAssay}
\alias{AssayNames}
\alias{AssayNames<-}
\alias{AssayNames.SpaceRover}
\title{Get Assay names}
\usage{
\method{AssayNames}{srAssay}(object, ...)

\method{AssayNames}{srAssay}(object, ...) <- value

AssayNames(object, ...)

AssayNames(object, ...) <- value

\method{AssayNames}{SpaceRover}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{AssayNames}: The name of the default assay

\code{AssayNames<-}: The coordinates updated
}
\description{
Given a SpaceRover object, give names of assays of some type, or main assay.
}
\concept{data-access}
