% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{automated_registration_rcpp}
\alias{automated_registration_rcpp}
\title{automated_registration}
\usage{
automated_registration_rcpp(
  ref_image,
  query_image,
  GOOD_MATCH_PERCENT = 0.15,
  MAX_FEATURES = 500
)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{GOOD_MATCH_PERCENT}{the percentage of good matching keypoints}

\item{MAX_FEATURES}{maximum number of detected features, i.e. keypoints}
}
\description{
automated registration with Rcpp/C++
}
