% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{changeAssayNames.vrSample}
\alias{changeAssayNames.vrSample}
\alias{changeAssayNames.vrBlock}
\title{changeAssayNames.vrSample}
\usage{
\method{changeAssayNames}{vrSample}(object, sample.metadata = NULL)

\method{changeAssayNames}{vrBlock}(object, ...)
}
\arguments{
\item{sample.metadata}{the sample metadata with NewAssayNames column which includes the new assay names}
}
\description{
Change the assay names of assays within a vrSample object

Change the assay names of assays within a vrBlock object
}
