% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metadata.R, R/objects.R
\name{changeSampleNames}
\alias{changeSampleNames}
\alias{changeSampleNames.vrMetadata}
\alias{changeSampleNames.VoltRon}
\title{changeSampleNames}
\usage{
changeSampleNames(object, ...)

\method{changeSampleNames}{vrMetadata}(object, sample_metadata_table)

\method{changeSampleNames}{VoltRon}(object, samples = NULL)
}
\arguments{
\item{object}{a VoltRon object}

\item{...}{Arguments passed to other methods}

\item{sample_metadata_table}{the sample metadata with old and new layers and samples passed from \code{changeSampleNames.VoltRon}}

\item{samples}{a single or a set of sample names}
}
\description{
change sample names of VoltRon or other objects

Change the sample names of the vrMetadata object and reorient layers if needed
This functions requires the new and old sample and layer names passed from \code{changeSampleNames.VoltRon}

Change the sample names of the VoltRon object and reorient layers if needed
}
