% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{combineChannels}
\alias{combineChannels}
\alias{combineChannels,VoltRon-method}
\alias{combineChannels,vrAssay-method}
\alias{combineChannels,vrAssayV2-method}
\alias{combineChannels,vrImage-method}
\alias{combineChannels,vrSpatial-method}
\title{combineChannels}
\usage{
combineChannels(object, ...)

\S4method{combineChannels}{VoltRon}(
  object,
  assay = NULL,
  spatial = NULL,
  reg = FALSE,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)

\S4method{combineChannels}{vrAssay}(
  object,
  spatial = NULL,
  reg = FALSE,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)

\S4method{combineChannels}{vrAssayV2}(
  object,
  spatial = NULL,
  reg = FALSE,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)

\S4method{combineChannels}{vrImage}(
  object,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)

\S4method{combineChannels}{vrSpatial}(
  object,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{spatial}{the name of the spatial coordinate system}

\item{reg}{TRUE if registered coordinates of the main image 
(\link{vrMainSpatial}) is requested}

\item{channels}{the name of the channel associated with the image}

\item{colors}{the colors associated with each channel}

\item{channel_key}{the name of the new channel name}
}
\description{
Combining channels into novel channels of the same image
}
