% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{combineGraphs}
\alias{combineGraphs}
\title{combineGraphs}
\usage{
combineGraphs(
  object,
  graph.names = NULL,
  graph.weights = NULL,
  graph.key = "combined"
)
}
\arguments{
\item{object}{a VoltRon Object}

\item{graph.names}{a vector of graph names}

\item{graph.weights}{the weights for edges of each graph.}

\item{graph.key}{the name of the combined graph}
}
\description{
Combining the edges of multiple graphs
}
