% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{computeAutomatedPairwiseTransform}
\alias{computeAutomatedPairwiseTransform}
\title{computeAutomatedPairwiseTransform}
\usage{
computeAutomatedPairwiseTransform(image_list, query_ind, ref_ind, input)
}
\arguments{
\item{image_list}{the list of images}

\item{query_ind}{the index of the query image}

\item{ref_ind}{the index of the reference image}

\item{input}{input}
}
\description{
Computing the registration matrix necessary for automated registration
}
