% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{computeTPSTransform}
\alias{computeTPSTransform}
\title{computeTPSTransform}
\usage{
computeTPSTransform(x, y)
}
\arguments{
\item{x}{fix landmarks from a k x m matrix}

\item{y}{moving landmarks from a k x m matrix}
}
\description{
calculate an affine transformation matrix, adapted from \code{Morpho} package.
}
