% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertAnnDataToVoltRon}
\alias{convertAnnDataToVoltRon}
\title{convertAnnDataToVoltRon}
\usage{
convertAnnDataToVoltRon(file, AssayID = NULL, ...)
}
\arguments{
\item{file}{h5ad file}

\item{AssayID}{the ID assays in the h5ad file}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
converting AnnData h5ad files to VoltRon objects
}
