% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R
\name{Coordinates.SpaceRover}
\alias{Coordinates.SpaceRover}
\alias{Coordinates}
\title{Coordinates}
\usage{
\method{Coordinates}{SpaceRover}(sr, ...)

Coordinates(object, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{object}{An object}
}
\value{
\code{Coordinates}: The name of the default assay
}
\description{
Get and set the coordinates of the main assay
}
\concept{data-access}
