% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{CreateL}
\alias{CreateL}
\title{Create Matrices necessary for Thin-Plate Spline}
\usage{
CreateL(matrix, lambda = 1e-08)
}
\arguments{
\item{matrix}{k x 3 or k x 2 matrix containing landmark coordinates.}

\item{lambda}{numeric: regularization factor}
}
\value{
{Matrix L as specified in Bookstein (1989)}
}
\description{
Create (Bending Engergy) Matrices necessary for Thin-Plate Spline, and sliding of Semilandmarks, adapted from \code{Morpho} package.
}
\note{
This function is not intended to be called directly - except for playing around to grasp the mechansims of the Thin-Plate Spline.
}
\references{
Gunz, P., P. Mitteroecker, and F. L. Bookstein. 2005. Semilandmarks in Three Dimensions, in Modern Morphometrics in Physical
Anthropology. Edited by D. E. Slice, pp. 73-98. New York: Kluwer
Academic/Plenum Publishers.

Bookstein FL. 1989. Principal Warps: Thin-plate splines and the
decomposition of deformations. IEEE Transactions on pattern analysis and
machine intelligence 11(6).
}
