% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSpaceRover}
\alias{CreateSpaceRover}
\title{CreateSpaceRover}
\usage{
CreateSpaceRover(
  data,
  metadata = NULL,
  image = NULL,
  coords,
  segments = NULL,
  sample.metadata = NULL,
  zstack = NULL,
  main.assay = "Custom_cell",
  assay.type = "cell",
  sample_name = NULL,
  layer_name = NULL,
  project = NULL,
  ...
)
}
\arguments{
\item{data}{the count table}

\item{metadata}{a metadata object of class \code{srMetadata}}

\item{image}{the image of the data}

\item{segments}{the segments of the spatial entities, optional}

\item{sample.metadata}{a data frame of the sample metadata}

\item{zstack}{the zstack graph to determine the adjacency of spatial entities across layers}

\item{main.assay}{the name of the main assay of the object}

\item{assay.type}{the type of the assay (cells, spots, ROIs)}

\item{sample_name}{the name of the sample}

\item{layer_name}{the name of the layer}

\item{project}{project name}

\item{coord}{the coordinates of the spatial entities}

\item{assay_name}{the name of the assay}
}
\description{
Create a SpaceRover object
}
