% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Embeddings.srAssay}
\alias{Embeddings.srAssay}
\alias{Embeddings<-.srAssay}
\alias{Embeddings}
\alias{Embeddings<-}
\alias{Embeddings.SpaceRover}
\alias{Embeddings<-.SpaceRover}
\title{Embeddings}
\usage{
\method{Embeddings}{srAssay}(object, type = "pca")

\method{Embeddings}{srAssay}(object, type = "pca", ...) <- value

Embeddings(object, ...)

Embeddings(object, ...) <- value

\method{Embeddings}{SpaceRover}(object, assay = NULL, type = "pca", ..., value)

\method{Embeddings}{SpaceRover}(object, assay = NULL, type = "pca", ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Embeddings}: The name of the default assay

\code{Embeddings}: The name of the default assay
}
\description{
Get embeddings of spatial points
}
\concept{data-access}
