% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/metadata.R,
%   R/objects.R, R/sample.R
\name{Entities.srAssay}
\alias{Entities.srAssay}
\alias{Entities}
\alias{Entities.srMetadata}
\alias{Entities.SpaceRover}
\alias{Entities.srSample}
\alias{Entities.srLayer}
\title{Entities}
\usage{
\method{Entities}{srAssay}(object, ...)

Entities(object, ...)

\method{Entities}{srMetadata}(object, ...)

\method{Entities}{SpaceRover}(object, ...)

\method{Entities}{srSample}(object, ...)

\method{Entities}{srLayer}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Entities}: The name of the default assay
}
\description{
Get and set spatial entities (cells, spots, ROI)
}
\concept{data-access}
