% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Features.vrAssay}
\alias{Features.vrAssay}
\alias{Features}
\alias{Features.VoltRon}
\title{Features}
\usage{
\method{Features}{vrAssay}(object, ...)

Features(object, ...)

\method{Features}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Features}: The name of the default assay
}
\description{
Get features from the main.assay
}
\concept{data-access}
