% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/clustering.R
\name{findNeighbors}
\alias{findNeighbors}
\alias{findNeighbors.SpaceRover}
\alias{findNeighbors.srAssay}
\title{Find neighbors}
\usage{
findNeighbors(object, ...)

\method{findNeighbors}{SpaceRover}(object, assay = NULL, assay.type = NULL, ...)

\method{findNeighbors}{srAssay}(object, assay = NULL, assay.type = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
find neighbors in an assay
}
\concept{clustering}
