% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{flipCoords.vrAssay}
\alias{flipCoords.vrAssay}
\alias{flipCoords}
\alias{flipCoords.VoltRon}
\title{flipCoords}
\usage{
\method{flipCoords}{vrAssay}(object, ...)

flipCoords(object, ...)

\method{flipCoords}{VoltRon}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{flipCoords}: The name of the default assay
}
\description{
Get and set the coordinates of the main assay
}
