% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{flipflop_keypoint}
\alias{flipflop_keypoint}
\title{Flip and Flop transformations for Keypoints}
\usage{
flipflop_keypoint(keypoints, image_limits, flipflop)
}
\arguments{
\item{keypoints}{dataset of keypoints}

\item{image_limits}{limits of the images}

\item{flipflop}{a flip or flop action as string}
}
\value{
keypoints
}
\description{
find transformed keypoints on image given any flip or flop action by magick
}
