% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{flipflopKeypoint}
\alias{flipflopKeypoint}
\title{flipflopKeypoint}
\usage{
flipflopKeypoint(keypoints, image_limits, flipflop)
}
\arguments{
\item{keypoints}{}

\item{image_limits}{limits of the images}

\item{flipflop}{a flip or flop action as string}
}
\description{
Find transformed keypoints on image given any flip or flop action by magick
}
