% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{FOVImage-class}
\alias{FOVImage-class}
\alias{FOVImage}
\title{The FOVImage class}
\description{
The FOVImage stores the morphology image from the Xenium assay
}
\section{Slots}{

\describe{
\item{\code{image}}{A magick-image class object from magick package}
}}

