% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{FromSegmentToCrop}
\alias{FromSegmentToCrop}
\title{FromSegmentToCrop}
\usage{
FromSegmentToCrop(segment, imageinfo)
}
\arguments{
\item{segment}{coordinates of a segment}

\item{imageinfo}{info of the image}
}
\description{
get magick crop information from coordinates of a segment
}
