% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateCosMxImage}
\alias{generateCosMxImage}
\title{generateCosMxImage}
\usage{
generateCosMxImage(
  dir.path,
  fov.position.file,
  increase.contrast = FALSE,
  output.path = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{CosMx folder of images}

\item{fov.position.file}{the file providing FOV positions}

\item{increase.contrast}{increase the contrast of the image before writing}

\item{output.path}{The path to the new morphology image created if the image should be saved to a location other than Xenium output folder.}

\item{verbose}{verbose}

\item{...}{additional parameters passed to the \link[EBImage]{writeImage} 
function}
}
\description{
Generates a low resolution Morphology image of the CosMx experiment
}
