% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateGeoJSON}
\alias{generateGeoJSON}
\title{generateGeoJSON}
\usage{
generateGeoJSON(segments, file)
}
\arguments{
\item{segments}{the segments, typically from \link{vrSegments}.}

\item{file}{the GeoJSON file, typically to be used by QuPath software.}
}
\description{
generating geojson files from segments
}
