% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateGeoJSONFromSegments}
\alias{generateGeoJSONFromSegments}
\title{generateSegmentsFromGeoJSON}
\usage{
generateGeoJSONFromSegments(segments, geojson.file)
}
\arguments{
\item{segments}{the segments, typically from \link{vrSegments}.}

\item{geojson.file}{the GeoJSON file, typically to be used by QuPath software.}
}
\description{
The function to import segments from a json data
}
