% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateSegments}
\alias{generateSegments}
\title{generateSegments}
\usage{
generateSegments(object, type = c("ROI", "cell"))
}
\arguments{
\item{geojson.file}{the GeoJSON file, typically generated by QuPath software}
}
\description{
The function to import segments from a geojson file
}
