% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateSegmentsFromGeoJSON}
\alias{generateSegmentsFromGeoJSON}
\title{generateSegmentsFromGeoJSON}
\usage{
generateSegmentsFromGeoJSON(geojson.file)
}
\arguments{
\item{geojson.file}{the GeoJSON file, typically generated by QuPath software}
}
\description{
The function to import segments from a json data
}
