% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{generateTileData}
\alias{generateTileData}
\alias{generateTileData,VoltRon-method}
\alias{generateTileData,vrAssay-method}
\alias{generateTileData,vrAssayV2-method}
\title{generateTileData}
\usage{
generateTileData(object, ...)

\S4method{generateTileData}{VoltRon}(object, assay = NULL, ...)

\S4method{generateTileData}{vrAssay}(object, spatial = NULL, reg = FALSE, channel = NULL)

\S4method{generateTileData}{vrAssayV2}(object, spatial = NULL, reg = FALSE, channel = NULL)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{additional parameters passed to vrAssay.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{spatial}{the name of the spatial coordinate system}

\item{reg}{TRUE if registered coordinates of the main
image (\link{vrMainSpatial}) is requested}

\item{channel}{the name of the channel associated with the image}
}
\description{
Generating data matrices for tile-based VoltRon objects from images
}
