% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_manipulation.R
\name{GenerateXeniumDAPIImage}
\alias{GenerateXeniumDAPIImage}
\title{GenerateXeniumDAPIImage}
\usage{
GenerateXeniumDAPIImage(
  dir.path,
  increase.contrast = TRUE,
  resolutionLevel = 7,
  output.path = NULL,
  ...
)
}
\arguments{
\item{dir.path}{Xenium output folder}

\item{increase.contrast}{increase the contrast of the image before writing}

\item{resolutionLevel}{the level of resolution within Xenium OME-TIFF image. Default: 7 (553x402)}

\item{output.path}{The path to the new morphology image created if the image should be saved to a location other than Xenium output folder.}

\item{...}{additional parameters passed to the EBImage::writeImage function}
}
\description{
Generates a low resolution DAPI image from
}
