% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateXeniumImage}
\alias{generateXeniumImage}
\title{generateXeniumImage}
\usage{
generateXeniumImage(
  dir.path,
  increase.contrast = TRUE,
  resolution_level = 7,
  overwrite_resolution = FALSE,
  output.path = NULL,
  file.name = "morphology_lowres.tif",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{Xenium output folder}

\item{increase.contrast}{increase the contrast of the image before writing}

\item{resolution_level}{the level of resolution within Xenium OME-TIFF
image. Default: 7 (553x402)}

\item{overwrite_resolution}{if TRUE, the image "file.name" will be generated
again although it exists at "dir.path"}

\item{output.path}{The path to the new morphology image created if the
image should be saved to a location other than Xenium output folder.}

\item{file.name}{the name of the lowred morphology image.
Default: morphology_lowres.tif}

\item{verbose}{verbose}

\item{...}{additional parameters passed to the \link[EBImage]{writeImage} 
function}
}
\description{
Generate a low resolution DAPI image of the Xenium experiment
}
\details{
The Xenium morphology_mip.ome.tif file that is found under the outs folder
comes is an hyperstack of different resolutions of the DAPI image.
\link{generateXeniumImage} allows extracting only one of these layers by
specifying the \code{resolution} parameter (Default: 7 for 553x402)
among 1 to 8. Lower incides of resolutions have higher higher resolutions,
e.g. 1 for 35416x25778. Note that you may need to allocate larger memory of
Java to import higher resolution images.
}
