% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getAlignmentTabPanel}
\alias{getAlignmentTabPanel}
\title{getRegisteredImageTabPanels}
\usage{
getAlignmentTabPanel(len_images, centre, register_ind)
}
\arguments{
\item{len_images}{the number of query images}

\item{centre}{center image index}

\item{register_ind}{query image indices}
}
\description{
The UI for a set of query spatial slides
}
