% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{getBasilisk}
\alias{getBasilisk}
\title{get the Python Basilisk environment}
\usage{
getBasilisk()
}
\description{
Defines a conda environment via Basilisk, which is used to convert R objects to Zarr stores.
}
