% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getConnectedSpatialPoints}
\alias{getConnectedSpatialPoints}
\title{getConnectedSpatialPoints}
\usage{
getConnectedSpatialPoints(object, spatialpoints = NULL)
}
\arguments{
\item{object}{the connectivity graph of the vrLayer}

\item{spatialpoints}{the set of spatial points}
}
\description{
get spatial points connected to other spatial points in the connectivity graph of vrLayer
}
