% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{getDeconSingle}
\alias{getDeconSingle}
\title{getDeconvolution}
\usage{
getDeconSingle(
  object,
  features = features,
  sc.object,
  sc.assay = "RNA",
  sc.cluster = "seurat_clusters",
  method = "RCTD",
  ...
)
}
\arguments{
\item{object}{a vrAssay object}

\item{features}{features}

\item{sc.object}{Seurat Object}

\item{sc.assay}{assay of the Seurat Object used for the single cell data reference}

\item{sc.cluster}{metadata column variable used for the single cell data reference}

\item{method}{Deconvolution method, RCTD (spot), SPOTlight (spot), MuSiC (ROI)}

\item{...}{additional parameters passed to \code{}}
}
\description{
Calculate deconvolution of spots and ROIs of a single vrAssay object
}
