% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{getDeconvolution}
\alias{getDeconvolution}
\title{getDeconvolution}
\usage{
getDeconvolution(
  object,
  assay = NULL,
  features = NULL,
  sc.object,
  sc.assay = "RNA",
  sc.cluster = "seurat_clusters",
  method = "RCTD",
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{features}{features}

\item{sc.object}{Seurat Object}

\item{sc.assay}{assay of the Seurat Object used for the single cell data reference}

\item{sc.cluster}{metadata column variable used for the single cell data reference}

\item{method}{Deconvolution method, RCTD (spot), SPOTlight (spot), MuSiC (ROI)}

\item{...}{additional parameters passed to \code{getDeconSingle}}
}
\description{
Calculate deconvolution of spatial spots and ROIs
}
