% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialexpression.R
\name{getDiffExp}
\alias{getDiffExp}
\title{getDiffExp}
\usage{
getDiffExp(
  object,
  assay = NULL,
  group.by,
  group.base = NULL,
  covariates = NULL,
  method = "DESeq2"
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{group.by}{the categorical variable from metadata to get differentially expressed features across}

\item{group.base}{Optional, the base category in \code{group.by} which is used as control group}

\item{covariates}{the covariate variable for the design formula}

\item{method}{the method for DE analysis, e.g. DESeq2}
}
\description{
Get differential expression with DESeq2
}
