% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialexpression.R
\name{getDiffExpDESeq2}
\alias{getDiffExpDESeq2}
\title{getDiffExpDESeq2}
\usage{
getDiffExpDESeq2(data, metadata, group.by, group.base = NULL, covariates)
}
\arguments{
\item{data}{the raw data set}

\item{metadata}{the metadata}

\item{group.by}{the categorical variable from metadata to get differentially expressed features across}

\item{group.base}{Optional, the base category in \code{group.by} which is used as control group}

\item{covariates}{the covariate variable for the design formula}
}
\description{
getDiffExpDESeq2
}
