% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures.VoltRon}
\alias{getFeatures.vrAssay}
\title{Get spatially variable feature}
\usage{
getFeatures(object, ...)

\method{getFeatures}{VoltRon}(object, assay = NULL, ...)

\method{getFeatures}{vrAssay}(object, max.count = 1, n = 3000)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
get spatially variable features in an assay
}
\concept{preprocessing}
