% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{getImage.SpaceRover}
\alias{getImage.SpaceRover}
\title{getImage.SpaceRover}
\usage{
getImage.SpaceRover(sr, ...)
}
\arguments{
\item{sr}{SpaceRover object}
}
\description{
get the image from a SpaceRover Object
}
