% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getImageTabPanels}
\alias{getImageTabPanels}
\title{getImageTabPanels}
\usage{
getImageTabPanels(len_images, type)
}
\arguments{
\item{len_images}{the number of query images}

\item{type}{Either reference (ref) or query (query) image}
}
\description{
The UI for a set of reference/query spatial slides
}
