% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getManualRegisteredImage}
\alias{getManualRegisteredImage}
\title{getManualRegisteredImage}
\usage{
getManualRegisteredImage(images, transmatrix, query_ind, ref_ind, input)
}
\arguments{
\item{images}{the list of images}

\item{transmatrix}{the transformation matrix}

\item{query_ind}{the index of the query image}

\item{ref_ind}{the index of the reference image}

\item{input}{input}
}
\description{
Generating the manually registered images
}
