% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getManualRegisterMatrix}
\alias{getManualRegisterMatrix}
\title{ManualRegisterMatrix}
\usage{
getManualRegisterMatrix(
  ManualRegisterMatrix_list,
  len_images,
  xyTable_list,
  xyTable_ref,
  input,
  output,
  session
)
}
\arguments{
\item{ManualRegisterMatrix_list}{a list of registration matrices of each query image}

\item{len_images}{length of query images}

\item{xyTable_list}{a list of keypoints x,y coordinates for query image}

\item{xyTable_ref}{the keypoints x,y coordinates of the reference image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Manuel registeration of images using manually entered keypoints
}
