% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{getMuSiC}
\alias{getMuSiC}
\title{getMuSiC}
\usage{
getMuSiC(
  object,
  features = NULL,
  sc.object,
  sc.assay = "RNA",
  sc.cluster = "seurat_clusters",
  sc.samples = NULL
)
}
\arguments{
\item{object}{A vrAssay object}

\item{features}{features}

\item{sc.object}{a Seurat object}

\item{sc.assay}{an assay in Seurat object where single cell count data is pulled}

\item{sc.cluster}{metadata column in Seurat provides the cell types in single cell data}

\item{sc.samples}{metadata column in Seurat that provides the samples in the single cell data}
}
\description{
Calculate MuSiC deconvolution for ROIs
}
