% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/clustering.R
\name{getNeighbors}
\alias{getNeighbors}
\alias{getNeighbors.VoltRon}
\title{Get neighbors}
\usage{
getNeighbors(object, ...)

\method{getNeighbors}{VoltRon}(
  object,
  assay = NULL,
  data.type = "pca",
  dims = 1:30,
  method = "kNN",
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{additional parameters passed to \code{FNN:get.knn}}

\item{assay}{assay}

\item{data.type}{the type of embedding used for neighborhood calculation, e.g. raw counts (raw), normalized counts (norm), PCA embeddings (pca), UMAP embeddings (umap) etc.}

\item{dims}{the set of dimensions of the embedding data}

\item{method}{the method used for graph construction, kNN or SNN}
}
\value{
Returns object after normalization
}
\description{
get neighbors in an assay
}
