% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getNicheAssay}
\alias{getNicheAssay}
\title{getNicheAssay}
\usage{
getNicheAssay(
  object,
  assay = NULL,
  label = NULL,
  graph.type = "delaunay",
  new_feature_name = "Niche"
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{label}{grouping label for Niche definition}

\item{graph.type}{the type of graph to determine spatial neighborhood}

\item{new_feature_name}{the name of the new feature set created for the niche assay. Default: "Niche"}
}
\description{
Create Niche Assays
}
