% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getObjectImage.Seurat}
\alias{getObjectImage.Seurat}
\title{getObjectImage.Seurat}
\usage{
getObjectImage.Seurat(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\description{
get the image from a Spatial assay
}
