% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getOmeTiffChannels}
\alias{getOmeTiffChannels}
\title{getOmeTiffChannels}
\usage{
getOmeTiffChannels(ome.tiff)
}
\arguments{
\item{ome.tiff}{location to ome.tiff file}
}
\value{
a vector of channel names
}
\description{
Function that returns the channel names of an image in a ome.tiff file
}
