% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/processing.R
\name{getPCA}
\alias{getPCA}
\alias{getPCA,data_matrix-method}
\alias{getPCA,VoltRon-method}
\title{getPCA}
\usage{
getPCA(object, ...)

\S4method{getPCA}{data_matrix}(object, dims = 30, n.workers = 1, data.type = "norm", seed = 1)

\S4method{getPCA}{VoltRon}(
  object,
  assay = NULL,
  features = NULL,
  feat_type = NULL,
  data.type = "norm",
  dims = 30,
  pca.key = "pca",
  n.workers = 1,
  overwrite = FALSE,
  seed = 1
)
}
\arguments{
\item{object}{a VoltRon object}

\item{...}{arguments passed to other methods.}

\item{dims}{the number of dimensions extracted from PCA}

\item{n.workers}{the number of cores/workers use for parallelization.}

\item{data.type}{the type of data used to calculate PCA from:
"norm" (default), "raw" or an existing embeddings \link{vrEmbeddingNames}.}

\item{seed}{seed}

\item{assay}{assay name (exp: Assay1) or assay class
(exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{features}{the selected features for PCA reduction}

\item{feat_type}{the feature set type}

\item{pca.key}{the key name for the embedding, default: pca}

\item{overwrite}{Whether the existing embedding with name 'type'
should be overwritten in \link{vrEmbeddings}}
}
\description{
calculate PCA of the VoltRon objects
}
