% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getProfileNeighbors}
\alias{getProfileNeighbors}
\title{Get profile specific neighbors}
\usage{
getProfileNeighbors(
  object,
  assay = NULL,
  data.type = "pca",
  dims = 1:30,
  k = 10,
  method = "kNN",
  graph.key = method,
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{data.type}{the type of embedding used for neighborhood calculation, e.g. raw counts (raw), normalized counts (norm), PCA embeddings (pca), UMAP embeddings (umap) etc.}

\item{dims}{the set of dimensions of the embedding data}

\item{k}{number of neighbors for kNN}

\item{method}{the method used for graph construction, SNN or kNN}

\item{graph.key}{the name of the graph}

\item{...}{additional parameters passed to \link{FNN::get.knn}}
}
\description{
Get neighbors of spatial points
}
