% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppManualRegistration}
\alias{getRcppManualRegistration}
\title{getRcppManualRegistration}
\usage{
getRcppManualRegistration(
  query_image,
  ref_image,
  query_landmark,
  reference_landmark,
  invert_query = FALSE,
  invert_ref = FALSE,
  method = "Homography",
  nonrigid = "TPS (OpenCV)",
  compute_accuracy = TRUE
)
}
\arguments{
\item{query_image}{query image}

\item{ref_image}{reference image}

\item{query_landmark}{query landmark points}

\item{reference_landmark}{refernece landmark points}

\item{invert_query}{invert query image}

\item{invert_ref}{invert reference image}

\item{method}{the automated registration method, either TPS or Homography+TPS}

\item{nonrigid}{the non-rigid registration method, "TPS (OpenCV)" or "BSpline (SimpleITK)"}

\item{compute_accuracy}{Should accuracy metrics be computed ?}
}
\description{
Manual registration workflow with Rcpp
}
