% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppWarpImageAuto}
\alias{getRcppWarpImageAuto}
\title{getRcppWarpImageAuto}
\usage{
getRcppWarpImageAuto(ref_image, query_image, mapping)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{mapping}{the list of mappings}
}
\description{
Warping a query image given a homography image
}
