% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppWarpImageManual}
\alias{getRcppWarpImageManual}
\title{getRcppWarpImageManual}
\usage{
getRcppWarpImageManual(ref_image, query_image, mapping)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{hmatrix}{the homography matrix}
}
\description{
Warping a query image given a homography image
}
