% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{getRCTD}
\alias{getRCTD}
\title{getRCTD}
\usage{
getRCTD(
  object,
  sc.object,
  sc.assay = "RNA",
  sc.cluster = "seurat_clusters",
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{sc.object}{Seurat Object}

\item{sc.assay}{assay of the Seurat Object used for the single cell data reference}

\item{sc.cluster}{metadata column variable used for the single cell data reference}

\item{...}{additional parameters passed to \code{create.RCTD} function}
}
\description{
Calculate RCTD deconvolution for spot transcriptomics
}
