% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredObject}
\alias{getRegisteredObject}
\title{getRegisteredObject}
\usage{
getRegisteredObject(obj, trans_matrix)
}
\arguments{
\item{obj}{Object}

\item{trans_matrix}{transformation function for the registration}
}
\description{
get a registered Spatial data object
}
