% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredObject.Seurat}
\alias{getRegisteredObject.Seurat}
\title{getRegisteredObject.Seurat}
\usage{
getRegisteredObject.Seurat(seu, mapping)
}
\arguments{
\item{seu}{Seurat object}

\item{mapping}{a list of transformation mapping matrices for the registration}
}
\description{
Get a Seurat Object wigth the registered spatial coordinates and images
}
