% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredObject.SpaceRover}
\alias{getRegisteredObject.SpaceRover}
\title{getRegisteredObject.SpaceRover}
\usage{
getRegisteredObject.SpaceRover(sr, mapping_list, register_ind, centre)
}
\arguments{
\item{mapping_list}{a list of transformation matrices}

\item{register_ind}{the indices of query images/spatialdatasets}

\item{centre}{the index of the central referance image/spatialdata}

\item{sr_list}{a list of SpaceRover objects}
}
\description{
get registered and merged SpaceRover object composed of several Samples
}
