% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredObject.VoltRon}
\alias{getRegisteredObject.VoltRon}
\title{getRegisteredObject.VoltRon}
\usage{
getRegisteredObject.VoltRon(
  sr,
  mapping_list,
  register_ind,
  centre,
  reg_mode = "manual"
)
}
\arguments{
\item{sr}{a list of VoltRon objects}

\item{mapping_list}{a list of transformation matrices}

\item{register_ind}{the indices of query images/spatialdatasets}

\item{centre}{the index of the central reference image/spatialdata}

\item{reg_mode}{the registration mode, either "auto" or "manual"}
}
\description{
Get registered and merged VoltRon object composed of several Samples
}
