% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredObjectListVoltRon}
\alias{getRegisteredObjectListVoltRon}
\title{getRegisteredObjectListVoltRon}
\usage{
getRegisteredObjectListVoltRon(
  sr,
  mapping_list,
  register_ind,
  centre,
  input,
  reg_mode = "manual",
  image_list = NULL
)
}
\arguments{
\item{sr}{a list of VoltRon objects}

\item{mapping_list}{a list of transformation matrices}

\item{register_ind}{the indices of query images/spatialdatasets}

\item{centre}{the index of the central reference image/spatialdata}

\item{input}{input}

\item{reg_mode}{the registration mode, either "auto" or "manual"}

\item{image_list}{the list of query/ref images}
}
\description{
Get registered and merged VoltRon object composed of several Samples
}
