% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredSeurat}
\alias{getRegisteredSeurat}
\title{getRegisteredSeurat}
\usage{
getRegisteredSeurat(seu, trans_matrix)
}
\arguments{
\item{seu}{Seurat object}

\item{trans_matrix}{transformation function for the registration}
}
\description{
get the cell data/locations from a Spatial assay
}
