% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteringCells}
\alias{getRegisteringCells}
\title{getRegisteringCells}
\usage{
getRegisteringCells(query_list)
}
\arguments{
\item{query_list}{the list of Seurat object with query cells/barcodes for registration}
}
\description{
get the cell data/locations from Spatial assays that will be used for registration
}
