% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteringImages}
\alias{getRegisteringImages}
\title{getRegisteringImages}
\usage{
getRegisteringImages(reference, query_list)
}
\arguments{
\item{reference}{Seurat object with reference spatial assay}

\item{query_list}{the list of Seurat object with query images for registration}
}
\description{
get the images from Spatial assays that will be used for registration
}
