% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getSeuratCellData}
\alias{getSeuratCellData}
\title{getSeuratCellData}
\usage{
getSeuratCellData(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\description{
get the cell data/locations from a Spatial assay
}
