% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getSharedFeatures}
\alias{getSharedFeatures}
\title{getSharedFeatures}
\usage{
getSharedFeatures(object, assay = NULL, n = 3000, ...)
}
\description{
get shared variable features across multiple assays
}
