% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getSpatialNeighbors}
\alias{getSpatialNeighbors}
\title{Get spatial neighbors}
\usage{
getSpatialNeighbors(
  object,
  assay = NULL,
  group.by = NULL,
  group.ids = NULL,
  method = "delaunay",
  k = 10,
  radius = numeric(0),
  graph.key = method,
  verbose = TRUE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities.}

\item{group.ids}{a subset of categories defined in metadata column from \code{group.by}.}

\item{method}{the method spatial connectivity: "delaunay", "spatialkNN", "radius".}

\item{k}{number of neighbors for kNN.}

\item{radius}{When \code{method = "radius"} selected, determines the radius of a neighborhood ball around each spatial point.}

\item{graph.key}{the name of the graph.}

\item{verbose}{verbose}
}
\description{
get neighbors in an assay given spatial coordinates
}
