% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{getSpotsFromCells}
\alias{getSpotsFromCells}
\title{getSpotsFromCells}
\usage{
getSpotsFromCells(
  from_object,
  from_metadata = NULL,
  to_object,
  features = NULL
)
}
\arguments{
\item{from_object}{The vrAssay object whose data transfer to the second assay}

\item{from_metadata}{the metadata associated with \code{from_object}}

\item{to_object}{The ID of the target vrAssay object where data is transfered to}

\item{features}{the name of the metadata feature to transfer, if NULL, the rawdata will be transfered}
}
\description{
Generate Psuedo counts per spots and insert to as a separate image assay to the Visium object
}
