% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Graph}
\alias{Graph}
\alias{Graph.VoltRon}
\title{Graph}
\usage{
Graph(object, ...)

\method{Graph}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Graph}: The name of the default assay
}
\description{
Get graph from the main.assay
}
\concept{data-access}
