% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{HeatmapPlot}
\alias{HeatmapPlot}
\title{HeatmapPlot}
\usage{
HeatmapPlot(
  object,
  assay = NULL,
  assay.type = NULL,
  group.by = "clusters",
  norm = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{norm}{}
}
\description{
HeatmapPlot
}
