% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{Image}
\alias{Image}
\alias{Image.Seurat}
\alias{Image.SpaceRover}
\alias{Image.srSample}
\alias{Image.srLayer}
\title{Image}
\usage{
Image(object, ...)

\method{Image}{Seurat}(seu, ...)

\method{Image}{SpaceRover}(sr, ...)

\method{Image}{srSample}(sr, ...)

\method{Image}{srLayer}(sr, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Image}: The name of the default assay
}
\description{
Get and set images of a spaceRover assay
}
\concept{data-access}
