% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{image_ggplot_keypoint}
\alias{image_ggplot_keypoint}
\title{Image to ggplot}
\usage{
image_ggplot_keypoint(image, keypoints)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{keypoints to draw on image}
}
\value{
ggplot object
}
\description{
Transform magick image to a ggplot layer with given keypoints
}
