% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{imageKeypoint}
\alias{imageKeypoint}
\title{imageKeypoint}
\usage{
imageKeypoint(image, keypoints)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{keypoints to draw on image}
}
\value{
ggplot object
}
\description{
add keypoints as points on ggplot object
}
