% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{ImageOutput}
\alias{ImageOutput}
\title{ImageOutput}
\usage{
ImageOutput(image_list, keypoints_list = NULL, centre, input, output, session)
}
\arguments{
\item{image_list}{a list of magick images}

\item{keypoints_list}{a list of data frames, each having a set of keypoints}

\item{centre}{the center image index}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Shiny outputs for a set of magick images with keypoints
}
