% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import10Xh5}
\alias{import10Xh5}
\title{import10Xh5}
\usage{
import10Xh5(filename)
}
\arguments{
\item{filename}{the path to h5 file}
}
\value{
a sparse matrix of 10X counts
}
\description{
import the sparse matrix from the H5 file
}
