% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCosMx}
\alias{importCosMx}
\title{importCosMx}
\usage{
importCosMx(
  path,
  assay_name = "CosMx",
  image = NULL,
  spatial = "main",
  image_name = NULL,
  import_molecules = FALSE,
  verbose = TRUE,
  method = "CSV",
  feature_name = NULL,
  ...
)
}
\arguments{
\item{path}{the path to the tiledb folder}

\item{assay_name}{the assay name, default: CosMx}

\item{image}{the reference morphology image of the CosMx assay}

\item{spatial}{the name of the coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the coordinate system, Default: main}

\item{import_molecules}{if TRUE, molecule assay will be created along
with cell assay.}

\item{verbose}{verbose}

\item{method}{the approach for importing the CosMx assay either by the
folder of CSVs or with TileDB array.}

\item{feature_name}{the name/key of the feature set.}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Import CosMx data
}
