% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importGenePS}
\alias{importGenePS}
\title{importGenePS}
\usage{
importGenePS(
  dir.path,
  assay_name = "GenePS",
  sample_name = NULL,
  use_image = TRUE,
  resolution_level = 7,
  spatial = "main",
  image_name = NULL,
  channel_name = "DAPI",
  import_molecules = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium output folder}

\item{assay_name}{the assay name of the SR object}

\item{sample_name}{the name of the sample}

\item{use_image}{if TRUE, the DAPI image will be used.}

\item{resolution_level}{the level of resolution within TIFF image. Default: 7 (971x638)}

\item{spatial}{the name of the coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the coordinate system, Default: main}

\item{channel_name}{the channel name of the image of the Xenium assay, Default: DAPI}

\item{import_molecules}{if TRUE, molecule assay will be created along with cell assay.}

\item{verbose}{verbose}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Importing GenePS data
}
