% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportGeoMxROISegments}
\alias{ImportGeoMxROISegments}
\title{ImportGeoMxROISegments}
\usage{
ImportGeoMxROISegments(ome.tiff)
}
\arguments{
\item{ome.tiff}{the OME.TIFF file of the GeoMx Experiment}
}
\description{
Import ROI polygons from the OME.TIFF file
}
