% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportGeoMxSegments}
\alias{ImportGeoMxSegments}
\title{ImportGeoMxSegments}
\usage{
ImportGeoMxSegments(ome.tiff, summary, imageinfo)
}
\arguments{
\item{ome.tiff}{the OME.TIFF file of the GeoMx Experiment}

\item{summary}{segmentation summary data frame}

\item{imageinfo}{image information}
}
\description{
Import ROI polygons from the OME.TIFF file
}
