% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importPhenoCycler}
\alias{importPhenoCycler}
\title{importPhenoCycler}
\usage{
importPhenoCycler(
  dir.path,
  assay_name = "PhenoCycler",
  sample_name = NULL,
  spatial = "main",
  image_name = NULL,
  type = c("inform", "processor", "qupath"),
  filter = "DAPI|Blank|Empty",
  inform.quant = c("mean", "total", "min", "max", "std"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{path to PhenoCycler output folder}

\item{assay_name}{the assay name of the SR object}

\item{sample_name}{the name of the sample}

\item{spatial}{the name of the spatial coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the spatial coordinate system, Default: main}

\item{type}{Specify which type matrix is being provided.
\itemize{
 \item \dQuote{\code{processor}}: matrix generated by CODEX Processor
 \item \dQuote{\code{inform}}: matrix generated by inForm
 \item \dQuote{\code{qupath}}: matrix generated by QuPath
}}

\item{filter}{A pattern to filter features by; pass \code{NA} to skip feature filtering}

\item{inform.quant}{When \code{type} is \dQuote{\code{inform}}, the quantification level to read in}

\item{verbose}{verbose}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\value{
VoltRon object
}
\description{
Importing PhenoCycler data
}
