% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importQuPathIF}
\alias{importQuPathIF}
\title{importImageData}
\usage{
importQuPathIF(
  measurements,
  image,
  segments,
  spatial = "main",
  image_name = NULL,
  channels = NULL,
  series = 1,
  resolution = NULL,
  ...
)
}
\arguments{
\item{measurements}{measurements}

\item{image}{a single or a list of image paths or magick-image objects}

\item{segments}{Either a list of segments. This will result in a second 
assay in the VoltRon object to be created}

\item{spatial}{the name of the spatial coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the spatial coordinate system, Default: main}

\item{channels}{the channel names of the images if multiple
images are provided}

\item{series}{the series IDs of the pyramidal image,
typically an integer starting from 1}

\item{resolution}{the resolution IDs of the
pyramidal image, typically an integer starting from 1}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
import an QuPath-quantified IF assay as VoltRon object
}
