% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportVisium_old}
\alias{ImportVisium_old}
\title{ImportVisium_old}
\usage{
ImportVisium_old(
  dir.path,
  assay_name = "Visium",
  layer_name = NULL,
  sample_name = NULL,
  InTissue = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium}

\item{assay_name}{the assay name}

\item{layer_name}{the layer name}

\item{sample_name}{the sample name}

\item{inTissue}{if TRUE, only barcodes that are in the tissue will be kept (default: TRUE)}
}
\description{
ImportVisium_old
}
