% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{initateKeypoints}
\alias{initateKeypoints}
\title{initateKeypoints}
\usage{
initateKeypoints(len_images, keypoints_list, input, output, session)
}
\arguments{
\item{len_images}{the number of query images}

\item{keypoints_list}{the keypoints list of query images}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a shiny reactive values object
}
\description{
Initiate shiny reactive values for keypoint dataframes for pairwise reference and query images
}
