% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{initiateQueryMatrices}
\alias{initiateQueryMatrices}
\title{initiateQueryMatrices}
\usage{
initiateQueryMatrices(len_images, input, output, session)
}
\arguments{
\item{len_images}{the number of query images}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Initiate shiny reactive values for registeration matrices
}
