% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{knn_annoy}
\alias{knn_annoy}
\title{knn_annoy}
\usage{
knn_annoy(data, query = data, k = 10, n_trees = 50, search_k = -1)
}
\arguments{
\item{data}{data}

\item{query}{query data (Default: data)}

\item{k}{number of neighbors for kNN}

\item{n_trees}{Number of trees during index build time. More trees gives higher accuracy}

\item{search_k}{Number of nodes to inspect during the query, or -1 for default value. Higher number gives higher accuracy}
}
\description{
knn engine employed by RcppAnnoy package, adapted from \code{BPCells} package.
}
\details{
**knn_annoy**: Use RcppAnnoy as knn engine
}
