% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{LoadNanostring}
\alias{LoadNanostring}
\title{LoadNanostring}
\usage{
LoadNanostring(
  data.dir,
  fov,
  assay = "Nanostring",
  type = c("centroids", "segmentations")
)
}
\arguments{
\item{data.dir}{Path to folder containing Nanostring SMI outputs}

\item{fov}{FOV name}

\item{assay}{assay}

\item{type}{the type of boundaries used for cells}
}
\description{
Load Nanostring CosMx data
}
