% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadVoltRon}
\alias{loadVoltRon}
\title{loadVoltRon}
\usage{
loadVoltRon(dir = "my_se")
}
\arguments{
\item{dir}{the directory that VoltRon object is found.}
}
\description{
load VoltRon object from memory or disk
}
