% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{MainAssay}
\alias{MainAssay}
\alias{MainAssay<-}
\alias{MainAssay.VoltRon}
\alias{MainAssay<-.VoltRon}
\title{Main Assay}
\usage{
MainAssay(object, ...)

MainAssay(object, ...) <- value

\method{MainAssay}{VoltRon}(object, ...)

\method{MainAssay}{VoltRon}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{MainAssay}: The name of the default assay

\code{MainAssay<-}: The name of the default assay updated
}
\description{
Get and set the main assay
}
\concept{data-access}
