% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{make_css}
\alias{make_css}
\title{make_css}
\usage{
make_css(..., file = NULL)
}
\arguments{
\item{...}{css style definitions. Each object you provide must be a list of three elements. The first element will be a vector of the selectors to be styled (e.g. table, th, an id or html class). If the first element is a vector of length greater than one then the selectors will be comma separated in the css. The second element will be a vector of the css definitions and the third element will a vector of the values of those definitions.}

\item{file}{Character sting. If a file name is provided then the css code will be printed into that file. If the argument is NULL (default) then a string will be returned.}
}
\description{
make_css from \code{tableHTML} package
}
