% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{manageKeypoints}
\alias{manageKeypoints}
\title{manageKeypoints}
\usage{
manageKeypoints(
  centre,
  register_ind,
  xyTable_list,
  image_list,
  input,
  output,
  session
)
}
\arguments{
\item{centre}{center image index}

\item{register_ind}{query image indices}

\item{xyTable_list}{a list of keypoints x,y coordinates for each magick image}

\item{image_list}{a lost of magick image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
A list of shiny observe events for tables and auxiliary operations for pairwise reference and query image
}
