% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melc_data}
\alias{melc_data}
\title{Example MELC Data}
\format{
An object of class \code{VoltRon} of length 1.
}
\source{
{Created to serve as an example.}
}
\usage{
melc_data
}
\description{
This VoltRon object is used for testing and as an example
}
\examples{
data(melc_data)
}
\keyword{datasets}
