% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{merge,vrBlock,ANY-method}
\alias{merge,vrBlock,ANY-method}
\title{Merging vrBlock objects}
\usage{
\S4method{merge}{vrBlock,ANY}(x, y, samples = NULL)
}
\arguments{
\item{x}{a vrSample object}

\item{y}{a list of vrSample objects}

\item{samples}{the sample names}
}
\description{
Given a vrBlock object, and a list of vrSample objects, merge all.
}
