% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{merge,vrMetadata,ANY-method}
\alias{merge,vrMetadata,ANY-method}
\title{Merging vrMetadata objects}
\usage{
\S4method{merge}{vrMetadata,ANY}(x, y)
}
\arguments{
\item{x}{a vrMetadata object}

\item{y}{a single or a list of vrMetadata objects}
}
\description{
Given a vrMetadata object, and a list of vrMetadata objects, merge all.
}
