% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge,VoltRon,ANY-method}
\alias{merge,VoltRon,ANY-method}
\alias{merge}
\title{Merging VoltRon objects}
\usage{
\S4method{merge}{VoltRon,ANY}(x, y, samples = NULL, main.assay = NULL, verbose = TRUE)
}
\arguments{
\item{x}{a VoltRon Object}

\item{y}{a single or a list of VoltRon objects}

\item{samples}{a single sample name or multiple sample names of the same size as the given VoltRon objects}

\item{main.assay}{the name of the main assay}

\item{verbose}{verbose}
}
\description{
Given a VoltRon object, and a list of VoltRon objects, merge all.
}
