% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{merge.sampleMetadata}
\alias{merge.sampleMetadata}
\title{merge.sampleMetadata}
\usage{
\method{merge}{sampleMetadata}(metadata_list)
}
\arguments{
\item{metadata_list}{a list of sample metadata of a VoltRon object}

\item{sample_names}{new sample names}
}
\description{
Merging sample.metadata from two VoltRon objects
}
