% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.SpaceRover}
\alias{merge.SpaceRover}
\title{Merging spacerover objects}
\usage{
\method{merge}{SpaceRover}(object, object_list, sample_name = NULL, main.assay = NULL)
}
\arguments{
\item{object}{a SpaceRover Object}

\item{object_list}{a list of SpaceRover objects}

\item{sample_name}{a single sample name if objects are of the same sample}

\item{main.assay}{name of the assay}
}
\description{
Given a spacerover object, and a list of spacerover object, merge all.
}
