% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.VoltRon}
\alias{merge.VoltRon}
\title{Merging VoltRon objects}
\usage{
\method{merge}{VoltRon}(object, object_list, samples = NULL, main.assay = NULL)
}
\arguments{
\item{object}{a VoltRon Object}

\item{object_list}{a list of VoltRon objects}

\item{samples}{a single sample name or multiple sample names of the same size as the given VoltRon objects}

\item{main.assay}{name of the assay}
}
\description{
Given a VoltRon object, and a list of VoltRon objects, merge all.
}
