% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.VoltRon_old}
\alias{merge.VoltRon_old}
\title{Merging VoltRon objects}
\usage{
\method{merge}{VoltRon_old}(object, object_list, sample_name = NULL, main.assay = NULL)
}
\arguments{
\item{object}{a VoltRon Object}

\item{object_list}{a list of VoltRon objects}

\item{sample_name}{a single sample name if objects are of the same sample}

\item{main.assay}{name of the assay}
}
\description{
Given a VoltRon object, and a list of VoltRon object, merge all.
}
