% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{merge.vrBlock}
\alias{merge.vrBlock}
\title{Merging vrBlock objects}
\usage{
\method{merge}{vrBlock}(object, object_list, samples = NULL)
}
\arguments{
\item{object}{a vrSample object}

\item{object_list}{a list of vrSample objects}

\item{samples}{the sample names}
}
\description{
Given a vrBlock object, and a list of vrSample objects, merge all.
}
