% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{merge.vrMetadata}
\alias{merge.vrMetadata}
\title{Merging vrMetadata objects}
\usage{
\method{merge}{vrMetadata}(object, object_list)
}
\arguments{
\item{object_list}{a list of vrMetadata objects}
}
\description{
Given a vrMetadata object, and a list of vrMetadata objects, merge all.
}
