% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{merge.vrSample}
\alias{merge.vrSample}
\title{Merging vrSample objects}
\usage{
\method{merge}{vrSample}(object, object_list, samples = NULL)
}
\arguments{
\item{object}{a vrSample object}

\item{object_list}{a list of vrSample objects}

\item{samples}{the sample names}
}
\description{
Given a vrSample object, and a list of vrSample objects, merge all.
}
