% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge_graphs}
\alias{merge_graphs}
\title{merge_graphs}
\usage{
merge_graphs(object, object_list)
}
\arguments{
\item{object}{a VoltRon Object}

\item{object_list}{a list of VoltRon objects}
}
\description{
Given a VoltRon object, and a list of VoltRon objects, merge their graphs.
}
