% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{merge_sampleMetadata}
\alias{merge_sampleMetadata}
\title{merge.sampleMetadata}
\usage{
merge_sampleMetadata(metadata_list)
}
\arguments{
\item{metadata_list}{a list of sample metadata of a VoltRon object}
}
\description{
Merging sample.metadata from two VoltRon objects
}
