% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_object}
\alias{merged_object}
\title{Example Transfer Data}
\format{
An object of class \code{VoltRon} of length 1.
}
\source{
{Created to serve as an example.}
}
\usage{
merged_object
}
\description{
This VoltRon object is used for testing and as an example for transfering 
information between layers and multilayer plotting.
}
\examples{
data(merged_object)
}
\keyword{datasets}
