% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{mod_app_server}
\alias{mod_app_server}
\title{App Server}
\usage{
mod_app_server(id, plot_g = NULL)
}
\arguments{
\item{id}{id of the module}

\item{plot_g}{the ggplot plot}
}
\description{
App Server
}
