% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{PCA}
\alias{PCA}
\alias{PCA.VoltRon}
\title{PCA}
\usage{
PCA(object, ...)

\method{PCA}{VoltRon}(object, assay = NULL, dims = 30)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{PCA}: The name of the default assay
}
\description{
calculate PCA of the VoltRon objects
}
\concept{data-access}
\concept{embedding}
