% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{point.in.circle}
\alias{point.in.circle}
\title{point.in.circle}
\usage{
point.in.circle(x, y, circle.x, circle.y, radius)
}
\description{
checks if points are in the defined circle ROI with a (x,y) 
center and radius
}
\keyword{internal}
