% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_manipulation.R
\name{PrepareXenium}
\alias{PrepareXenium}
\title{PrepareXenium}
\usage{
PrepareXenium(dir.path, increase.contrast = TRUE, resolutionLevel = 7, ...)
}
\arguments{
\item{dir.path}{Xenium output folder}

\item{increase.contrast}{increase the contrast of the image before writing}

\item{resolutionLevel}{the level of resolution within Xenium OME-TIFF image. Default: 7 (553x402)}

\item{...}{additional parameters passed to the EBImage::writeImage function}
}
\description{
Generates and prepares auxiliary files for a Xenium assay in the output folder
}
