% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\docType{data}
\name{py_env}
\alias{py_env}
\title{The Python Basilisk environment}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
py_env
}
\description{
Defines a conda environment via Basilisk, which is used to convert R objects to Zarr stores.
}
\keyword{internal}
