% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{QueryKeypointTable}
\alias{QueryKeypointTable}
\title{QueryKeypointTable}
\usage{
QueryKeypointTable(xyTable_list, image_list, input, output, session)
}
\arguments{
\item{xyTable_list}{a list of keypoints x,y coordinates for each magick image}

\item{image_list}{a lost of magick image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
A list of shiny observe events for keypoints tables, updating keypoints and auxiliary operations for query image
}
