% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{QueryMatrices}
\alias{QueryMatrices}
\title{QueryMatrices}
\usage{
QueryMatrices(len_images, input, output, session)
}
\arguments{
\item{len_images}{the number of query images}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a shiny reactive values object
}
\description{
Initiate shiny reactive values for registeration matrices
}
