% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{QueryTabPanel}
\alias{QueryTabPanel}
\title{QueryTabPanel}
\usage{
QueryTabPanel(len_images, type)
}
\arguments{
\item{len_images}{the number of query images}
}
\value{
tabsetpanel
}
\description{
The UI for a set of reference/query spatial slides
}
