% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{QueryTabPanels}
\alias{QueryTabPanels}
\title{QueryTabPanels}
\usage{
QueryTabPanels(len_images)
}
\arguments{
\item{len_images}{the number of query images}
}
\value{
tabsetpanel
}
\description{
The UI for a set of query spatial slides
}
