% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readDCC}
\alias{readDCC}
\title{readDCC}
\usage{
readDCC(file)
}
\arguments{
\item{file}{A character string containing the path to the DCC file.}
}
\description{
Read a NanoString GeoMx Digital Count Conversion (DCC) file.
}
