% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readPKC}
\alias{readPKC}
\title{readPKC}
\usage{
readPKC(file, default_pkc_vers = NULL)
}
\arguments{
\item{file}{A character string containing the path to the PKC file.}

\item{default_pkc_vers}{Optional list of pkc file names to use as default if more than one pkc version of each module is provided.}
}
\description{
Read a NanoString Probe Kit Configuration (PKC) file. Adapted from \code{GeomxTools} package.
}
