% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{ReferenceImageOutput}
\alias{ReferenceImageOutput}
\title{ReferenceImageOutput}
\usage{
ReferenceImageOutput(image, xyTable, input, output, session)
}
\arguments{
\item{image}{magick image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{keypoints}{keypoint data frame}
}
\description{
Shiny outputs for a set of magick images with keypoints
}
