% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{ReferenceKeypoints}
\alias{ReferenceKeypoints}
\title{ReferenceKeypoints}
\usage{
ReferenceKeypoints(keypoints = NULL, input, output, session)
}
\arguments{
\item{keypoints}{the keypoints of reference image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a shiny reactive values object
}
\description{
Initiate shiny reactive values for keypoint dataframes for the reference image
}
