% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{ReferenceKeypointTable}
\alias{ReferenceKeypointTable}
\title{ReferenceKeypointTable}
\usage{
ReferenceKeypointTable(xyTable, image, input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{xyTable_list}{a list of keypoints x,y coordinates for each magick image}

\item{image_list}{a lost of magick image}
}
\description{
A list of shiny observe events for keypoints tables, updating keypoints and auxiliary operations for reference image
}
