% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{register_slides}
\alias{register_slides}
\title{register_slides}
\usage{
register_slides(reference_visium, query_image_paths, keypoints = NULL)
}
\arguments{
\item{reference_visium}{visium data}

\item{query_images}{a list of xenium data as query}
}
\value{
projection matrix
}
\description{
A mini shiny app to calculate spatial cell(barcode) projection matrix between a visium slide and a xenium slide
}
