% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rescaleGeoMxPoints}
\alias{rescaleGeoMxPoints}
\title{rescaleGeoMxROIs}
\usage{
rescaleGeoMxPoints(pts, summary, imageinfo)
}
\arguments{
\item{pts}{coordinates of the cells from the Xenium assays}

\item{summary}{segmentation summary data frame}

\item{imageinfo}{image information}
}
\description{
Rescale GeoMx point (center or polygon corners of ROI) coordinates for image registration
}
