% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{rescaleXeniumCells}
\alias{rescaleXeniumCells}
\title{rescaleXeniumCells}
\usage{
rescaleXeniumCells(cells, bbox, image)
}
\arguments{
\item{cells}{coordinates of the cells from the Xenium assays}

\item{bbox}{the surrounding box of the Xenium cell coordinates}

\item{image}{reference image}
}
\description{
rescale Xenium cells coordinates for image registration
}
