% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{rotate_keypoint}
\alias{rotate_keypoint}
\title{rotate_keypoint}
\usage{
rotate_keypoint(
  keypoints,
  angle,
  origin,
  limits,
  rotated_origin,
  rotated_limits
)
}
\arguments{
\item{keypoints}{dataset of keypoints}

\item{angle}{angle of rotation [0,360]}

\item{origin}{center of the image}

\item{limits}{limits of the image}

\item{rotated_origin}{center of the rotated image}

\item{rotated_limits}{limits of the rotated image}
}
\value{
keypoints
}
\description{
Find transformations of keypoints under clockwise rotations of the image
}
