% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{RunMuSiC}
\alias{RunMuSiC}
\title{RunMuSiC}
\usage{
RunMuSiC(
  object,
  sc.object,
  sc.assay = "RNA",
  sc.cluster = "seurat_clusters",
  sc.Samples = NULL
)
}
\arguments{
\item{object}{An srAssay object}

\item{sc.object}{a Seurat object}

\item{sc.assay}{an assay in Seurat object where single cell count data is pulled}

\item{sc.cluster}{metadata column in Seurat provides the cell types in single cell data}

\item{sc.Samples}{metadata column in Seurat that provides the samples in the single cell data}
}
\description{
RunMuSiC
}
