% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{SampleMetadata}
\alias{SampleMetadata}
\alias{SampleMetadata.SpaceRover}
\title{SampleMetadata}
\usage{
SampleMetadata(object, ...)

\method{SampleMetadata}{SpaceRover}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{SampleMetadata}: The name of the default assay
}
\description{
Get the sample metadata
}
\concept{data-access}
