% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Segments.srAssay}
\alias{Segments.srAssay}
\alias{Segments<-.srAssay}
\alias{Segments}
\alias{Segments<-}
\alias{Segments.SpaceRover}
\alias{Segments<-.SpaceRover}
\title{Segments}
\usage{
\method{Segments}{srAssay}(object, reg = FALSE, ...)

\method{Segments}{srAssay}(object, reg = FALSE, ...) <- value

Segments(object, ...)

Segments(object, ...) <- value

\method{Segments}{SpaceRover}(object, reg = FALSE, assay = NULL, ...)

\method{Segments}{SpaceRover}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Segments}: The name of the default assay

\code{Segments<-}: The coordinates updated
}
\description{
Get and set the segments of spatian points of the main assay
}
\concept{data-access}
