% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setSRMetadata}
\alias{setSRMetadata}
\title{setSRMetadata}
\usage{
setSRMetadata(cell, spot, ROI)
}
\arguments{
\item{cell}{cell data frame}

\item{spot}{spot data frame}

\item{ROI}{ROI data frame}
}
\description{
setSRMetadata
}
