% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setSRSampleMetadata}
\alias{setSRSampleMetadata}
\title{setSRSampleMetadata}
\usage{
setSRSampleMetadata(samples)
}
\arguments{
\item{samples}{a list of srSample object}
}
\description{
setSRSampleMetadata
}
