% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setVRSampleMetadata}
\alias{setVRSampleMetadata}
\title{setVRSampleMetadata}
\usage{
setVRSampleMetadata(samples)
}
\arguments{
\item{samples}{a list of vrSample object}
}
\description{
setVRSampleMetadata
}
