% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SFPlot}
\alias{SFPlot}
\title{SFPlot}
\usage{
SFPlot(object, features, ncol = 3, nrow = NULL, sample = NULL, font.size = 2)
}
\arguments{
\item{object}{SpaceRover object}

\item{features}{a vector of features to visualize}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{sample}{sample names matching spatial points (i.e. entities)}

\item{font.size}{font sizes}
}
\description{
Functions for plotting spatial objects
}
