% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SFPlotMulti}
\alias{SFPlotMulti}
\title{SFPlotMulti}
\usage{
SFPlotMulti(
  object,
  features,
  group.by = NULL,
  assay = "GeoMx",
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  keep.scale = "feature"
)
}
\arguments{
\item{object}{SpaceRover object}

\item{features}{a vector of features to visualize}

\item{group.by}{a grouping label for the spatial entities}

\item{assay}{the assay type name}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{font.size}{font sizes}

\item{keep.scale}{whether unify all scales for all features or not}
}
\description{
Functions for plotting spatial objects
}
