% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SFPlotMulti2}
\alias{SFPlotMulti2}
\title{SFPlotMulti2}
\usage{
SFPlotMulti2(
  object,
  features,
  sample = NULL,
  split.by = NULL,
  assay = "GeoMx",
  ncol = 3,
  nrow = NULL,
  font.size = 2,
  keep.scale = "feature"
)
}
\arguments{
\item{object}{SpaceRover object}

\item{features}{a vector of features to visualize}

\item{sample}{sample names matching spatial points (i.e. entities)}

\item{split.by}{split by a metadata element}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{font.size}{font sizes}

\item{keep.scale}{whether unify all scales for all features or not}

\item{mode}{plotting mode for different spatial entities, ROI, Spot or cell}
}
\description{
Functions for plotting spatial objects
}
