% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{shinyInputLabel}
\alias{shinyInputLabel}
\title{Shiny's internal \code{shinyInputLabel} function}
\usage{
shinyInputLabel(inputId, label = NULL)
}
\description{
Reproduced since it is not exported in the Shiny namespace.
}
\keyword{internal}
