% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{slotApply}
\alias{slotApply}
\title{slotApply}
\usage{
slotApply(x, FUN, ...)
}
\arguments{
\item{x}{object}

\item{FUN}{function}

\item{...}{arguments passed to \code{FUN}}
}
\description{
apply to slots
}
