% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SpaceRover-class}
\alias{SpaceRover-class}
\alias{SpaceRover}
\title{The SpaceRover Class}
\description{
The SpaceRover Class
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of samples for the this project}

\item{\code{integrated.datasets}}{A list of integrated data objects that indicate integrated spatial layers}

\item{\code{meta.data}}{Contains meta-information about each sample}

\item{\code{project}}{Name of the project}
}}

