% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spaceRover-package}
\alias{spaceRover}
\alias{spaceRover-package}
\title{spaceRover: spaceRover for Spatial Data Analysis}
\description{
spaceRover is a spatial data analysis toolbox for spatial data analysis, integration and image registration
}
