% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatFeatPlotSingle}
\alias{SpatFeatPlotSingle}
\title{SpatFeatPlotSingle}
\usage{
SpatFeatPlotSingle(
  assay,
  metadata,
  feature,
  limits,
  group.by = "label",
  font.size = 2,
  pt.size = 2,
  alpha = 0.6,
  label = FALSE,
  plot_title = NULL,
  legend_title = NULL,
  background = "image",
  crop = FALSE
)
}
\arguments{
\item{crop}{}
}
\description{
A single Spatial Feature plot of spacerover objects
}
