% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{SpatialRegistration}
\alias{SpatialRegistration}
\title{register_slides}
\usage{
SpatialRegistration(
  spatial_data_list = NULL,
  reference_spatdata = NULL,
  query_spatdata = NULL,
  keypoints = NULL
)
}
\arguments{
\item{spatial_data_list}{a list of spatial data sets}

\item{reference_spatdata}{a reference spatial data set, used only if \code{spatial_data_list} is \code{NULL}}

\item{query_spatdata}{a query spatial data set, used only if \code{spatial_data_list} is \code{NULL}}
}
\value{
keypoints
}
\description{
A mini shiny app to calculate spatial cell(barcode) projection matrix between a visium slide and a xenium slide
}
