% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatPlot}
\alias{SpatPlot}
\title{SpatPlot}
\usage{
SpatPlot(
  object,
  group.by = "label",
  assay = "Visium",
  assay.type = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  alpha = 0.6,
  label = FALSE,
  background = "image",
  crop = FALSE,
  common.legend = TRUE
)
}
\arguments{
\item{object}{SpaceRover object}

\item{group.by}{a grouping label for the spatial entities}

\item{assay}{the assay name}

\item{assay.type}{the assay type name: 'cell', 'spot' or 'ROI'}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{alpha}{alpha}

\item{background}{background color or image}

\item{common.legend}{whether to use a common legend for all plots}
}
\description{
Functions for plotting spatial objects
}
