% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatPlotSingle}
\alias{SpatPlotSingle}
\title{SpatFeatPlotSingle}
\usage{
SpatPlotSingle(
  assay,
  metadata,
  limits,
  group.by = "label",
  font.size = 2,
  pt.size = 10,
  alpha = 0.6,
  label = FALSE,
  plot_title = NULL,
  background = "image"
)
}
\arguments{
\item{background}{}
}
\description{
A single Spatial Feature plot of spacerover objects
}
