% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{srAssay-class}
\alias{srAssay-class}
\alias{srAssay}
\title{The srAssay (SpaceRover Assay) Class}
\description{
The srAssay (SpaceRover Assay) Class
}
\section{Slots}{

\describe{
\item{\code{image}}{}
}}

