% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{srImageOutput}
\alias{srImageOutput}
\title{srImageOutput}
\usage{
srImageOutput(image_list, keypoints_list = NULL, input, output, session)
}
\arguments{
\item{image_list}{a list of magick images}

\item{keypoints_list}{a list of data frames, each having a set of keypoints}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Shiny outputs for a set of magick images with keypoints
}
