% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{srLayer-class}
\alias{srLayer-class}
\alias{srLayer}
\title{The srLayer (SpaceRover Layer) Class}
\description{
The srLayer (SpaceRover Layer) Class
}
\section{Slots}{

\describe{
\item{\code{assay}}{}
}}

