% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{srMetadata-class}
\alias{srMetadata-class}
\alias{srMetadata}
\title{The srMetadata (SpaceRover Metadata) Class}
\description{
The srMetadata (SpaceRover Metadata) Class
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of layers for the this srSample object}
}}

