% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{srSample-class}
\alias{srSample-class}
\alias{srSample}
\title{The srSample (SpaceRover Sample) Class}
\description{
The srSample (SpaceRover Sample) Class
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of layers for the this srSample object}
}}

