% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{[[,VoltRon,character,missing-method}
\alias{[[,VoltRon,character,missing-method}
\title{}
\usage{
\S4method{[[}{VoltRon,character,missing}(x, i, j, ...)
}
\description{
# update sample metadata and metadata
  x@sample.metadata[[i]] <- value
  Metadata(x)[[i]] <- value
}
\details{

}
