% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subset,vrAssay-method}
\alias{subset,vrAssay-method}
\title{Subsetting vrAssay objects}
\usage{
\S4method{subset}{vrAssay}(x, subset, spatialpoints = NULL, features = NULL, image = NULL)
}
\arguments{
\item{x}{a vrAssay object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{features}{the set of features to subset the object}

\item{image}{the subseting string passed to \link[magick]{image_crop}}
}
\description{
Given a vrAssay object, subset the object given one of the attributes
}
