% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subset,vrAssayV2-method}
\alias{subset,vrAssayV2-method}
\title{Subsetting vrAssayV2 objects}
\usage{
\S4method{subset}{vrAssayV2}(x, subset, spatialpoints = NULL, features = NULL, image = NULL)
}
\arguments{
\item{x}{a vrAssayV2 object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{features}{the set of features to subset the object}

\item{image}{the subseting string passed to \link[magick]{image_crop}}
}
\description{
Given a vrAssayV2 object, subset the object given one of the attributes
}
