% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{subset,vrLayer-method}
\alias{subset,vrLayer-method}
\title{Subsetting vrLayer objects}
\usage{
\S4method{subset}{vrLayer}(x, subset, assays = NULL, spatialpoints = NULL, image = NULL)
}
\arguments{
\item{x}{a vrLayer object}

\item{subset}{the subset statement}

\item{assays}{the set of assays to subset the object}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \link[magick]{image_crop}}
}
\description{
Given a vrLayer object, subset the object given one of the attributes
}
