% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{subset,vrMetadata-method}
\alias{subset,vrMetadata-method}
\title{Subsetting vrMetadata objects}
\usage{
\S4method{subset}{vrMetadata}(x, subset, samples = NULL, assays = NULL, spatialpoints = NULL)
}
\arguments{
\item{x}{a vrMetadata object}

\item{subset}{the subset statement}

\item{samples}{the set of samples to subset the object}

\item{assays}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{spatialpoints}{the set of spatial points to subset the object}
}
\description{
Given a vrMetadata object, subset the object given one of the attributes
}
