% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{subset,vrSpatial-method}
\alias{subset,vrSpatial-method}
\title{Subsetting vrSpatial objects}
\usage{
\S4method{subset}{vrSpatial}(x, subset, spatialpoints = NULL, image = NULL)
}
\arguments{
\item{x}{A vrSpatial object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \link[magick]{image_crop}}
}
\description{
Given a vrSpatial object, subset the object given one of the attributes.
}
