% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{subset.Connectivity}
\alias{subset.Connectivity}
\title{subset.Connectivity}
\usage{
\method{subset}{Connectivity}(object, spatialpoints = NULL)
}
\arguments{
\item{object}{the connectivity graph of the vrLayer}

\item{spatialpoints}{the set of spatial points}
}
\description{
Subsetting the connectivity graph of vrLayer using spatial points
}
