% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{subset.vrMetadata}
\alias{subset.vrMetadata}
\title{Subsetting vrMetadata objects}
\usage{
\method{subset}{vrMetadata}(metadata, subset, samples = NULL, assays = NULL, spatialpoints = NULL)
}
\arguments{
\item{samples}{the set of samples to subset the object}

\item{assays}{the set of assays to subset the object}

\item{spatialpoints}{the set of spatial points to subset the object}
}
\description{
Given a vrMetadata object, subset the object given one of the attributes
}
