% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{subset.vrSpatial}
\alias{subset.vrSpatial}
\title{Subsetting vrSpatial objects}
\usage{
\method{subset}{vrSpatial}(object, ...)
}
\arguments{
\item{object}{A vrSpatial object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \link{image_crop}}
}
\description{
Given a vrSpatial object, subset the object given one of the attributes.
}
