% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{subset_graphs}
\alias{subset_graphs}
\title{subset_graphs}
\usage{
subset_graphs(object, metadata)
}
\arguments{
\item{object}{a VoltRon Object}

\item{metadata}{a vrMetadata Object}
}
\description{
Given a VoltRon object and a vrMetadata, subset the graph
}
