% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{subset_sampleMetadata}
\alias{subset_sampleMetadata}
\title{subset_sampleMetadata}
\usage{
subset_sampleMetadata(metadata, samples = NULL, assays = NULL)
}
\arguments{
\item{metadata}{sample metadata of a VoltRon object}

\item{samples}{the set of samples to subset the object}

\item{assays}{the set of assays to subset the object}
}
\description{
Subseting sample metadata
}
