% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{textInputwithButton}
\alias{textInputwithButton}
\title{internal Text input with button}
\usage{
textInputwithButton(
  textinputId,
  label,
  buttoninputId,
  value = "",
  width = NULL,
  placeholder = NULL
)
}
\description{
Reproduced since it is not exported in the Shiny namespace.
}
\keyword{internal}
