% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/integration.R
\name{transferData}
\alias{transferData}
\alias{transferData,VoltRon,character-method}
\alias{transferData,VoltRon,Seurat-method}
\alias{transferData,VoltRon,SingleCellExperiment-method}
\title{transferData}
\usage{
transferData(object, from, ...)

\S4method{transferData}{VoltRon,character}(
  object,
  from = NULL,
  to = NULL,
  features = NULL,
  expand = NULL,
  new_feature_name = NULL
)

\S4method{transferData}{VoltRon,Seurat}(
  object,
  from,
  to = NULL,
  features = NULL,
  new_feature_name = NULL,
  sc.assay = NULL,
  pc.dims = 30,
  seed = 1
)

\S4method{transferData}{VoltRon,SingleCellExperiment}(
  object,
  from,
  to = NULL,
  features = NULL,
  new_feature_name = NULL,
  sc.assay = NULL,
  pc.dims = 30,
  seed = 1
)
}
\arguments{
\item{object}{a VoltRon object}

\item{from}{if a character, the name of the assay whose data transfered to the second assay, 
otherwise a single cell object of class Seurat or SingleCellExperiment whose data is 
transfered to the VoltRon assay provided in \code{to}.}

\item{...}{arguments passed to other methods.}

\item{to}{the name or class of target assay where data is transfered to}

\item{features}{the set of features from \link{vrFeatures} or metadata columns from \link{Metadata} 
that are transferred. Only one metadata feature can be transferred at a time.}

\item{expand}{if TRUE, metadata features will be transformed into
dummy features where each category in the feature will be a new feature.
If FALSE, metadata features will not be transformed and transfered as
metadata columns, else the decision will be made automatically.}

\item{new_feature_name}{the name of the new feature set created
from the source assay defined in \code{from} argument.
Only used when a new assay in created. If a \code{features} is a metadata 
column and \code{from} is a single cell object of Seurat or 
SingleCellExperiment, then \code{new_feature_name} is the name of the new 
metadata column in the VoltRon object.}

\item{sc.assay}{When \code{from} is a single cell object (Seurat or 
SingleCellExperiment), sc.assay can be provided to choose a specific assay
from the single cell object.}

\item{pc.dims}{the number of dimensions extracted from PCA. Only used 
when \code{sc.object} is a Seurat or SingleCellExperiment object.}

\item{seed}{seed}
}
\description{
Transfering features across spatial omics assays and single cell assays (Seurat or SingleCellExperiment). 
This functions facilitates the transfer of features between both (i) VoltRon assays in a common 
coordinate system and tissue block, and also (ii) features between a VoltRon assay and a single cell assay 
with Seurat or SingleCellExperiment object.
}
