% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{transform_keypoints}
\alias{transform_keypoints}
\title{transform_keypoints}
\usage{
transform_keypoints(image, keypoints, extension, input, session)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{keypoints visualized on image}

\item{extension}{name extension for the shiny input parameter}

\item{input}{input}

\item{session}{session}
}
\value{
magick image
}
\description{
Apply transformations to keypoints given transformed images to find the keypoints locations in the original image
}
