% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{transform_magick_image}
\alias{transform_magick_image}
\title{Rotate and Transform magick Image}
\usage{
transform_magick_image(image, extension, input, session)
}
\arguments{
\item{image}{magick image}

\item{extension}{name extension for the shiny input parameter}

\item{input}{input}

\item{session}{session}
}
\value{
magick image
}
\description{
apply given transformations to a magick image
}
