% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{transform_magick_image_keypoints}
\alias{transform_magick_image_keypoints}
\title{transform_magick_image_keypoints}
\usage{
transform_magick_image_keypoints(image, keypoints, extension, input, session)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{a set of keypoints}

\item{extension}{name extension for the shiny input parameter}

\item{input}{input}

\item{session}{session}
}
\value{
a list of magick image and keypoints
}
\description{
Apply given transformations to a magick image and keypoints simultaneously for plotting
}
