% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{transform_magick_image_query_list}
\alias{transform_magick_image_query_list}
\title{transform_magick_image}
\usage{
transform_magick_image_query_list(image_list, input, session)
}
\arguments{
\item{input}{shiny input}

\item{session}{shiny session}

\item{image}{magick image}

\item{extension}{name extension for the shiny input parameter}
}
\value{
magick image
}
\description{
Apply given transformations to a magick image
}
