% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{transformImageKeypoints}
\alias{transformImageKeypoints}
\title{transformImageKeypoints}
\usage{
transformImageKeypoints(image, keypoints, extension, input, session)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{a set of keypoints}

\item{extension}{name extension for the shiny input parameter}

\item{input}{shiny input}

\item{session}{shiny session}
}
\description{
Apply given transformations to a magick image and keypoints for plotting
}
