% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{transformImageQueryList}
\alias{transformImageQueryList}
\title{transformImageQueryList}
\usage{
transformImageQueryList(image_list, input, session)
}
\arguments{
\item{image_list}{magick image}

\item{input}{shiny input}

\item{session}{shiny session}
}
\description{
Apply given transformations to a list of magick image and return shiny reactive
}
