% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{transformKeypoints}
\alias{transformKeypoints}
\title{transformKeypoints}
\usage{
transformKeypoints(image, keypoints, extension, input)
}
\arguments{
\item{image}{magick image}

\item{keypoints}{keypoints visualized on image}

\item{extension}{name extension for the shiny input parameter}

\item{input}{shiny input}
}
\description{
Apply transformations to keypoints given transformed images to find the keypoints locations in the original image
}
