% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{updateAssay}
\alias{updateAssay}
\title{Update Assay}
\usage{
updateAssay(object, ...)
}
\arguments{
\item{object}{a VoltRon object.}

\item{...}{arguments passed to other methods.}
}
\description{
update assays from vrAssay to vrAssayV2
}
