% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{updateGraphAssay}
\alias{updateGraphAssay}
\title{updateGraphAssay}
\usage{
updateGraphAssay(object1, object2)
}
\arguments{
\item{object1}{VoltRon object}

\item{object2}{VoltRon object}
}
\description{
updateGraphAssay
}
