% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{updateParameterPanels}
\alias{updateParameterPanels}
\title{updateParameterPanels}
\usage{
updateParameterPanels(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
A function for managing which parameter panels or input boxes to appear on UI
}
