% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{updateSequentialTabPanels}
\alias{updateSequentialTabPanels}
\title{updateSequentialTabPanels}
\usage{
updateSequentialTabPanels(input, output, session, centre, register_ind)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{centre}{center image index}

\item{register_ind}{query image indices}
}
\description{
A function for automatized selection of reference/query tab panels
}
