% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visium_data}
\alias{visium_data}
\title{Example Visium Data}
\format{
An object of class \code{VoltRon} of length 1.
}
\source{
{Created to serve as an example.}
}
\usage{
visium_data
}
\description{
This VoltRon object is used for testing and as an example
}
\examples{
data(visium_data)
}
\keyword{datasets}
