% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{vrAssay-class}
\alias{vrAssay-class}
\alias{vrAssay}
\title{The vrAssay (VoltRon Assay) Class}
\description{
The vrAssay (VoltRon Assay) Class
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{raw data}

\item{\code{normdata}}{normalized data}

\item{\code{featuredata}}{feature metadata}

\item{\code{embeddings}}{list of embeddings}

\item{\code{image}}{a list of vrSpatial objects}

\item{\code{params}}{additional parameters used by different assay types}

\item{\code{type}}{the type of the assay (tile, molecule, cell, spot, ROI)}

\item{\code{name}}{the assay name}

\item{\code{main_image}}{the key of the main image}
}}

