% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/metadata.R,
%   R/objects.R
\name{vrAssayNames.vrAssay}
\alias{vrAssayNames.vrAssay}
\alias{vrAssayNames<-.vrAssay}
\alias{vrAssayNames}
\alias{vrAssayNames<-}
\alias{vrAssayNames.vrMetadata}
\alias{vrAssayNames.VoltRon}
\title{Get Assay names}
\usage{
\method{vrAssayNames}{vrAssay}(object)

\method{vrAssayNames}{vrAssay}(object, ...) <- value

vrAssayNames(object, assay = NULL)

vrAssayNames(object, ...) <- value

\method{vrAssayNames}{vrMetadata}(object)

\method{vrAssayNames}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{assay name}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Given a VoltRon, vrMetadata or vrAssay object, get/set names of assays.
}
