% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R
\name{vrAssayParams.vrAssay}
\alias{vrAssayParams.vrAssay}
\alias{vrAssayParams}
\title{Get Assay types}
\usage{
\method{vrAssayParams}{vrAssay}(object, param = NULL)

vrAssayParams(object, ...)
}
\arguments{
\item{object}{An object}

\item{param}{the parameter value to return}

\item{...}{Arguments passed to other methods}
}
\description{
Given a VoltRon object, if there are any, get a list of parameters of the assay(s)
}
