% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{vrAssayTypes}
\alias{vrAssayTypes}
\alias{vrAssayTypes,VoltRon-method}
\alias{vrAssayTypes,vrAssay-method}
\alias{vrAssayTypes,vrAssayV2-method}
\title{Get assay types}
\usage{
vrAssayTypes(object, ...)

\S4method{vrAssayTypes}{VoltRon}(object, assay = NULL)

\S4method{vrAssayTypes}{vrAssay}(object)

\S4method{vrAssayTypes}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Given a VoltRon or vrAssay object, get types of assays. 
Here, an assay type is of either tile, molecule, cell, spot or ROI.
}
