% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{vrAssayV2-class}
\alias{vrAssayV2-class}
\alias{vrAssayV2}
\title{The vrAssayV2 (VoltRon Assay) Class}
\description{
The vrAssayV2 (VoltRon Assay) Class
}
\section{Slots}{

\describe{
\item{\code{data}}{list of count/normalized datasets}

\item{\code{featuredata}}{list of feature metadata}

\item{\code{embeddings}}{list of embeddings}

\item{\code{image}}{a list of vrSpatial objects}

\item{\code{params}}{additional parameters used by different assay types}

\item{\code{type}}{the type of the assay (tile, molecule, cell, spot, ROI)}

\item{\code{name}}{the assay name}

\item{\code{main_image}}{the key of the main image}

\item{\code{main_featureset}}{the key of the main feature set}
}}

