% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{vrBlock-class}
\alias{vrBlock-class}
\alias{vrBlock}
\title{The vrBlock (VoltRon Block) Class}
\description{
The vrBlock (VoltRon Block) Class
}
\section{Slots}{

\describe{
\item{\code{layer}}{A list of layers (vrLayer)}

\item{\code{zlocation}}{a vector of z coordinates of layers}
}}

