% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingFeaturePlotSingle}
\alias{vrEmbeddingFeaturePlotSingle}
\title{vrEmbeddingPlotSingle}
\usage{
vrEmbeddingFeaturePlotSingle(
  assay,
  embedding = "pca",
  metadata,
  feature = NULL,
  limits,
  font.size = 2,
  pt.size = 1,
  plot_title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{assay}{vrAssay object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{metadata}{the metadata associated with the assay}

\item{feature}{a feature, either from the rows of rawdata, normdata or columns of the metadata}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{plot_title}{the title of the single plot}
}
\description{
Plotting a single assay from a VoltRon object. We plot the identification of spatially resolved cells, spots, and ROI on associated images.
}
