% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{vrEmbeddingNames}
\alias{vrEmbeddingNames}
\alias{vrEmbeddingNames,VoltRon-method}
\alias{vrEmbeddingNames,vrAssay-method}
\alias{vrEmbeddingNames,vrAssayV2-method}
\title{Get Embedding names}
\usage{
vrEmbeddingNames(object, ...)

\S4method{vrEmbeddingNames}{VoltRon}(object, assay = NULL)

\S4method{vrEmbeddingNames}{vrAssay}(object)

\S4method{vrEmbeddingNames}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Given a VoltRon or vrAssay object, give names of embeddings
}
