% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingPlot_old}
\alias{vrEmbeddingPlot_old}
\title{vrEmbeddingPlot_old}
\usage{
vrEmbeddingPlot_old(
  object,
  embedding = "pca",
  group.by = "label",
  assay = NULL,
  assay.type = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 1,
  label = FALSE,
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{group.by}{a grouping label for the spatial entities}

\item{assay}{the assay name}

\item{assay.type}{the assay type name: 'cell', 'spot' or 'ROI'}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting embeddings of cells and spots on associated images from multiple assays in a VoltRon object.
}
