% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingPlotSingle}
\alias{vrEmbeddingPlotSingle}
\title{vrEmbeddingPlotSingle}
\usage{
vrEmbeddingPlotSingle(
  assay,
  embedding = "pca",
  metadata,
  group.by = "label",
  font.size = 2,
  pt.size = 1,
  plot_title = NULL
)
}
\arguments{
\item{assay}{vrAssay object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{metadata}{the metadata associated with the assay}

\item{group.by}{a grouping label for the spatial entities}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{plot_title}{the title of the single plot}
}
\description{
Plotting embedding of a single assay from a VoltRon object. We plot the identification of cells and spots on associated images.
}
