% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrEmbeddings.vrAssay}
\alias{vrEmbeddings.vrAssay}
\alias{vrEmbeddings<-.vrAssay}
\alias{vrEmbeddings}
\alias{vrEmbeddings<-}
\alias{vrEmbeddings.VoltRon}
\alias{vrEmbeddings<-.VoltRon}
\title{vrEmbeddings}
\usage{
\method{vrEmbeddings}{vrAssay}(object, type = "pca")

\method{vrEmbeddings}{vrAssay}(object, type = "pca", ...) <- value

vrEmbeddings(object, ...)

vrEmbeddings(object, ...) <- value

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", ..., value)

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{vrEmbeddings}: The name of the default assay

\code{vrEmbeddings}: The name of the default assay
}
\description{
Get embeddings of spatial points
}
\concept{data-access}
