% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrFeatures}
\alias{vrFeatures}
\alias{vrFeatures.VoltRon}
\alias{vrFeatures.vrAssay}
\title{vrFeatures}
\usage{
vrFeatures(object, assay = NULL)

\method{vrFeatures}{VoltRon}(object, assay = NULL)

\method{vrFeatures}{vrAssay}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Get features names
}
