% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{vrImage_to_zarr}
\alias{vrImage_to_zarr}
\title{Title}
\usage{
vrImage_to_zarr(vrimage, out_path, image_id = "main_image")
}
\arguments{
\item{vrimage}{VoltRon image}

\item{out_path}{output path to ome.zarr}

\item{image_id}{image name}
}
\description{
Title
}
