% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{vrLayer-class}
\alias{vrLayer-class}
\alias{vrLayer}
\title{The vrLayer (VoltRon Layer) Class}
\description{
The vrLayer (VoltRon Layer) Class
}
\section{Slots}{

\describe{
\item{\code{assay}}{A list of assays (vrAssay)}

\item{\code{connectivity}}{the connectivity graph}
}}

