% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{vrMainAssay}
\alias{vrMainAssay}
\alias{vrMainAssay<-}
\alias{vrMainAssay.VoltRon}
\alias{vrMainAssay<-.VoltRon}
\title{Main Assay}
\usage{
vrMainAssay(object, ...)

vrMainAssay(object, ...) <- value

\method{vrMainAssay}{VoltRon}(object, ...)

\method{vrMainAssay}{VoltRon}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{new assay name}
}
\description{
Get and set the main assay of a VoltRon object
}
