% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrMainFeatureType}
\alias{vrMainFeatureType}
\alias{vrMainFeatureType.VoltRon}
\alias{vrMainFeatureType.vrAssay}
\alias{vrMainFeatureType<-.VoltRon}
\alias{vrMainFeatureType<-.vrAssay}
\title{vrMainFeatureType}
\usage{
vrMainFeatureType(object, assay = NULL)

\method{vrMainFeatureType}{VoltRon}(object, assay = NULL)

\method{vrMainFeatureType}{vrAssay}(object)

\method{vrMainFeatureType}{VoltRon}(object, assay = NULL) <- value

\method{vrMainFeatureType}{vrAssay}(object) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
If NULL, the default assay will be used, see \link{vrMainAssay}. If given as "all", then provides a summary of spatial systems across all assays.}

\item{value}{the name of main feature set}
}
\description{
Get the main feature type
}
