% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{class}
\name{vrMetadata-class}
\alias{vrMetadata-class}
\alias{vrMetadata}
\title{The vrMetadata (VoltRon Metadata) Class}
\description{
The vrMetadata (VoltRon Metadata) Class
}
\section{Slots}{

\describe{
\item{\code{tile}}{the metadata of tiles}

\item{\code{molecule}}{the metadata of molecules}

\item{\code{cell}}{the metadata of cells}

\item{\code{spot}}{the metadata of spot}

\item{\code{ROI}}{the metadata of ROI}
}}

