% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrProportionPlot}
\alias{vrProportionPlot}
\title{vrPercentagePlot}
\usage{
vrProportionPlot(
  object,
  assay = NULL,
  x.label = NULL,
  split.by = NULL,
  split.method = "facet_wrap",
  ncol = 2,
  nrow = NULL
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{x.label}{labels of the x axis}

\item{split.by}{the column to split the barplots by}

\item{split.method}{either facet_grid or facet_wrap, not affected 
if \code{split.by} is \code{NULL}}

\item{ncol}{column wise number of plots, for \link[ggplot2]{ggarrange}}

\item{nrow}{row wise number of plots, for \link[ggplot2]{ggarrange}}
}
\description{
vrPercentagePlot
}
