% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{vrSample-class}
\alias{vrSample-class}
\alias{vrSample}
\title{The vrSample (VoltRon Sample) Class}
\description{
The vrSample (VoltRon Sample) Class
}
\section{Slots}{

\describe{
\item{\code{layer}}{A list of layers (vrLayer)}
}}

