% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrScatterPlot}
\alias{vrScatterPlot}
\title{vrScatterPlot}
\usage{
vrScatterPlot(
  object,
  feature.1,
  feature.2,
  norm = TRUE,
  assay = NULL,
  pt.size = 2,
  font.size = 2,
  group.by = "label",
  label = FALSE,
  trend = FALSE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{feature.1}{first feature}

\item{feature.2}{second feature}

\item{norm}{if TRUE, the normalize data will be used}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{pt.size}{point size}

\item{font.size}{font size}

\item{group.by}{a column from metadata to label points}

\item{label}{whether labels are visualized or not}

\item{trend}{inserting a trend line two the scatter plot}
}
\description{
get a scatter plot between two features
}
