% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R,
%   R/image.R
\name{vrSegments}
\alias{vrSegments}
\alias{vrSegments,VoltRon-method}
\alias{vrSegments,vrAssay-method}
\alias{vrSegments,vrAssayV2-method}
\alias{vrSegments,vrImage-method}
\alias{vrSegments,vrSpatial-method}
\alias{vrSegments<-,VoltRon-method}
\alias{vrSegments<-,vrAssay-method}
\alias{vrSegments<-,vrAssayV2-method}
\alias{vrSegments<-,vrImage-method}
\alias{vrSegments<-,vrSpatial-method}
\alias{vrSegments<-}
\title{vrSegments}
\usage{
vrSegments(object, ...)

\S4method{vrSegments}{VoltRon}(
  object,
  assay = NULL,
  image_name = NULL,
  spatial = NULL,
  reg = FALSE,
  as.data.frame = FALSE
)

\S4method{vrSegments}{vrAssay}(object, image_name = NULL, spatial = NULL, reg = FALSE)

\S4method{vrSegments}{vrAssayV2}(object, image_name = NULL, spatial = NULL, reg = FALSE)

\S4method{vrSegments}{vrImage}(object)

\S4method{vrSegments}{vrSpatial}(object)

\S4method{vrSegments}{VoltRon}(object, image_name = NULL, spatial = NULL, reg = FALSE) <- value

\S4method{vrSegments}{vrAssay}(object, image_name = NULL, spatial = NULL, reg = FALSE) <- value

\S4method{vrSegments}{vrAssayV2}(object, image_name = NULL, spatial = NULL, reg = FALSE) <- value

\S4method{vrSegments}{vrImage}(object) <- value

\S4method{vrSegments}{vrSpatial}(object) <- value

vrSegments(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{image_name}{(deprecated, use \code{spatial}) the name/key of the image associated with the coordinates}

\item{spatial}{the name/key of the spatial system associated with the coordinates}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainImage}) is requested}

\item{as.data.frame}{if TRUE, the coordinates of segment nodes will be returned as a data frame}

\item{value}{segments}
}
\description{
Given a VoltRon, vrAssay or vrSpatial object, get and set the segment coordinates of assays and coordinate systems
}
