% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialFeaturePlotSingle}
\alias{vrSpatialFeaturePlotSingle}
\title{vrSpatialFeaturePlotSingle}
\usage{
vrSpatialFeaturePlotSingle(
  assay,
  metadata,
  feature,
  graph = NULL,
  limits,
  group.by = "label",
  norm = TRUE,
  log = FALSE,
  font.size = 2,
  pt.size = 2,
  title.size = 10,
  alpha = 0.6,
  label = FALSE,
  plot_title = NULL,
  legend_title = NULL,
  background = NULL,
  reg = FALSE,
  crop = FALSE
)
}
\arguments{
\item{assay}{vrAssay object}

\item{metadata}{the metadata associated with the assay}

\item{feature}{a feature, either from the rows of rawdata, normdata or columns of the metadata}

\item{graph}{if not NULL, the graph is added to the plot}

\item{limits}{limits of the legend of the plot}

\item{group.by}{a grouping label for the spatial entities}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) will be log transformed}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{title.size}{title size of legend and plot}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{label}{if TRUE, labels of ROIs will be visualized too}

\item{plot_title}{the main title of the single plot}

\item{legend_title}{the legend title of the single plot}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{reg}{if TRUE, the registered coordinates will be used}

\item{crop}{whether to crop an image of a spot assay}
}
\description{
A single Spatial Feature plot of VoltRon objects
}
