% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrSpatialNames}
\alias{vrSpatialNames}
\alias{vrSpatialNames.VoltRon}
\alias{vrSpatialNames.vrAssay}
\alias{vrSpatialNames.vrAssayV2}
\title{vrSpatialNames}
\usage{
vrSpatialNames(object, assay = NULL)

\method{vrSpatialNames}{VoltRon}(object, assay = NULL)

\method{vrSpatialNames}{vrAssay}(object)

\method{vrSpatialNames}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Get names of all spatial systems
}
