% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{vrSpatialPlotBackgroundJob}
\alias{vrSpatialPlotBackgroundJob}
\title{vrSpatialPlotBackgroundJob}
\usage{
vrSpatialPlotBackgroundJob(
  host = getOption("shiny.host", "127.0.0.1"),
  plot = NULL
)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
}
\description{
Run an interactive spatial plot as a background job and show it in the viewer pane.
Adapted from \code{gptstudio} package
}
