% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{vrSpatialPlotInteractive}
\alias{vrSpatialPlotInteractive}
\title{vrSpatialPlotInteractive}
\usage{
vrSpatialPlotInteractive(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port"),
  plot_g = NULL
)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port between 3000:8000, excluding ports that are blocked
by Google Chrome for being considered unsafe: 3659, 4045, 5060,
5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random
ports will be tried.}

\item{plot_g}{the ggplot plot}
}
\value{
This function has no return value.
}
\description{
vrSpatialPlotInteractive
}
