% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialPlotSingleTiling}
\alias{vrSpatialPlotSingleTiling}
\title{vrSpatialPlotSingleTiling}
\usage{
vrSpatialPlotSingleTiling(g, data, n.tile, alpha = 1)
}
\arguments{
\item{g}{the ggplot figure}

\item{n.tile}{should points be aggregated into tiles before visualization (see \code{geom_tile}). Applicable only for cells and molecules}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{assay}{vrAssay object}

\item{metadata}{the metadata associated with the assay}

\item{group.by}{a grouping label for the spatial entities}

\item{plot.segments}{plot segments instead of points}

\item{group.ids}{a subset of categories defined with in the grouping label \code{group.by}}

\item{graph}{if not NULL, the graph is added to the plot}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{cell.shape}{the shape of the points representing cells, see \code{help(geom_point)}}

\item{plot_title}{the title of the single plot}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{reg}{if TRUE, the registered coordinates will be used}

\item{crop}{whether to crop an image of a spot assay}

\item{legend.pt.size}{the size of points at the legend}

\item{scale.image}{should the image be scaled down to a low resolution (width: 1000px)}
}
\description{
Plotting a tiled version of the vrSpatialPlot
}
