% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{vrSpatialPlotVitessce}
\alias{vrSpatialPlotVitessce}
\title{vrSpatialPlotInteractive}
\usage{
vrSpatialPlotVitessce(
  zarr.file,
  group.by = "Sample",
  plot.segments = FALSE,
  group.ids = NULL,
  assay = NULL,
  reduction = "umap",
  background = NULL,
  reg = FALSE,
  crop = FALSE
)
}
\arguments{
\item{zarr.file}{The zarr file of a VoltRon object}

\item{group.by}{a grouping label for the spatial entities}

\item{plot.segments}{plot segments instead of points}

\item{group.ids}{a subset of categories defined with in the grouping label \code{group.by}}

\item{assay}{the assay name}

\item{reduction}{The name of the reduction to visualize an embedding alongside with the spatial plot.}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{reg}{if TRUE, the registered coordinates will be used}

\item{crop}{whether to crop an image of a spot assay}
}
\description{
Interactive Plotting identification of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
