% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{warpSimpleITKImage}
\alias{warpSimpleITKImage}
\title{getRcppWarpImage}
\usage{
warpSimpleITKImage(query_image, mapping)
}
\arguments{
\item{query_image}{query image}

\item{mapping}{a list of the homography matrices and TPS keypoints}
}
\description{
Warping a query image onto a reference image using SimpleITK.
}
