% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xenium_data}
\alias{xenium_data}
\title{Example Xenium Data}
\format{
An object of class \code{VoltRon} of length 1.
}
\source{
{Created to serve as an example.}
}
\usage{
xenium_data
}
\description{
This VoltRon object is used for testing and as an example
}
\examples{
data(xenium_data)
}
\keyword{datasets}
