% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisk.R
\name{zarrcreateGroup}
\alias{zarrcreateGroup}
\title{zarrcreateGroup}
\usage{
zarrcreateGroup(store, name)
}
\arguments{
\item{store}{the location of (zarr) store}

\item{name}{name of the group}
}
\description{
get information of an ImgArray object
}
