% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/metadata.R, R/objects.R
\name{addAssay}
\alias{addAssay}
\alias{addAssay,vrMetadata-method}
\alias{addAssay,VoltRon-method}
\title{Add Assay}
\usage{
addAssay(object, ...)

\S4method{addAssay}{vrMetadata}(
  object,
  metadata = NULL,
  assay,
  assay_name,
  sample = "Sample1",
  layer = "Section1"
)

\S4method{addAssay}{VoltRon}(
  object,
  assay,
  metadata = NULL,
  assay_name,
  sample = "Sample1",
  layer = "Section1"
)
}
\arguments{
\item{object}{a VoltRon object.}

\item{...}{arguments passed to other methods.}

\item{metadata}{a predefined metadata}

\item{assay}{a vrAssay object}

\item{assay_name}{assay name of the new added assay}

\item{sample}{sample name}

\item{layer}{layer name}
}
\description{
add a new assay (vrAssay object) to a VoltRon object
}
