% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metadata.R, R/objects.R
\name{AddAssay}
\alias{AddAssay}
\alias{AddAssay.vrMetadata}
\alias{AddAssay.VoltRon}
\title{Add Assay}
\usage{
AddAssay(object, ...)

\method{AddAssay}{vrMetadata}(object, assay, assay_name, sample = "Sample1", layer = "Section1")

\method{AddAssay}{VoltRon}(object, assay, assay_name, sample = "Sample1", layer = "Section1")
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{AddAssay}: The name of the default assay
}
\description{
add assay to the object
}
\concept{data-access}
