% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metadata.R, R/objects.R
\name{addAssay}
\alias{addAssay}
\alias{addAssay.vrMetadata}
\alias{addAssay.VoltRon}
\title{Add Assay}
\usage{
addAssay(object, ...)

\method{addAssay}{vrMetadata}(
  object,
  metadata = NULL,
  assay,
  assay_name,
  sample = "Sample1",
  layer = "Section1"
)

\method{addAssay}{VoltRon}(
  object,
  assay,
  metadata = NULL,
  assay_name,
  sample = "Sample1",
  layer = "Section1"
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{metadata}{a predefined metadata}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{assay_name}{assay name}

\item{sample}{sample name}

\item{layer}{layer name}
}
\description{
add assay to the object
}
