% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{addAssay}
\alias{addAssay}
\title{Add Assay}
\usage{
addAssay(object, ...)
}
\arguments{
\item{object}{a VoltRon object.}

\item{...}{arguments passed to other methods.}
}
\description{
add a new assay (vrAssay object) to a VoltRon object
}
