% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{addAssay}
\alias{addAssay}
\title{Add Assay}
\usage{
addAssay(
  object,
  assay,
  metadata = NULL,
  assay_name,
  sample = "Sample1",
  layer = "Section1"
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{a vrAssay object}

\item{metadata}{a predefined metadata}

\item{assay_name}{assay name of the new added assay}

\item{sample}{sample name}

\item{layer}{layer name}
}
\description{
add a new assay (vrAssay object) to a VoltRon object
}
