% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{addFeature}
\alias{addFeature}
\title{addFeature}
\usage{
addFeature(object, assay = NULL, data, feature_name)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}
}
\description{
add a new feature set to a vrAssay (or VoltRon) object
}
