% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{addFOVImage}
\alias{addFOVImage}
\title{addFOVImage}
\usage{
addFOVImage(seu, file, fov = "fov", overwrite = FALSE)
}
\arguments{
\item{seu}{Seurat Object with Xenium Data}

\item{file}{the morphology image file created by \code{generateXeniumImage}.}

\item{fov}{FOV name, preferably the name used in the Seurat Object}

\item{overwrite}{Overwrite the existing FOV image}
}
\description{
Adding the Xenium image to the Seurat Object. Please run \code{generateXeniumImage} first.
}
