% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateSpatialData}
\alias{annotateSpatialData}
\title{annotateSpatialData}
\usage{
annotateSpatialData(
  object,
  label = "annotation",
  assay = NULL,
  annotation_assay = "ROIAnnotation",
  use.image.only = FALSE,
  shiny.options = list(launch.browser = getOption("shiny.launch.browser", interactive())),
  spatial = NULL,
  channel = NULL,
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{label}{the name of the new metadata column (default: annotation) annotating spatial points by selected polygons}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{annotation_assay}{name of the annotation assay ()}

\item{use.image.only}{if TRUE, use only the image}

\item{shiny.options}{a list of shiny options (launch.browser, host, port etc.) passed \code{options} arguement of \link{shinyApp}. For more information, see \link{runApp}}

\item{spatial}{the name of the spatial coordinate system}

\item{channel}{the name of the main channel}

\item{...}{additional parameters passed to \link{vrSpatialPlot}.}
}
\description{
A mini shiny app to for annotating spatial points
}
\examples{
\dontrun{
# Annotate based on images
visium_data <- annotateSpatialData(visium_data, use.image.only = TRUE)

# Annotate based on spatial plot
xenium_data <- annotateSpatialData(xenium_data, group.by = "clusters")
}
}
