% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateSpatialData}
\alias{annotateSpatialData}
\title{annotateSpatialData}
\usage{
annotateSpatialData(
  object,
  label = "annotation",
  assay = NULL,
  use.image = FALSE,
  image_name = NULL,
  channel = NULL,
  ...
)
}
\arguments{
\item{object}{a list of VoltRon (or Seurat) objects}

\item{label}{the name of the new metadata feature (annotation) of selected spatial points}

\item{assay}{a reference spatial data set, used only if \code{object_list} is \code{NULL}}

\item{use.image}{if TRUE, use only the image}

\item{channel}{the name of the main channel}

\item{...}{additional parameters passed to \code{vrSpatialPlot}.}

\item{name}{the name of the main image}
}
\value{
a vector of annotations
}
\description{
A mini shiny app to for annotating spatial points
}
