% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.AnnData}
\alias{as.AnnData}
\title{as.AnnData}
\usage{
as.AnnData(
  object,
  file,
  assay = NULL,
  type = c("image", "spatial"),
  flip_coordinates = FALSE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{file}{the name of the h5ad file}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}

\item{flip_coordinates}{if TRUE, the spatial coordinates (including segments) will be flipped}
}
\description{
Converting a VoltRon object into a AnnData (.h5ad) object
}
