% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.AnnData}
\alias{as.AnnData}
\alias{as.AnnData.VoltRon}
\title{as.AnnData}
\usage{
as.AnnData(object, ...)

\method{as.AnnData}{VoltRon}(
  object,
  file,
  assay = NULL,
  image_key = "fov",
  type = c("image", "spatial")
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{file}{the name of the h5ad file}

\item{assay}{the name(type) of the assay to be converted}

\item{image_key}{the name (or prefix) of the image(s)}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}
}
\description{
Generic methods for conversion into a AnnData object
}
