% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.Giotto}
\alias{as.Giotto}
\title{as.Giotto}
\usage{
as.Giotto(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Generic methods for conversion into a Giotto object
}
