% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.OmeTiff}
\alias{as.OmeTiff}
\title{as.OmeTiff}
\usage{
as.OmeTiff(object, out_path, image_id = "image_1", python.path = NULL)
}
\arguments{
\item{object}{a magick-image object}

\item{out_path}{output path to ome.tiff file}

\item{image_id}{image name}

\item{python.path}{the path to the python binary, otherwise either \code{basilisk} package is used or \code{getOption("voltron.python.path")} should be not NULL.}
}
\description{
Converting VoltRon (magick) images to ome.tiff
}
