% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.Seurat}
\alias{as.Seurat}
\alias{as.Seurat.VoltRon}
\title{as.Seurat}
\usage{
as.Seurat(object, ...)

\method{as.Seurat}{VoltRon}(
  object,
  assay = NULL,
  image_key = "fov",
  type = c("image", "spatial")
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{the name(type) of the assay to be converted}

\item{image_key}{the name (or prefix) of the image(s)}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}
}
\value{
\code{as.Seurat}: The name of the default assay
}
\description{
Generic methods for conversion into a Seurat object
}
