% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.Seurat}
\alias{as.Seurat}
\title{as.Seurat}
\usage{
as.Seurat(
  object,
  cell.assay = NULL,
  molecule.assay = NULL,
  image_key = "fov",
  type = c("image", "spatial"),
  reg = FALSE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{cell.assay}{the name(type) of the cell assay to be converted}

\item{molecule.assay}{the name(type) of the molecule assay to be added to 
the cell assay in Seurat object}

\item{image_key}{the name (or prefix) of the image(s)}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}

\item{reg}{if TRUE, registered coordinates will be used}
}
\description{
Converting a VoltRon object into a Seurat object
}
