% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.Seurat}
\alias{as.Seurat}
\alias{as.Seurat.VoltRon}
\title{as.Seurat}
\usage{
as.Seurat(object, ...)

\method{as.Seurat}{VoltRon}(object, image = "fov")
}
\arguments{
\item{object}{A VoltRon object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{as.Seurat}: The name of the default assay
}
\description{
Generic methods for conversion into a Seurat object
}
