% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.VoltRon}
\alias{as.VoltRon}
\alias{as.VoltRon.Seurat}
\alias{as.VoltRon.Giotto}
\title{as.VoltRon}
\usage{
as.VoltRon(object, ...)

\method{as.VoltRon}{Seurat}(object)

\method{as.VoltRon}{Giotto}(object)
}
\arguments{
\item{object}{A Giotto object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{as.VoltRon}: The name of the default assay
}
\description{
Generic methods for conversion into a VoltRon object
}
