% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.VoltRon}
\alias{as.VoltRon}
\alias{as.VoltRon.Seurat}
\title{as.VoltRon}
\usage{
as.VoltRon(object, ...)

\method{as.VoltRon}{Seurat}(object, type = c("image", "spatial"), assay_name = NULL, ...)
}
\arguments{
\item{object}{a Seurat object}

\item{...}{Additional parameter passed to \link{formVoltRon}}

\item{type}{the spatial data type of Seurat object: "image" or "spatial"}

\item{assay_name}{the assay name}
}
\description{
Generic methods for conversion into a VoltRon object
}
