% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/conversion.R
\name{as.Zarr}
\alias{as.Zarr}
\alias{as.Zarr.magick-image}
\title{as.Zarr}
\usage{
as.Zarr(object, out_path, image_id)

\method{as.Zarr}{`magick-image`}(object, out_path, image_id = "image_1")
}
\arguments{
\item{object}{a VoltRon or magick-image object}

\item{out_path}{output path to zarr file}

\item{image_id}{image name}
}
\description{
Generic methods to save VoltRon or magick-image objects as zarr files
}
