% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/conversion.R
\name{as.Zarr}
\alias{as.Zarr}
\alias{as.Zarr.VoltRon}
\alias{as.Zarr.magick-image}
\title{as.Zarr}
\usage{
as.Zarr(object, out_path, image_id, ...)

\method{as.Zarr}{VoltRon}(object, out_path, image_id = "main_image")

\method{as.Zarr}{`magick-image`}(object, out_path, image_id = "main_image")
}
\arguments{
\item{object}{a VoltRon image}

\item{out_path}{output path to ome.zarr}

\item{image_id}{image name}

\item{...}{Arguments passed to other methods}
}
\description{
Generic methods for conversion into a Seurat object
}
