% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{AssayNames}
\alias{AssayNames}
\alias{AssayNames.SpaceRover}
\title{Get Assay names}
\usage{
AssayNames(object, ...)

\method{AssayNames}{SpaceRover}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{AssayNames}: The name of the default assay
}
\description{
Given a SpaceRover object, give names of assays of some type, or main assay.
}
\concept{data-access}
