% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{AssayTypes.srAssay}
\alias{AssayTypes.srAssay}
\alias{AssayTypes}
\alias{AssayTypes.SpaceRover}
\title{Get Assay types}
\usage{
\method{AssayTypes}{srAssay}(object, ...)

AssayTypes(object, ...)

\method{AssayTypes}{SpaceRover}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{AssayTypes}: The name of the default assay
}
\description{
Given a SpaceRover object, give types of assays of some type, or of the main assay.
Here, an assay type is of either cell, spot or ROI.
}
\concept{data-access}
