% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{combineChannels}
\alias{combineChannels}
\alias{combineChannels.VoltRon}
\alias{combineChannels.vrAssay}
\alias{combineChannels.vrImage}
\title{combineChannels}
\usage{
combineChannels(object, ...)

\method{combineChannels}{VoltRon}(object, ...)

\method{combineChannels}{vrAssay}(object, name = NULL, reg = FALSE, ...)

\method{combineChannels}{vrImage}(
  object,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)
}
\arguments{
\item{object}{An object}

\item{...}{arguements passed to \code{combineChannels.vrImage}}

\item{name}{the name of the main image}

\item{reg}{TRUE if registered images are assigned}

\item{channels}{the list of channels}

\item{colors}{the colors associated with each channel}

\item{channel_key}{the name of the new channel name}
}
\description{
Combining channels into novel channels of the same image
}
