% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{combineChannels}
\alias{combineChannels}
\alias{combineChannels.VoltRon}
\alias{combineChannels.vrAssay}
\alias{combineChannels.vrImage}
\title{combineChannels}
\usage{
combineChannels(
  object,
  assay = NULL,
  name = NULL,
  reg = FALSE,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)

\method{combineChannels}{VoltRon}(object, assay = NULL, ...)

\method{combineChannels}{vrAssay}(object, name = NULL, reg = FALSE, ...)

\method{combineChannels}{vrImage}(
  object,
  channels = NULL,
  colors = NULL,
  channel_key = "combined"
)
}
\arguments{
\item{object}{A VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the name of the image}

\item{reg}{TRUE if registered coordinates of the main image (\code{vrMainImage}) is requested}

\item{channels}{the name of the channel associated with the image}

\item{colors}{the colors associated with each channel}

\item{channel_key}{the name of the new channel name}

\item{...}{arguements passed to other methods}
}
\description{
Combining channels into novel channels of the same image
}
