% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{computeManualPairwiseTransform}
\alias{computeManualPairwiseTransform}
\title{computeManualPairwiseTransform}
\usage{
computeManualPairwiseTransform(image_list, keypoints_list, query_ind, ref_ind)
}
\arguments{
\item{image_list}{the list of images}

\item{keypoints_list}{the list of keypoint matrices}

\item{query_ind}{the index of the query image}

\item{ref_ind}{the index of the reference image}
}
\description{
Computing transformation matrix of manual registration
}
