% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Coordinates.vrAssay}
\alias{Coordinates.vrAssay}
\alias{Coordinates<-.vrAssay}
\alias{Coordinates}
\alias{Coordinates<-}
\alias{Coordinates.VoltRon}
\alias{Coordinates<-.VoltRon}
\title{Coordinates}
\usage{
\method{Coordinates}{vrAssay}(object, reg = FALSE, ...)

\method{Coordinates}{vrAssay}(object, reg = FALSE, ...) <- value

Coordinates(object, ...)

Coordinates(object, ...) <- value

\method{Coordinates}{VoltRon}(object, reg = FALSE, assay = NULL, ...)

\method{Coordinates}{VoltRon}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Coordinates}: The name of the default assay

\code{Coordinates<-}: The coordinates updated
}
\description{
Get and set the coordinates of the main assay
}
\concept{data-access}
