% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Coordinates.srAssay}
\alias{Coordinates.srAssay}
\alias{Coordinates<-.srAssay}
\alias{Coordinates}
\alias{Coordinates<-}
\alias{Coordinates.SpaceRover}
\alias{Coordinates<-.SpaceRover}
\title{Coordinates}
\usage{
\method{Coordinates}{srAssay}(object, ...)

\method{Coordinates}{srAssay}(object, ...) <- value

Coordinates(object, ...)

Coordinates(object, ...) <- value

\method{Coordinates}{SpaceRover}(object, ...)

\method{Coordinates}{SpaceRover}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Coordinates}: The name of the default assay

\code{Coordinates<-}: The coordinates updated
}
\description{
Get and set the coordinates of the main assay
}
\concept{data-access}
