% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSpaceRover}
\alias{CreateSpaceRover}
\title{CreateSpaceRover}
\usage{
CreateSpaceRover(
  samples,
  metadata = NULL,
  sample.metadata = NULL,
  zstack = NULL,
  main.assay = NULL,
  project = NULL
)
}
\arguments{
\item{samples}{a list of samples list of class \code{srSamples}}

\item{metadata}{a metadata object of class \code{srMetadata}}

\item{sample.metadata}{a data frame of the sample metadata}

\item{zstack}{the zstack graph to determine the adjacency of spatial entities across layers}

\item{main.assay}{the name of the main assay of the object}

\item{project}{project name}
}
\description{
Create a SpaceRover object
}
