% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{Data.srAssay}
\alias{Data.srAssay}
\alias{Data}
\alias{Data.SpaceRover}
\title{Data}
\usage{
\method{Data}{srAssay}(object, ...)

Data(object, ...)

\method{Data}{SpaceRover}(object, type = NULL, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Data}: The name of the default assay
}
\description{
Get data from the main.assay
}
\concept{data-access}
