% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{demuxVoltRon}
\alias{demuxVoltRon}
\title{demuxVoltRon}
\usage{
demuxVoltRon(object, scale_width = 800, use_points = FALSE)
}
\arguments{
\item{object}{a VoltRon object}

\item{scale_width}{the initial width of the object image}
}
\description{
Subsetting/demultiplexing of the VoltRon Object using interactive shiny app
}
