% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{demuxVoltRon}
\alias{demuxVoltRon}
\title{demuxVoltRon}
\usage{
demuxVoltRon(
  object,
  max.pixel.size = 1200,
  use.points.only = FALSE,
  shiny.options = list(launch.browser = getOption("shiny.launch.browser", interactive()))
)
}
\arguments{
\item{object}{a VoltRon object}

\item{max.pixel.size}{the initial width of the object image}

\item{use.points.only}{use spatial points instead of the reference image}

\item{shiny.options}{a list of shiny options (launch.browser, host, 
port etc.) passed \code{options} arguement of \link{shinyApp}. 
For more information, see \link{runApp}}
}
\description{
Subsetting/demultiplexing of the VoltRon Object using interactive shiny app
}
