% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R
\name{Distances.srAssay}
\alias{Distances.srAssay}
\alias{Distances}
\title{Distances}
\usage{
\method{Distances}{srAssay}(object, reg = FALSE, method = "euclidean", ...)

Distances(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Distances}: The name of the default assay
}
\description{
Get distances between spatial points using coordinates
}
\concept{data-access}
