% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metadata.R, R/objects.R
\name{Entities}
\alias{Entities}
\alias{Entities.srMetadata}
\alias{Entities.SpaceRover}
\title{Entities}
\usage{
Entities(object, ...)

\method{Entities}{srMetadata}(object, ...)

\method{Entities}{SpaceRover}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Entities}: The name of the default assay
}
\description{
Get and set spatial entities (cells, spots, ROI)
}
\concept{data-access}
