% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/metadata.R,
%   R/objects.R, R/sample.R
\name{Entities.vrAssay}
\alias{Entities.vrAssay}
\alias{Entities}
\alias{Entities.vrMetadata}
\alias{Entities.VoltRon}
\alias{Entities.vrSample}
\alias{Entities.vrLayer}
\title{Entities}
\usage{
\method{Entities}{vrAssay}(object, ...)

Entities(object, ...)

\method{Entities}{vrMetadata}(object, ...)

\method{Entities}{VoltRon}(object, ...)

\method{Entities}{vrSample}(object, ...)

\method{Entities}{vrLayer}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Entities}: The name of the default assay
}
\description{
Get and set spatial entities (cells, spots, ROI)
}
\concept{data-access}
