% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R
\name{FeatureData.vrAssay}
\alias{FeatureData.vrAssay}
\alias{FeatureData<-.vrAssay}
\alias{FeatureData}
\alias{FeatureData<-}
\title{Feature Data}
\usage{
\method{FeatureData}{vrAssay}(object, ...)

\method{FeatureData}{vrAssay}(object, ...) <- value

FeatureData(object, ...)

FeatureData(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{FeatureData}: The name of the default assay

\code{FeatureData<-}: The coordinates updated
}
\description{
Get and set feature data from the main.assay
}
\concept{data-access}
