% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{flipCoordinates.vrAssay}
\alias{flipCoordinates.vrAssay}
\alias{flipCoordinates}
\alias{flipCoordinates.VoltRon}
\title{flipCoordinates}
\usage{
\method{flipCoordinates}{vrAssay}(object, image_name = NULL, ...)

flipCoordinates(object, ...)

\method{flipCoordinates}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{image_name}{the key of the image}

\item{...}{Arguments passed to other methods}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Flip the coordinates of the spatial points in the y axis direction
}
