% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{flipCoordinates}
\alias{flipCoordinates}
\alias{flipCoordinates.VoltRon}
\alias{flipCoordinates.vrAssay}
\title{flipCoordinates}
\usage{
flipCoordinates(object, assay = NULL, image_name = NULL, ...)

\method{flipCoordinates}{VoltRon}(object, assay = NULL, image_name = NULL, ...)

\method{flipCoordinates}{vrAssay}(object, image_name = NULL, ...)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{image_name}{the name/key of the image}

\item{...}{additional parameters passed to \code{vrCoordinates} and \code{vrSegments}}
}
\description{
Flip the coordinates of the spatial points in the y axis direction
}
