% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{flipCoordinates}
\alias{flipCoordinates}
\alias{flipCoordinates.VoltRon}
\alias{flipCoordinates.vrAssay}
\title{flipCoordinates}
\usage{
flipCoordinates(
  object,
  assay = NULL,
  spatial_name = NULL,
  image_name = NULL,
  ...
)

\method{flipCoordinates}{VoltRon}(
  object,
  assay = NULL,
  image_name = NULL,
  spatial_name = NULL,
  ...
)

\method{flipCoordinates}{vrAssay}(object, image_name = NULL, spatial_name = NULL, ...)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{spatial_name}{the name/key of the spatial system associated with the coordinates}

\item{image_name}{(deprecated) the name/key of the image}

\item{...}{additional parameters passed to \link{vrCoordinates} and \link{vrSegments}}
}
\description{
Flip the coordinates of spatial points in the y axis direction.
}
