% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{flipCoordinates.vrAssay}
\alias{flipCoordinates.vrAssay}
\alias{flipCoordinates}
\alias{flipCoordinates.VoltRon}
\title{flipCoordinates}
\usage{
\method{flipCoordinates}{vrAssay}(object, ...)

flipCoordinates(object, ...)

\method{flipCoordinates}{VoltRon}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{flipCoordinates}: The name of the default assay
}
\description{
Flip the coordinates of the spatial points in the y axis direction
}
