% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{formAssay}
\alias{formAssay}
\title{formAssay}
\usage{
formAssay(
  data = NULL,
  coords = NULL,
  segments = list(),
  image = NULL,
  params = list(),
  type = "ROI",
  name = "Assay1",
  main_spatial = "image_1",
  main_image = NULL,
  main_featureset = NULL,
  assay_version = "v2",
  ...
)
}
\arguments{
\item{data}{the feature matrix of spatialpoints}

\item{coords}{the coordinates of the spatial points}

\item{segments}{the list of segments each associated with a spatial point (optional)}

\item{image}{a singleton or list of images as magick-image objects}

\item{params}{additional parameters of the object}

\item{type}{the type of the assay (tile, molecule, cell, spot or ROI)}

\item{name}{the name of the assay}

\item{main_spatial}{the name of the main spatial coordinate system}

\item{main_image}{(deprecated) the name of the main spatial coordinate system}

\item{main_featureset}{the name of the main_featureset}

\item{assay_version}{the assay version}

\item{...}{additional arguments passed to \link{formSpatial}}
}
\description{
Create a vrAssay (VoltRon assay) object
}
