% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{formImage}
\alias{formImage}
\title{formImage}
\usage{
formImage(coords, segments = list(), image = NULL, main_channel = NULL)
}
\arguments{
\item{coords}{the coordinates of the spatial points}

\item{segments}{the segments of the spatial points, optional}

\item{image}{the image of the data}

\item{main_channel}{the key of the main channel of vrImage object}
}
\description{
Create a vrImage (VoltRon image) object
}
