% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{formSpatial}
\alias{formSpatial}
\title{formSpatial}
\usage{
formSpatial(
  coords = NULL,
  segments = list(),
  image = NULL,
  main_channel = NULL
)
}
\arguments{
\item{coords}{the coordinates of the spatial points}

\item{segments}{the list of segments each associated with a spatial point}

\item{image}{a singelton or list of images as magick-image objects}

\item{main_channel}{the key of the main channel of vrImage object}
}
\description{
Create a vrImage (VoltRon image) object
}
