% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{formVoltRon}
\alias{formVoltRon}
\title{formVoltRon}
\usage{
formVoltRon(
  data,
  metadata = NULL,
  image = NULL,
  coords,
  segments = list(),
  sample.metadata = NULL,
  graph = NULL,
  main.assay = "Custom_cell",
  assay.type = "cell",
  params = list(),
  sample_name = NULL,
  layer_name = NULL,
  project = NULL,
  ...
)
}
\arguments{
\item{data}{the count table}

\item{metadata}{a metadata object of class \code{vrMetadata}}

\item{image}{the image of the data}

\item{segments}{the segments of the spatial points, optional}

\item{sample.metadata}{a data frame of the sample metadata}

\item{graph}{the graph to determine the adjacency of spatial points across layers}

\item{main.assay}{the name of the main assay of the object}

\item{assay.type}{the type of the assay (cells, spots, ROIs)}

\item{params}{additional parameters of the object}

\item{sample_name}{the name of the sample}

\item{layer_name}{the name of the layer}

\item{project}{project name}

\item{...}{additional parameters passed to VoltRon object}

\item{coord}{the coordinates of the spatial points}

\item{assay_name}{the name of the assay}
}
\description{
Create a VoltRon object
}
