% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{formVoltRon}
\alias{formVoltRon}
\title{formVoltRon}
\usage{
formVoltRon(
  data = NULL,
  metadata = NULL,
  image = NULL,
  coords,
  segments = list(),
  sample.metadata = NULL,
  main.assay = NULL,
  assay.type = "cell",
  params = list(),
  sample_name = NULL,
  layer_name = NULL,
  image_name = NULL,
  feature_name = NULL,
  project = NULL,
  version = "v2",
  ...
)
}
\arguments{
\item{data}{the feature matrix of spatialpoints}

\item{metadata}{a metadata object of class \link{vrMetadata}}

\item{image}{a singelton or list of images as magick-image objects}

\item{coords}{the coordinates of the spatial points}

\item{segments}{the list of segments each associated with a spatial point}

\item{sample.metadata}{a data frame of the sample metadata, see \link{SampleMetadata}}

\item{main.assay}{the name of the main assay}

\item{assay.type}{the type of the assay (tile, molecule, cell, spot or ROI)}

\item{params}{additional parameters}

\item{sample_name}{the name of the sample}

\item{layer_name}{the name of the layer}

\item{image_name}{the name/key of the image}

\item{feature_name}{the name/key of the feature set}

\item{project}{project name}

\item{version}{the assay version, V1 or V2}

\item{...}{additional parameters passed to \link{formAssay}}
}
\description{
Create a VoltRon object
}
