% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateCosMxImage}
\alias{generateCosMxImage}
\title{generateCosMxImage}
\usage{
generateCosMxImage(dir.path, increase.contrast = TRUE, output.path = NULL, ...)
}
\arguments{
\item{dir.path}{CosMx output folder}

\item{increase.contrast}{increase the contrast of the image before writing}

\item{output.path}{The path to the new morphology image created if the image should be saved to a location other than Xenium output folder.}

\item{...}{additional parameters passed to the \link{writeImage} function}
}
\description{
Generates a low resolution Morphology image of the CosMx experiment
}
