% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{generateSegments}
\alias{generateSegments}
\title{generateSegments}
\usage{
generateSegments(object, type = c("ROI", "cell"))
}
\arguments{
\item{object}{an object of sf class with each element having a geometry}

\item{type}{spatial entity type: ROI or cell}
}
\description{
The function to import segments from a geojson file
}
