% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{generateTileData}
\alias{generateTileData}
\alias{generateTileData.VoltRon}
\alias{generateTileData.vrAssay}
\alias{generateTileData.vrAssayV2}
\title{generateTileData}
\usage{
generateTileData(
  object,
  assay = NULL,
  name = NULL,
  reg = FALSE,
  channel = NULL
)

\method{generateTileData}{VoltRon}(object, assay = NULL, ...)

\method{generateTileData}{vrAssay}(object, name = NULL, reg = FALSE, channel = NULL)

\method{generateTileData}{vrAssayV2}(object, name = NULL, reg = FALSE, channel = NULL)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the name of the main spatial system}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainSpatial}) is requested}

\item{channel}{the name of the channel associated with the image}

\item{...}{additional parameters passed to \link{generateTileData.vrAssay}}
}
\description{
Generating data matrices for tile-based VoltRon objects from images
}
