% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getAutomatedRegisteration}
\alias{getAutomatedRegisteration}
\title{getManualRegisteration}
\usage{
getAutomatedRegisteration(
  registered_spatdata_list,
  spatdata_list,
  image_list,
  input,
  output,
  session
)
}
\arguments{
\item{registered_spatdata_list}{a list of registered Spatial data object of the query images, updated with image registration}

\item{spatdata_list}{a list of Spatial data object of the query images}

\item{image_list}{the list of query images}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Manuel registeration of images using manually entered keypoints
}
