% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClusters}
\alias{getClusters}
\title{getClusters}
\usage{
getClusters(
  object,
  resolution = 1,
  assay = NULL,
  label = "clusters",
  graph = "kNN",
  seed = 1
)
}
\arguments{
\item{object}{a VoltRon object}

\item{resolution}{the resolution parameter for leiden clustering}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{label}{the name for the newly created clustering column in the metadata}

\item{graph}{the graph type to be used}

\item{seed}{seed}
}
\description{
Get clustering of the VoltRon object
}
