% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClusters}
\alias{getClusters}
\title{}
\usage{
getClusters(object, resolution = 1, assay = NULL, label = "clusters")
}
\arguments{
\item{object}{A VoltRon object}

\item{resolution}{the resolution parameter for leiden clustering}

\item{assay}{assay}

\item{label}{the name for the newly created clustering column in the metadata}
}
\description{
# sample metadata
  sample.metadata <- SampleMetadata(object)
}
\details{

}
