% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures.VoltRon}
\alias{getFeatures.vrAssay}
\title{Get spatially variable feature}
\usage{
getFeatures(object, assay = NULL, max.count = 1, n = 3000)

\method{getFeatures}{VoltRon}(object, assay = NULL, ...)

\method{getFeatures}{vrAssay}(object, max.count = 1, n = 3000)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{max.count}{maximum count (across spatial points) for low count filtering}

\item{n}{the top number of variable features}

\item{...}{arguements passed to other methods}
}
\description{
Get spatially variable features in an assay
}
