% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/processing.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures.VoltRon}
\alias{getFeatures.vrAssay}
\alias{getFeatures.vrAssayV2}
\title{getFeatures}
\usage{
getFeatures(object, assay = NULL, max.count = 1, n = 3000)

\method{getFeatures}{VoltRon}(object, assay = NULL, max.count = 1, n = 3000)

\method{getFeatures}{vrAssay}(object, max.count = 1, n = 3000)

\method{getFeatures}{vrAssayV2}(object, max.count = 1, n = 3000)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{max.count}{maximum count (across spatial points) for low count filtering}

\item{n}{the top number of variable features}
}
\description{
Get variable features of assays
}
