% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures.VoltRon}
\alias{getFeatures.vrAssay}
\title{Get spatially variable feature}
\usage{
getFeatures(object, ...)

\method{getFeatures}{VoltRon}(object, assay = NULL, ...)

\method{getFeatures}{vrAssay}(object, max.count = 1, n = 3000)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{max.count}{max count (maximum across spatial points) for low count filtering}

\item{n}{the number of features}
}
\description{
get spatially variable features in an assay
}
