% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getHotSpotAnalysis}
\alias{getHotSpotAnalysis}
\title{getHotSpotAnalysis}
\usage{
getHotSpotAnalysis(
  object,
  assay = NULL,
  method = "Getis-Ord",
  features,
  graph.type = NULL,
  group.ids = NULL,
  alpha.value = 0.01,
  norm = TRUE,
  n.tile = 0,
  verbose = TRUE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{method}{the statistical method of conducting hot spot analysis. Default is "Getis-Ord"}

\item{features}{a set of features to be visualized, either from \link{vrFeatures} of raw or normalized data or columns of the \link{Metadata}.}

\item{graph.type}{the type of graph to determine spatial neighborhood}

\item{group.ids}{a subset of categories defined in metadata column from \code{features}, 
throws an error if the feature does not include}

\item{alpha.value}{the alpha value for the hot spot analysis test. Default is 0.01}

\item{norm}{if TRUE, the normalized data is used}

\item{n.tile}{should points be rasterized along x-and y-axes, typically used for molecule assays to generate features for Getis-Ord statistics}

\item{verbose}{verbose}
}
\description{
Conduct hot spot detection
}
