% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{getImage}
\alias{getImage}
\alias{getImage.Seurat}
\alias{getImage.SpaceRover}
\title{getImage}
\usage{
getImage(object, ...)

\method{getImage}{Seurat}(seu, ...)

\method{getImage}{SpaceRover}(sr, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{getImage}: The name of the default assay
}
\description{
Get and set images of a spaceRover assay
}
\concept{data-access}
