% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getManualRegisteration}
\alias{getManualRegisteration}
\title{getManualRegisteration}
\usage{
getManualRegisteration(
  ManualRegisterMatrix_list,
  ManualRegisterCells_list,
  len_images,
  image_list,
  ref_image,
  cells_data_list,
  xyTable_list,
  xyTable_ref,
  input,
  output,
  session
)
}
\arguments{
\item{ManualRegisterMatrix_list}{a list of registration matrices of each query image}

\item{ManualRegisterCells_list}{a list of registered cells/barcodes of each query image}

\item{len_images}{length of query images}

\item{image_list}{the list of query images}

\item{ref_image}{reference image}

\item{cells_data_list}{the list of cells/barcodes locations for each image}

\item{xyTable_list}{a list of keypoints x,y coordinates for query image}

\item{xyTable_ref}{the keypoints x,y coordinates of the reference image}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Manuel registeration of images using manually entered keypoints
}
