% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getManualRegisteration}
\alias{getManualRegisteration}
\title{getManualRegisteration}
\usage{
getManualRegisteration(
  registration_mapping_list,
  spatdata_list,
  image_list,
  keypoints_list,
  centre,
  register_ind,
  input,
  output,
  session
)
}
\arguments{
\item{registration_mapping_list}{a list of mapping matrices used for registering VoltRon objects}

\item{spatdata_list}{a list of Spatial data object of the query images}

\item{image_list}{the list of query images}

\item{keypoints_list}{a list of keypoints x,y coordinates for query image}

\item{centre}{center image index}

\item{register_ind}{query image indices}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Manual registeration of images using manually entered keypoints
}
