% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getManualRegisteredImage}
\alias{getManualRegisteredImage}
\title{}
\usage{
getManualRegisteredImage(query_image, ref_image, transmatrix)
}
\arguments{
\item{query_image}{query image}

\item{ref_image}{reference image}

\item{transmatrix}{the transformation matrix}
}
\description{

}
\details{

}
