% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/clustering.R
\name{getNeighbors}
\alias{getNeighbors}
\alias{getNeighbors.VoltRon}
\alias{getNeighbors.vrAssay}
\title{Get neighbors}
\usage{
getNeighbors(object, ...)

\method{getNeighbors}{VoltRon}(object, assay = NULL, data.type = "pca", ...)

\method{getNeighbors}{vrAssay}(object, data.type = "pca", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
get neighbors in an assay
}
\concept{clustering}
