% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/clustering.R
\name{getNeighbors}
\alias{getNeighbors}
\alias{getNeighbors.SpaceRover}
\alias{getNeighbors.srAssay}
\title{Get neighbors}
\usage{
getNeighbors(object, ...)

\method{getNeighbors}{SpaceRover}(object, assay = NULL, assay.type = NULL, ...)

\method{getNeighbors}{srAssay}(object, assay = NULL, assay.type = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
get neighbors in an assay
}
\concept{clustering}
