% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getPCA}
\alias{getPCA}
\alias{getPCA.VoltRon}
\title{getPCA}
\usage{
getPCA(object, ...)

\method{getPCA}{VoltRon}(object, assay = NULL, dims = 30)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{dims}{the number of dimensions extracted from PCA}
}
\description{
calculate getPCA of the VoltRon objects
}
