% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getPCA}
\alias{getPCA}
\alias{getPCA.VoltRon}
\title{getPCA}
\usage{
getPCA(object, ...)

\method{getPCA}{VoltRon}(object, assay = NULL, features = NULL, dims = 30, seed = 1, ...)
}
\arguments{
\item{object}{An object}

\item{...}{additional parameters passed to \code{vrEmbeddings}}

\item{assay}{assay}

\item{features}{the selected features for PCA reduction}

\item{dims}{the number of dimensions extracted from PCA}

\item{seed}{seed}
}
\description{
calculate getPCA of the VoltRon objects
}
