% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getPCA}
\alias{getPCA}
\title{getPCA}
\usage{
getPCA(
  object,
  assay = NULL,
  features = NULL,
  dims = 30,
  type = "pca",
  overwrite = FALSE,
  seed = 1,
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{features}{the selected features for PCA reduction}

\item{dims}{the number of dimensions extracted from PCA}

\item{type}{the key name for the embedding, default: pca}

\item{overwrite}{Whether the existing embedding with name 'type' should be overwritten in \link{vrEmbeddings}}

\item{seed}{seed}
}
\description{
calculate PCA of the VoltRon objects
}
