% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppAutomatedRegistration}
\alias{getRcppAutomatedRegistration}
\title{getRcppAutomatedRegistration}
\usage{
getRcppAutomatedRegistration(
  ref_image,
  query_image,
  invert_query = FALSE,
  invert_ref = FALSE,
  flipflop_query = "None",
  flipflop_ref = "None",
  rotate_query = "0",
  rotate_ref = "0"
)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{invert_query}{invert query image?}

\item{invert_ref}{invert reference image}

\item{flipflop_query}{flip or flop the query image}

\item{flipflop_ref}{flip or flop the reference image}

\item{rotate_query}{rotation of query image}

\item{rotate_ref}{rotation of reference image}
}
\description{
Automated registration with Rcpp
}
