% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppAutomatedRegistration}
\alias{getRcppAutomatedRegistration}
\title{getRcppAutomatedRegistration}
\usage{
getRcppAutomatedRegistration(
  ref_image,
  query_image,
  GOOD_MATCH_PERCENT = 0.15,
  MAX_FEATURES = 500,
  invert_query = FALSE,
  invert_ref = FALSE,
  flipflop_query = "None",
  flipflop_ref = "None",
  rotate_query = "0",
  rotate_ref = "0",
  matcher = "FLANN",
  method = "Homography",
  nonrigid = "TPS (OpenCV)"
)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{GOOD_MATCH_PERCENT}{the percentage of good matching keypoints, used by "Brute force" method}

\item{MAX_FEATURES}{maximum number of detected features, i.e. keypoints, used by "Brute force" method}

\item{invert_query}{invert query image?}

\item{invert_ref}{invert reference image}

\item{flipflop_query}{flip or flop the query image}

\item{flipflop_ref}{flip or flop the reference image}

\item{rotate_query}{rotation of query image}

\item{rotate_ref}{rotation of reference image}

\item{matcher}{the matching method for landmarks/keypoints FLANN or BRUTE-FORCE}

\item{method}{the automated registration method, Homography or Homography+TPS}
}
\description{
Automated registration workflos with Rcpp
}
