% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRcppManualRegistration}
\alias{getRcppManualRegistration}
\title{getRcppManualRegistration}
\usage{
getRcppManualRegistration(
  query_image,
  ref_image,
  query_landmark,
  reference_landmark
)
}
\arguments{
\item{query_image}{query image}

\item{ref_image}{reference image}

\item{query_landmark}{query landmark points}

\item{reference_landmark}{refernece landmark points}
}
\description{
Manual registration workflow with Rcpp
}
