% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredObject}
\alias{getRegisteredObject}
\title{getRegisteredObject}
\usage{
getRegisteredObject(obj_list, trans_matrix_list)
}
\arguments{
\item{obj_list}{a list of spatial data object}

\item{trans_matrix_list}{a list of transformation matrices}
}
\description{
get a registered Spatial data object
}
