% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredObject}
\alias{getRegisteredObject}
\title{getRegisteredObject}
\usage{
getRegisteredObject(obj)
}
\arguments{
\item{obj}{Object}
}
\description{
get a registered Spatial data object
}
