% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{getRegisteredObject}
\alias{getRegisteredObject}
\title{getRegisteredObject}
\usage{
getRegisteredObject(obj_list, mapping_list, register_ind, centre, input, ...)
}
\arguments{
\item{obj_list}{a list of spatial data object}

\item{mapping_list}{a list of transformation matrices}

\item{register_ind}{the indices of query images/spatialdatasets}

\item{centre}{the index of the central referance image/spatialdata}

\item{input}{input}

\item{...}{additional parameters passed to \code{getRegisteredObjectListVoltRon} or \code{getRegisteredObject.Seurat}}
}
\description{
Get a registered VoltRon or Seurat object
}
