% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredObject.Seurat}
\alias{getRegisteredObject.Seurat}
\title{getRegisteredObject.Seurat}
\usage{
getRegisteredObject.Seurat(seu, trans_matrix)
}
\arguments{
\item{seu}{Seurat object}

\item{trans_matrix}{transformation function for the registration}
}
\description{
get the cell data/locations from a Seurat Object
}
