% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{getRegisteredObject.SpaceRover}
\alias{getRegisteredObject.SpaceRover}
\title{getRegisteredObject.SpaceRover}
\usage{
getRegisteredObject.SpaceRover(sr, trans_matrix_list)
}
\arguments{
\item{trans_matrix_list}{transformation function for the registration}

\item{sr_list}{a list of SpaceRover objects}
}
\description{
get registered and merged SpaceRover object composed of several Samples
}
