% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getSpatialFeatures}
\alias{getSpatialFeatures}
\alias{getSpatialFeatures.SpaceRover}
\alias{getSpatialFeatures.srAssay}
\title{Get spatially variable feature}
\usage{
getSpatialFeatures(object, ...)

\method{getSpatialFeatures}{SpaceRover}(object, assay = NULL, ...)

\method{getSpatialFeatures}{srAssay}(object, max.count = 1, n = 3000)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
get spatially variable features in an assay
}
\concept{preprocessing}
