% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getSpatialNeighbors}
\alias{getSpatialNeighbors}
\title{Get spatial neighbors}
\usage{
getSpatialNeighbors(
  object,
  assay = NULL,
  method = "delaunay",
  radius = numeric(0),
  graph.key = method,
  ...
)
}
\arguments{
\item{object}{An object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{method}{the method spatial connectivity: "delaunay", "spatialkNN", "radius"}

\item{radius}{When \code{method = "radius"} selected, determines the radius of a neighborhood ball around each spatial point}

\item{graph.key}{the name of the graph}

\item{...}{additional parameters passed to \code{FNN:get.knn}}
}
\description{
get neighbors in an assay given spatial coordinates
}
