% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{getSpatialNeighbors}
\alias{getSpatialNeighbors}
\alias{getSpatialNeighbors.VoltRon}
\title{Get spatial neighbors}
\usage{
getSpatialNeighbors(object, ...)

\method{getSpatialNeighbors}{VoltRon}(
  object,
  assay = NULL,
  method = "delaunay",
  radius = numeric(0),
  graph.key = method,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{additional parameters passed to \code{FNN:get.knn}}

\item{assay}{assay}

\item{method}{the method spatial connectivity: "delaunay", "spatialkNN", "radius"}

\item{radius}{When \code{method = "radius"} selected, determines the radius of a neighborhood ball around each spatial point}

\item{graph.key}{the name of the graph}
}
\description{
get neighbors in an assay given spatial coordinates
}
