% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{getSpotsFromCells}
\alias{getSpotsFromCells}
\title{getSpotsFromCells}
\usage{
getSpotsFromCells(Visium, Xenium)
}
\arguments{
\item{Visium}{Visium Seurat Object}

\item{Xenium}{Xenium Seurat Object}
}
\description{
Generate Psuedo counts per spots and insert to as a separate image assay to the Visium object
}
