% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getUMAP}
\alias{getUMAP}
\alias{getUMAP.VoltRon}
\title{getUMAP}
\usage{
getUMAP(object, ...)

\method{getUMAP}{VoltRon}(object, assay = NULL, data.type = "pca", dims = 1:30, seed = 1)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{data.type}{the type of data used to calculate UMAP from: "pca" (default), "raw" or "norm"}

\item{dims}{the number of dimensions extracted from PCA}

\item{seed}{seed}
}
\description{
calculate getUMAP of the VoltRon objects
}
