% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{getUMAP}
\alias{getUMAP}
\alias{getUMAP.VoltRon}
\title{getUMAP}
\usage{
getUMAP(object, ...)

\method{getUMAP}{VoltRon}(object, assay = NULL, dims = 30, seed = 1)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{dims}{the number of dimensions extracted from PCA}

\item{seed}{seed}
}
\value{
\code{getUMAP}: The name of the default assay
}
\description{
calculate getUMAP of the VoltRon objects
}
