% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getVariableFeatures}
\alias{getVariableFeatures}
\title{getSharedFeatures}
\usage{
getVariableFeatures(object, assay = NULL, n = 3000, ...)
}
\arguments{
\item{object}{a Voltron Object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{n}{the number of features}

\item{...}{additional arguements passed to \code{vrFeatureData}}
}
\description{
get shared variable features across multiple assays
}
