% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getVariableFeatures}
\alias{getVariableFeatures}
\title{getVariableFeatures}
\usage{
getVariableFeatures(object, assay = NULL, n = 3000, ...)
}
\arguments{
\item{object}{a Voltron Object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{n}{the number of features}

\item{...}{additional arguements passed to \link{vrFeatureData}}
}
\description{
get shared variable features across multiple assays
}
