% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{Image}
\alias{Image}
\alias{Image.Seurat}
\alias{Image.VoltRon}
\alias{Image.vrSample}
\alias{Image.vrLayer}
\title{Image}
\usage{
Image(object, ...)

\method{Image}{Seurat}(seu, ...)

\method{Image}{VoltRon}(sr, ...)

\method{Image}{vrSample}(vr, ...)

\method{Image}{vrLayer}(vr, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Image}: The name of the default assay
}
\description{
Get images of a VoltRon assay
}
\concept{data-access}
