% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCosMx}
\alias{importCosMx}
\title{importCosMx}
\usage{
importCosMx(
  tiledbURI,
  assay_name = "CosMx",
  image = NULL,
  image_name = "DAPI",
  ome.tiff = NULL,
  ...
)
}
\arguments{
\item{tiledbURI}{the path to the tiledb folder}

\item{assay_name}{the assay name, default: CosMx}

\item{image}{the reference morphology image of the CosMx assay}

\item{image_name}{the image name of the CosMx assay, Default: DAPI}

\item{ome.tiff}{the OME.TIFF file of the CosMx experiment if exists}

\item{...}{additional parameters passed to \code{formVoltRon}}

\item{import_molecules}{if TRUE, molecule assay will be created along with cell assay.}
}
\description{
Import CosMx data
}
