% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importCosMx}
\alias{importCosMx}
\title{importCosMx}
\usage{
importCosMx(
  dcc.path,
  pkc.file,
  summarySegment,
  summarySegmentSheetName,
  assay_name = "CosMx",
  image = NULL,
  segment_polygons = FALSE,
  ome.tiff = NULL,
  ...
)
}
\arguments{
\item{dcc.path}{path to the folder where the dcc files are found}

\item{pkc.file}{path to the pkc file}

\item{summarySegment}{the metadata csv (sep = ";") or excel file, if the file is an excel file, \code{summarySegmentSheetName} should be provided as well.}

\item{summarySegmentSheetName}{the sheet name of the excel file, \code{summarySegment}}

\item{assay_name}{the assay name, default: CosMx}

\item{image}{the reference morphology image of the CosMx assay}

\item{segment_polygons}{if TRUE, the ROI polygons are parsed from the OME.TIFF file}

\item{ome.tiff}{the OME.TIFF file of the CosMx experiment if exists}

\item{...}{additional parameters passed to \code{formVoltRon}}
}
\description{
Import CosMx data
}
