% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importDBITSeq}
\alias{importDBITSeq}
\title{importDBITSeq}
\usage{
importDBITSeq(
  path.rna,
  path.prot = NULL,
  size = 10,
  assay_name = "DBIT-Seq",
  sample_name = NULL,
  image_name = "main",
  channel_name = "H&E",
  ...
)
}
\arguments{
\item{path.rna}{path to rna count matrix}

\item{path.prot}{path to protein count matrix}

\item{size}{the size of the in situ pixel (defualt is 10 (micron))}

\item{assay_name}{the assay name}

\item{sample_name}{the name of the sample}

\item{image_name}{the image name of the Visium assay, Default: main}

\item{channel_name}{the channel name of the image of the Visium assay, Default: H&E}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Importing DBIT-Seq data
}
