% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importGeoMx}
\alias{importGeoMx}
\title{importGeoMx}
\usage{
importGeoMx(
  dcc.path,
  pkc.file,
  summarySegment,
  summarySegmentSheetName,
  assay_name = "GeoMx",
  image = NULL,
  segment_polygons = FALSE,
  ome.tiff = NULL,
  resolution_level = 3,
  spatial = "main",
  image_name = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dcc.path}{path to the folder where the dcc files are found}

\item{pkc.file}{path to the pkc file}

\item{summarySegment}{the metadata csv (sep = ";") or excel file,
if the file is an excel file, \code{summarySegmentSheetName} should
be provided as well.}

\item{summarySegmentSheetName}{the sheet name of the excel file,
\code{summarySegment}}

\item{assay_name}{the assay name, default: GeoMx}

\item{image}{the reference morphology image of the GeoMx assay}

\item{segment_polygons}{if TRUE, the ROI polygons are parsed from the
OME.TIFF file}

\item{ome.tiff}{the OME.TIFF file of the GeoMx experiment if exists}

\item{resolution_level}{the level of resolution within GeoMx OME-TIFF
image, Default: 3}

\item{spatial}{the name of the coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the coordinate system, Default: main}

\item{verbose}{verbose}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Import GeoMx data
}
