% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportGeoMx}
\alias{ImportGeoMx}
\title{ImportGeoMx}
\usage{
ImportGeoMx(
  dir.path,
  pkc_file,
  summarySegment,
  summarySegmentSheetName,
  assay_name = "GeoMx",
  ...
)
}
\arguments{
\item{dir.path}{path to GeoMx run}

\item{pkc_file}{path to the pkc file}

\item{assay_name}{the assay name}

\item{...}{additional parameters passed to \code{CreateSpaceRover}}
}
\description{
ImportGeoMx
}
