% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportGeoMx}
\alias{ImportGeoMx}
\title{ImportGeoMx}
\usage{
ImportGeoMx(
  dir.path,
  pkc_file,
  summarySegment,
  summarySegmentSheetName,
  assay_name = "GeoMx",
  segment_polygons = FALSE,
  ome.tiff = NULL,
  ...
)
}
\arguments{
\item{dir.path}{path to GeoMx run}

\item{pkc_file}{path to the pkc file}

\item{summarySegment}{the annotation excel file}

\item{summarySegmentSheetName}{the sheet name of the excel file, \code{summarySegment}}

\item{assay_name}{the assay name, default: GeoMx}

\item{segment_polygons}{if TRUE, the ROI polygons are parsed from the OME.TIFF file}

\item{ome.tiff}{the OME.TIFF file of the GeoMx experiment if exists}

\item{...}{additional parameters passed to \code{CreateSpaceRover}}
}
\description{
ImportGeoMx
}
