% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importImageData}
\alias{importImageData}
\title{importImageData}
\usage{
importImageData(
  image,
  tile.size = 10,
  segments = NULL,
  spatial = "main",
  image_name = NULL,
  channels = NULL,
  series = 1,
  resolution = NULL,
  is.RGB = TRUE,
  ...
)
}
\arguments{
\item{image}{a single or a list of image paths or magick-image objects}

\item{tile.size}{the size of tiles}

\item{segments}{Either a list of segments or a GeoJSON file. This will
result in a second assay in the VoltRon object to be created}

\item{spatial}{the name of the coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the coordinate system, Default: main}

\item{channels}{the channel names of the images if multiple
images are provided}

\item{series}{the series IDs of the pyramidal image,
typical an integer starting from 1}

\item{resolution}{the resolution IDs of the
pyramidal image, typical an integer starting from 1}

\item{is.RGB}{If TRUE, all three channel images will be converted to RGB 
images.}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
import an image as VoltRon object
}
\examples{
# single image
imgfile <- system.file("extdata", "DAPI.tif", package = "VoltRon")
vrdata <- importImageData(imgfile, spatial = "main")

# multiple images
imgfile <- c(system.file("extdata", "DAPI.tif", package = "VoltRon"),
             system.file("extdata", "DAPI.tif", package = "VoltRon"))
vrdata <- importImageData(imgfile, spatial = "main",
                                   channels = c("DAPI", "DAPI2"))

}
