% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importImageData}
\alias{importImageData}
\title{importImageData}
\usage{
importImageData(image.path, tile.size = 10, ...)
}
\arguments{
\item{image.path}{the path to an image file}

\item{tile.size}{the size of tiles}
}
\description{
import an image as VoltRon object
}
