% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importImageData}
\alias{importImageData}
\title{importImageData}
\usage{
importImageData(image, tile.size = 10, stack.id = 1, segments = NULL, ...)
}
\arguments{
\item{image}{the path to an image file}

\item{tile.size}{the size of tiles}

\item{stack.id}{the id of the stack when the magick image composed of multiple layers}

\item{segments}{Either a list of segments or a GeoJSON file. This will result in a second assay in the VoltRon object to be created}

\item{...}{additional parameters passed to \code{formVoltRon}}
}
\description{
import an image as VoltRon object
}
