% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importOpenST}
\alias{importOpenST}
\title{importOpenST}
\usage{
importOpenST(
  h5ad.path,
  assay_name = "OpenST",
  sample_name = NULL,
  spatial = "main",
  image_name = NULL,
  channel_name = "H&E",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{h5ad.path}{path to h5ad file of STOmics output}

\item{assay_name}{the assay name}

\item{sample_name}{the name of the sample}

\item{spatial}{the name of the spatial coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the spatial coordinate system, Default: main}

\item{channel_name}{the channel name of the image of the Visium assay, Default: H&E}

\item{verbose}{verbose}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Importing OpenST data
}
