% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importQuPathIF}
\alias{importQuPathIF}
\title{importImageData}
\usage{
importQuPathIF(
  measurements,
  image,
  segments,
  image_name = "main",
  channels = NULL,
  series = 1,
  resolution = NULL,
  ...
)
}
\arguments{
\item{measurements}{measurements}

\item{image}{a single or a list of image paths or magick-image objects}

\item{segments}{Either a list of segments or a GeoJSON file. This will
result in a second assay in the VoltRon object to be created}

\item{image_name}{the image name of the Image assay, Default: main}

\item{channels}{the channel names of the images if multiple
images are provided}

\item{series}{the series IDs of the pyramidal image,
typically an integer starting from 1}

\item{resolution}{the resolution IDs of the
pyramidal image, typically an integer starting from 1}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
import an QuPath-quantified IF assay as VoltRon object
}
\examples{
# single image
imgfile <- system.file("extdata", "DAPI.tif", package = "VoltRon")
vrdata <- importImageData(imgfile, image_name = "main")

# multiple images
imgfile <- c(system.file("extdata", "DAPI.tif", package = "VoltRon"),
             system.file("extdata", "DAPI.tif", package = "VoltRon"))
vrdata <- importImageData(imgfile, image_name = "main",
                                   channel_name = c("DAPI", "DAPI2"))

}
