% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importVisium}
\alias{importVisium}
\title{importVisium}
\usage{
importVisium(
  dir.path,
  selected_assay = "Gene Expression",
  assay_name = "Visium",
  sample_name = NULL,
  spatial = "main",
  image_name = NULL,
  channel_name = "H&E",
  inTissue = TRUE,
  resolution_level = "lowres",
  ...
)
}
\arguments{
\item{dir.path}{path to Visium output folder}

\item{selected_assay}{selected assay from Visium}

\item{assay_name}{the assay name}

\item{sample_name}{the name of the sample}

\item{spatial}{the name of the coordinate system, Default: main}

\item{image_name}{(deprecated) the name of the coordinate system, Default: main}

\item{channel_name}{the channel name of the image of the Visium assay,
Default: H&E}

\item{inTissue}{if TRUE, only barcodes that are in the tissue will
be kept (default: TRUE)}

\item{resolution_level}{the level of resolution of Visium image:
"lowres" (default) or "hires"}

\item{...}{additional parameters passed to \link{formVoltRon}}
}
\description{
Importing Visium data
}
