% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ImportVisium}
\alias{ImportVisium}
\title{ImportVisium}
\usage{
ImportVisium(
  dir.path,
  layer_name = NULL,
  sample_name = NULL,
  assay_name = "Visium",
  assay_type = "Spot",
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium}
}
\description{
ImportVisium
}
