% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportVisium}
\alias{ImportVisium}
\title{ImportVisium}
\usage{
ImportVisium(dir.path, assay_name = "Visium", InTissue = TRUE, ...)
}
\arguments{
\item{dir.path}{path to Xenium}

\item{assay_name}{the assay name}

\item{...}{additional parameters passed to \code{CreateVoltRon}}

\item{inTissue}{if TRUE, only barcodes that are in the tissue will be kept (default: TRUE)}
}
\description{
ImportVisium
}
