% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importVisium}
\alias{importVisium}
\title{importVisium}
\usage{
importVisium(
  dir.path,
  selected_assay = "Gene Expression",
  assay_name = "Visium",
  inTissue = TRUE,
  ...
)
}
\arguments{
\item{dir.path}{path to Visium output folder}

\item{selected_assay}{selected assay from Visium}

\item{assay_name}{the assay name}

\item{inTissue}{if TRUE, only barcodes that are in the tissue will be kept (default: TRUE)}

\item{...}{additional parameters passed to \code{formVoltRon}}
}
\description{
Importing Visium data
}
