% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportXenium}
\alias{ImportXenium}
\title{ImportXenium}
\usage{
ImportXenium(
  dir.path,
  selected_assay = "Gene Expression",
  assay_name = "Xenium",
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium}

\item{selected_assay}{selected assay}

\item{assay_name}{the assay name of the SR object}

\item{...}{additional parameters passed to \code{formVoltRon}}
}
\description{
Importing a Xenium data
}
