% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ImportXenium}
\alias{ImportXenium}
\title{ImportXenium}
\usage{
ImportXenium(
  dir.path,
  selected_assay = "Gene Expression",
  layer_name = NULL,
  sample_name = NULL,
  assay_name = "Xenium",
  assay_type = "cell",
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium}

\item{selected_assay}{selected assay}
}
\description{
ImportXenium
}
