% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importXenium}
\alias{importXenium}
\title{importXenium}
\usage{
importXenium(
  dir.path,
  selected_assay = "Gene Expression",
  assay_name = "Xenium",
  sample_name = NULL,
  use_image = TRUE,
  morphology_image = "morphology_lowres.tif",
  resolution_level = 7,
  overwrite_resolution = FALSE,
  image_name = "main",
  channel_name = "DAPI",
  import_molecules = FALSE,
  ...
)
}
\arguments{
\item{dir.path}{path to Xenium output folder}

\item{selected_assay}{selected assay from Xenium}

\item{assay_name}{the assay name of the SR object}

\item{sample_name}{the name of the sample}

\item{use_image}{if TRUE, the DAPI image will be used.}

\item{morphology_image}{the name of the lowred morphology image. Default: morphology_lowres.tif}

\item{resolution_level}{the level of resolution within Xenium OME-TIFF image, see \code{generateXeniumImage}. Default: 7 (553x402)}

\item{overwrite_resolution}{if TRUE, the image "file.name" will be generated again although it exists at "dir.path"}

\item{image_name}{the image name of the Xenium assay, Default: main}

\item{channel_name}{the channel name of the image of the Xenium assay, Default: DAPI}

\item{import_molecules}{if TRUE, molecule assay will be created along with cell assay.}

\item{...}{additional parameters passed to \code{formVoltRon}}
}
\description{
Importing Xenium data
}
