% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{initateKeypoints}
\alias{initateKeypoints}
\title{initateKeypoints}
\usage{
initateKeypoints(len_images, keypoints_list, input, output, session)
}
\arguments{
\item{len_images}{the length of images}

\item{keypoints_list}{the list of keypoint pairs}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Initiate shiny reactive values for keypoint dataframes for pairwise reference and query images
}
