% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{initateKeypoints}
\alias{initateKeypoints}
\title{}
\usage{
initateKeypoints(len_images, keypoints_list, input, output, session)
}
\arguments{
\item{len_images}{the length of images}

\item{keypoints_list}{the list of keypoint pairs}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
# plot with raster
  ref_image_raster <- as.raster(images[[ref_ind]]) |> as.matrix() |> terra::rast()
  query_image_raster <- as.raster(images[[query_ind]]) |> as.matrix() |> terra::rast()
  if(terra::nlyr(query_image_raster) == 1){
    query_image_raster <- raster::stack(query_image_raster)
    query_image_raster <- raster::stack(query_image_raster, query_image_raster, query_image_raster)
  } else {
    query_image_raster <- raster::stack(query_image_raster)
  }
}
\details{

}
