% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{LoadNanostring}
\alias{LoadNanostring}
\title{LoadNanostring}
\usage{
LoadNanostring(
  data.dir,
  fov,
  assay = "Nanostring",
  type = c("centroids", "segmentations")
)
}
\arguments{
\item{assay}{}
}
\description{
Load Nanostring CosMx data
}
