% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{MainAssay}
\alias{MainAssay}
\alias{MainAssay<-}
\alias{MainAssay.SpaceRover}
\alias{MainAssay<-.SpaceRover}
\title{#' merge
#'
#' Merge multiple spaceRover objects
#'
#' @param object An object
#' @param ... Arguments passed to other methods
#'
#' @return \code{merge}: The name of the default assay
#'
#' @rdname merge
#' @export merge
#'
merge <- function(object, ...) {
  UseMethod(generic = 'merge', object = object)
}
Main Assay}
\usage{
MainAssay(object, ...)

MainAssay(object, ...) <- value

\method{MainAssay}{SpaceRover}(object, ...)

\method{MainAssay}{SpaceRover}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{MainAssay}: The name of the default assay

\code{MainAssay<-}: The name of the default assay updated
}
\description{
Get and set the main assay
}
\concept{data-access}
