% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.VoltRon}
\alias{merge.VoltRon}
\alias{merge}
\title{Merging VoltRon objects}
\usage{
\method{merge}{VoltRon}(object, object_list, samples = NULL, main.assay = NULL, verbose = TRUE)
}
\arguments{
\item{object}{a VoltRon Object}

\item{object_list}{a list of VoltRon objects}

\item{samples}{a single sample name or multiple sample names of the same size as the given VoltRon objects}

\item{main.assay}{the name of the main assay}

\item{verbose}{verbose}
}
\description{
Given a VoltRon object, and a list of VoltRon objects, merge all.
}
