% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.SpaceRover}
\alias{merge.SpaceRover}
\title{}
\usage{
\method{merge}{SpaceRover}(object, object_list, sample_name = NULL, main.assay = NULL)
}
\arguments{
\item{object}{a SpaceRover Object}

\item{object_list}{a list of SpaceRover objects}

\item{sample_name}{a single sample name if objects are of the same sample}

\item{main.assay}{name of the assay}
}
\description{
# combine all elements
  if(!is.list(object_list))
    object_list <- list(object_list)
  object_list <- c(object, object_list)
}
\details{

}
