% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Metadata}
\alias{Metadata}
\alias{Metadata<-}
\alias{Metadata.VoltRon}
\alias{Metadata<-.VoltRon}
\title{Metadata}
\usage{
Metadata(object, ...)

Metadata(object, ...) <- value

\method{Metadata}{VoltRon}(object, type = "cell")

\method{Metadata}{VoltRon}(object, type = "cell", ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{type}{the assay type: ROI, spot or cell}

\item{value}{new metadata}
}
\value{
\code{Metadata}: The name of the default assay

\code{Metadata<-}: The coordinates updated
}
\description{
Get the metadata
}
\concept{data-access}
