% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Metadata}
\alias{Metadata}
\alias{Metadata.SpaceRover}
\title{Metadata}
\usage{
Metadata(object, ...)

\method{Metadata}{SpaceRover}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Metadata}: The name of the default assay
}
\description{
Get the metadata
}
\concept{data-access}
