% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Metadata}
\alias{Metadata}
\alias{Metadata<-}
\alias{Metadata.VoltRon}
\alias{Metadata<-.VoltRon}
\title{Metadata}
\usage{
Metadata(object, assay = NULL, type = NULL)

Metadata(object, assay = NULL, type = NULL) <- value

\method{Metadata}{VoltRon}(object, assay = NULL, type = NULL)

\method{Metadata}{VoltRon}(object, type = NULL, assay = NULL) <- value
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{type}{the assay type: ROI, spot or cell, or all for the entire metadata object}

\item{value}{new metadata}
}
\description{
Get the metadata of a VoltRon object.
}
