% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{modulateImage}
\alias{modulateImage}
\alias{modulateImage,VoltRon-method}
\alias{modulateImage,vrAssay-method}
\alias{modulateImage,vrAssayV2-method}
\alias{modulateImage,vrImage-method}
\alias{modulateImage,vrSpatial-method}
\title{modulateImage}
\usage{
modulateImage(object, ...)

\S4method{modulateImage}{VoltRon}(
  object,
  assay = NULL,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\S4method{modulateImage}{vrAssay}(
  object,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\S4method{modulateImage}{vrAssayV2}(
  object,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\S4method{modulateImage}{vrImage}(
  object,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\S4method{modulateImage}{vrSpatial}(
  object,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{spatial}{the name of the spatial coord. system}

\item{reg}{TRUE if registered coordinates of the main image 
(\link{vrMainSpatial}) is requested}

\item{channel}{the name of the channel associated with the image}

\item{brightness}{modulation of brightness as percentage of the 
current value (100 for no change)}

\item{saturation}{modulation of saturation as percentage of the 
current value (100 for no change)}

\item{hue}{modulation of hue is an absolute rotation of -180 degrees 
to +180 degrees from the current position corresponding to an argument 
range of 0 to 200 (100 for no change)}

\item{force}{if TRUE, all channels will be modulated given no specific 
channel name}
}
\description{
Modulating Magick images
}
