% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{modulateImage}
\alias{modulateImage}
\alias{modulateImage.VoltRon}
\alias{modulateImage.vrAssay}
\alias{modulateImage.vrImage}
\alias{modulateImage.vrSpatial}
\title{modulateImage}
\usage{
modulateImage(
  object,
  assay = NULL,
  name = NULL,
  reg = FALSE,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\method{modulateImage}{VoltRon}(object, assay = NULL, ...)

\method{modulateImage}{vrAssay}(object, name = NULL, reg = FALSE, ...)

\method{modulateImage}{vrImage}(
  object,
  channel = NULL,
  brightness = 100,
  saturation = 100,
  hue = 100,
  force = FALSE
)

\method{modulateImage}{vrSpatial}(object, ...)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the name of the image}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainSpatial}) is requested}

\item{channel}{the name of the channel associated with the image}

\item{brightness}{modulation of brightness as percentage of the current value (100 for no change)}

\item{saturation}{modulation of saturation as percentage of the current value (100 for no change)}

\item{hue}{modulation of hue is an absolute rotation of -180 degrees to +180 degrees from the current position corresponding to an argument range of 0 to 200 (100 for no change)}

\item{force}{if TRUE, all channels will be modulated given no specific channel name}

\item{...}{arguements passed to other methods}
}
\description{
Modulating Magick images
}
