% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{normalizeData}
\alias{normalizeData}
\alias{normalizeData.SpaceRover}
\alias{normalizeData.srAssay}
\title{Normalize Data}
\usage{
normalizeData(object, ...)

\method{normalizeData}{SpaceRover}(object, assay = NULL, ...)

\method{normalizeData}{srAssay}(object, method = "LogNorm", desiredQuantile = 0.9)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
Normalize the count data present in a given assay.
}
\concept{preprocessing}
