% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{normalizeData}
\alias{normalizeData}
\alias{normalizeData.VoltRon}
\alias{normalizeData.vrAssay}
\title{Normalize Data}
\usage{
normalizeData(object, ...)

\method{normalizeData}{VoltRon}(object, assay = NULL, ...)

\method{normalizeData}{vrAssay}(
  object,
  method = "LogNorm",
  desiredQuantile = 0.9,
  scale = 0.2,
  sizefactor = 10000
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{method}{the normalization method: "LogNorm", "Q3Norm", "LogQ3Norm" or "CLR"}

\item{desiredQuantile}{the quantile of the data if "QuanNorm" or "LogQuanNorm" is selected as \code{method}}

\item{scale}{the scale parameter for the hyperbolic arcsine transformation}

\item{sizefactor}{size factor if \code{method} is selected as \code{LogNorm}}
}
\description{
Normalize the count data present in a given assay.
}
