% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{NormalizeData}
\alias{NormalizeData}
\alias{NormalizeData.SpaceRover}
\alias{NormalizeData.srAssay}
\alias{NormalizeData.default}
\title{Normalize Data}
\usage{
NormalizeData(object, ...)

\method{NormalizeData}{SpaceRover}(object, ...)

\method{NormalizeData}{srAssay}(object, method = "LogNormalize")

\method{NormalizeData}{default}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
Normalize the count data present in a given assay.
}
\concept{preprocessing}
