% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{register_slides}
\alias{register_slides}
\title{register_slides}
\usage{
register_slides(reference_seu, query_list_seu, keypoints = NULL)
}
\arguments{
\item{reference_seu}{Seurat object with reference spatial assay}

\item{query_list_seu}{a list of Seurat Object with query images}
}
\value{
keypoints
}
\description{
A mini shiny app to calculate spatial cell(barcode) projection matrix between a visium slide and a xenium slide
}
