% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{register_slides}
\alias{register_slides}
\title{Image registration between slides}
\usage{
register_slides(Visium_data, Xenium_data_image_path)
}
\arguments{
\item{Visium_data}{visium data}

\item{Xenium_data_image_path}{xenium image}
}
\value{
projection matrix
}
\description{
A mini shiny app to calculate spatial cell(barcode) projection matrix between a visium slide and a xenium slide
}
