% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registerSpatialData}
\alias{registerSpatialData}
\title{registerSpatialData}
\usage{
registerSpatialData(
  object_list = NULL,
  reference_spatdata = NULL,
  query_spatdata = NULL,
  keypoints = NULL,
  mapping_parameters = list(),
  interactive = TRUE,
  shiny.options = list(launch.browser = getOption("shiny.launch.browser", interactive()))
)
}
\arguments{
\item{object_list}{a list of VoltRon (or Seurat) objects}

\item{reference_spatdata}{a reference spatial data set, used only if \code{object_list} is \code{NULL}}

\item{query_spatdata}{a query spatial data set, used only if \code{object_list} is \code{NULL}}

\item{keypoints}{(DEPRECATED) a list of tables, each points to matching keypoints from registered images.}

\item{mapping_parameters}{for manual image registration, a list of tables, each points to matching keypoints from registered images, and for automated image registration, a set of mapping parameters}

\item{interactive}{if TRUE, the shiny application for image registration will be triggered, otherwise 'mapping_parameters' or 'keypoints' should be defined.}

\item{shiny.options}{a list of shiny options (launch.browser, host, port etc.) passed \code{options} arguement of \link{shinyApp}. For more information, see \link{runApp}}
}
\description{
A mini shiny app for registering images and spatial coordinates of multiple consecutive VoltRon objects.
}
