% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registerSpatialData}
\alias{registerSpatialData}
\title{register_slides}
\usage{
registerSpatialData(
  data_list = NULL,
  reference_spatdata = NULL,
  query_spatdata = NULL,
  keypoints = NULL
)
}
\arguments{
\item{data_list}{a list of VoltRon objects}

\item{reference_spatdata}{a reference spatial data set, used only if \code{spatial_data_list} is \code{NULL}}

\item{query_spatdata}{a query spatial data set, used only if \code{spatial_data_list} is \code{NULL}}

\item{keypoints}{keypoints tables for each registration}
}
\description{
A mini shiny app to for registering images and spatial coordinates of multiple consequtive spatial datasets
}
