% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registerSpatialData}
\alias{registerSpatialData}
\title{registerSpatialData}
\usage{
registerSpatialData(
  object_list = NULL,
  reference_spatdata = NULL,
  query_spatdata = NULL,
  keypoints = NULL,
  shiny.options = list(launch.browser = getOption("shiny.launch.browser", interactive()))
)
}
\arguments{
\item{object_list}{a list of VoltRon (or Seurat) objects}

\item{reference_spatdata}{a reference spatial data set, used only if \code{object_list} is \code{NULL}}

\item{query_spatdata}{a query spatial data set, used only if \code{object_list} is \code{NULL}}

\item{keypoints}{a list of tables, each points to matching keypoints from registered images.}

\item{shiny.options}{a list of shiny options (launch.browser, host, port etc.) passed \code{options} arguement of \link{shinyApp}. For more information, see \link{runApp}}
}
\description{
A mini shiny app to for registering images and spatial coordinates of multiple consequtive spatial datasets
}
