% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{resizeImage}
\alias{resizeImage}
\alias{resizeImage.VoltRon}
\alias{resizeImage.vrAssay}
\title{resizeImage}
\usage{
resizeImage(object, ...)

\method{resizeImage}{VoltRon}(object, ...)

\method{resizeImage}{vrAssay}(object, size)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{resizeImage}: The name of the default assay
}
\description{
Resizing Magick images
}
\concept{data-access}
