% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{resizeImage}
\alias{resizeImage}
\alias{resizeImage.VoltRon}
\alias{resizeImage.vrAssay}
\alias{resizeImage.vrImage}
\alias{resizeImage.vrSpatial}
\title{resizeImage}
\usage{
resizeImage(object, assay = NULL, name = NULL, reg = FALSE, size)

\method{resizeImage}{VoltRon}(object, assay = NULL, ...)

\method{resizeImage}{vrAssay}(object, name = NULL, reg = FALSE, ...)

\method{resizeImage}{vrImage}(object, size)

\method{resizeImage}{vrSpatial}(object, ...)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the name of the image}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainImage}) is requested}

\item{size}{the width of the resized image}

\item{...}{arguements passed to other methods}
}
\description{
Resizing Magick images
}
