% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{resizeImage}
\alias{resizeImage}
\alias{resizeImage.VoltRon}
\alias{resizeImage.vrAssay}
\alias{resizeImage.vrImage}
\title{resizeImage}
\usage{
resizeImage(object, ...)

\method{resizeImage}{VoltRon}(object, ...)

\method{resizeImage}{vrAssay}(object, name = NULL, reg = FALSE, ...)

\method{resizeImage}{vrImage}(object, size)
}
\arguments{
\item{object}{An object}

\item{...}{arguements passed to \code{vrImages.vrImage}}

\item{name}{the name of the main image}

\item{reg}{TRUE if registered images are assigned}

\item{size}{the width of the resized image}
}
\description{
Resizing Magick images
}
