% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{ResizeImage}
\alias{ResizeImage}
\alias{ResizeImage.VoltRon}
\alias{ResizeImage.vrAssay}
\title{ResizeImage}
\usage{
ResizeImage(object, ...)

\method{ResizeImage}{VoltRon}(object, ...)

\method{ResizeImage}{vrAssay}(object, size)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{ResizeImage}: The name of the default assay
}
\description{
Resizing Magick images
}
\concept{data-access}
