% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{SampleMetadata}
\alias{SampleMetadata}
\alias{SampleMetadata.VoltRon}
\title{SampleMetadata}
\usage{
SampleMetadata(object, ...)

\method{SampleMetadata}{VoltRon}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get the sample metadata
}
