% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{setVRMetadata}
\alias{setVRMetadata}
\title{setVRMetadata}
\usage{
setVRMetadata(molecule, cell, spot, ROI)
}
\arguments{
\item{molecule}{molecule data frame}

\item{cell}{cell data frame}

\item{spot}{spot data frame}

\item{ROI}{ROI data frame}
}
\description{
setVRMetadata
}
