% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SFPlot}
\alias{SFPlot}
\title{SFPlot}
\usage{
SFPlot(
  object,
  features,
  ncol = 3,
  nrow = NULL,
  sample = NULL,
  font.size = 2,
  keep.scale = "feature",
  mode = "ROI"
)
}
\arguments{
\item{object}{SpaceRover object}

\item{features}{a vector of features to visualize}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{sample}{sample names matching spatial points (i.e. entities)}

\item{font.size}{font sizes}

\item{keep.scale}{whether unify all scales for all features or not}

\item{mode}{plotting mode for different spatial entities, ROI, Spot or cell}

\item{split.by}{split by a metadata element}
}
\description{
Functions for plotting spatial objects
}
