% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatFeatPlot}
\alias{SpatFeatPlot}
\title{SpatFeatPlot}
\usage{
SpatFeatPlot(
  object,
  features,
  group.by = "label",
  assay = NULL,
  assay.type = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  alpha = 0.6,
  keep.scale = "feature",
  label = FALSE,
  background = "image",
  crop = FALSE,
  common.legend = TRUE
)
}
\arguments{
\item{object}{SpaceRover object}

\item{features}{a set of features, either from the rows of rawdata, normdata or columns of the metadata}

\item{group.by}{a grouping label for the spatial entities}

\item{assay}{the assay name}

\item{assay.type}{the assay type name: 'cell', 'spot' or 'ROI'}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{keep.scale}{whether unify all scales for all features or not}

\item{label}{if TRUE, labels of ROIs will be visualized too}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{crop}{whether to crop an image of a spot assay}

\item{common.legend}{whether to use a common legend for all plots}
}
\description{
Plotting single/multiple features of spatially resolved cells, spots, and ROI on associated images from multiple assays in a SpaceRover object.
}
