% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatFeatPlotSingle}
\alias{SpatFeatPlotSingle}
\title{SpatFeatPlotSingle}
\usage{
SpatFeatPlotSingle(
  coords,
  image,
  feature,
  limits,
  group.by = "label",
  font.size = 2,
  keep.scale = "feature",
  label = FALSE,
  plot_title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{legend_title}{}
}
\description{
SpatFeatPlotSingle
}
