% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{srAssay-class}
\alias{srAssay-class}
\alias{srAssay}
\title{The srAssay (SpaceRover Assay) Class}
\description{
The srAssay (SpaceRover Assay) Class
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{raw count table}

\item{\code{normdata}}{normalized count table}

\item{\code{coord}}{spatial coordinates of the assay}

\item{\code{image}}{image of the spatial assay}
}}

