% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{subset,vrImage-method}
\alias{subset,vrImage-method}
\title{Subsetting vrImage objects}
\usage{
\S4method{subset}{vrImage}(x, subset, spatialpoints = NULL, image = NULL)
}
\arguments{
\item{x}{A vrImage object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \link{image_crop}}
}
\description{
Given a vrImage object, subset the object given one of the attributes.
}
