% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/objects.R
\name{subset.vrAssay}
\alias{subset.vrAssay}
\alias{subset.VoltRon}
\title{Subsetting VoltRon objects}
\usage{
\method{subset}{vrAssay}(object, subset, spatialpoints = NULL, features = NULL, image = NULL)

\method{subset}{VoltRon}(
  object,
  subset,
  samples = NULL,
  assays = NULL,
  spatialpoints = NULL,
  features = NULL,
  image = NULL,
  interactive = FALSE
)
}
\arguments{
\item{object}{A vrAssay object}

\item{subset}{Logical statement for subsetting}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{features}{the set of features to subset the object}

\item{image}{the subseting string passed to \code{magick::image_crop}}

\item{samples}{the set of samples to subset the object}

\item{interactive}{TRUE if interactive subsetting on the image is demanded}
}
\description{
Given a VoltRon object, subset the object given one of the attributes
}
