% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{subset.sampleMetadata}
\alias{subset.sampleMetadata}
\title{subset.sampleMetadata}
\usage{
\method{subset}{sampleMetadata}(metadata, samples = NULL, assays = NULL)
}
\arguments{
\item{metadata}{sample metadata of a spaceRover object}

\item{samples}{samples}

\item{assays}{assays}
}
\description{
subset.sampleMetadata
}
