% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{subset.vrBlock}
\alias{subset.vrBlock}
\title{Subsetting vrBlock objects}
\usage{
\method{subset}{vrBlock}(object, ...)
}
\description{
Given a vrBlock object, subset the object given one of the attributes
}
