% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{subset.vrImage}
\alias{subset.vrImage}
\title{Subsetting vrImage objects}
\usage{
\method{subset}{vrImage}(object, subset, spatialpoints = NULL, image = NULL)
}
\arguments{
\item{object}{A vrImage object}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \code{magick::image_crop}}
}
\description{
Given a vrImage object, subset the object given one of the attributes.
}
