% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{subset.vrLayer}
\alias{subset.vrLayer}
\title{Subsetting vrLayer objects}
\usage{
\method{subset}{vrLayer}(object, subset, assays = NULL, spatialpoints = NULL, image = NULL)
}
\arguments{
\item{assays}{the set of assays to subset the object}

\item{spatialpoints}{the set of spatial points to subset the object}

\item{image}{the subseting string passed to \code{magick::image_crop}}
}
\description{
Given a vrLayer object, subset the object given one of the attributes
}
