% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subsetCoordinates}
\alias{subsetCoordinates}
\title{subsetCoordinates}
\usage{
subsetCoordinates(coords, image, crop_info)
}
\arguments{
\item{coords}{coordinates}

\item{image}{the magick image associated with the coordinates}

\item{crop_info}{the subseting string passed to \code{magick::image_crop}}
}
\description{
subsetting coordinates given cropping parameters of a magick image objects
}
