% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subsetSegments}
\alias{subsetSegments}
\title{subsetSegments}
\usage{
subsetSegments(segments, image, crop_info)
}
\arguments{
\item{image}{the magick image associated with the coordinates}

\item{crop_info}{image crop parameters in character, see \code{image_crop}}

\item{coords}{coordinates}
}
\description{
subsetting segments given cropping parameters of a magick image objects
}
