% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subsetSegments}
\alias{subsetSegments}
\title{subsetSegments}
\usage{
subsetSegments(segments, imageinfo, crop_info)
}
\arguments{
\item{segments}{the list of segments each associated with a spatial point}

\item{imageinfo}{the magick image info associated with the image}

\item{crop_info}{the subseting string passed to \link{image_crop}}
}
\description{
subsetting segments given cropping parameters of a magick image objects
}
