% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{subsetSegments}
\alias{subsetSegments}
\title{subsetSegments}
\usage{
subsetSegments(segments, image, crop_info)
}
\arguments{
\item{segments}{a list of coordinates for each segment}

\item{image}{the magick image associated with the coordinates}

\item{crop_info}{the subseting string passed to \code{magick::image_crop}}
}
\description{
subsetting segments given cropping parameters of a magick image objects
}
