% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{transferData}
\alias{transferData}
\title{transferData}
\usage{
transferData(object, from = NULL, to = NULL, features = NULL)
}
\arguments{
\item{object}{a VoltRon object}

\item{from}{The ID of assay whose data transfer to the second assay}

\item{to}{The ID of the target assay where data is transfered to}

\item{features}{The set of data or metadata features that are transfered. Only one metadata feature can be transfered at a time.}
}
\description{
transfer data across assays
}
