% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{transferData}
\alias{transferData}
\title{transferData}
\usage{
transferData(
  object,
  from = NULL,
  to = NULL,
  features = NULL,
  new_assay_name = NULL
)
}
\arguments{
\item{object}{a VoltRon object}

\item{from}{The ID of assay whose data transfered to the second assay}

\item{to}{The ID of the target assay where data is transfered to}

\item{features}{The set of data or metadata features that are transferred. Only one metadata feature can be transfered at a time.}

\item{new_assay_name}{the name of the new assay created from the source assay defined in \code{from} argument.}
}
\description{
transfer data across assays
}
