% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{transferData}
\alias{transferData}
\title{transferData}
\usage{
transferData(
  object,
  from = NULL,
  to = NULL,
  features = NULL,
  new_feature_name = NULL
)
}
\arguments{
\item{object}{a VoltRon object}

\item{from}{the name or class of assay whose data transfered to the second assay}

\item{to}{the name or class of target assay where data is transfered to}

\item{features}{the set of features from \link{vrFeatures} or metadata columns from \link{Metadata} that are transferred. 
Only one metadata feature can be transfered at a time.}

\item{new_feature_name}{the name of the new feature set created from the source assay defined in \code{from} argument.}
}
\description{
transfer data across assays
}
