% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registering.R
\name{transform_keypoints}
\alias{transform_keypoints}
\title{Rotate and Transform Keypoints}
\usage{
transform_keypoints(
  keypoints,
  orig_image,
  trans_image,
  angle,
  extension,
  input,
  session
)
}
\arguments{
\item{keypoints}{keypoints visualized on image}

\item{orig_image}{original magick image}

\item{trans_image}{transformed magick image}

\item{angle}{rotation angle}

\item{extension}{name extension for the shiny input parameter}

\item{input}{input}

\item{session}{session}
}
\value{
magick image
}
\description{
apply transformations to keypoints given images
}
