% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{transform_magick_image_query_list}
\alias{transform_magick_image_query_list}
\title{transform_magick_image_query_list}
\usage{
transform_magick_image_query_list(image_list, input, session)
}
\arguments{
\item{image_list}{magick image}

\item{input}{shiny input}

\item{session}{shiny session}
}
\value{
magick image
}
\description{
Apply given transformations to a list of magick image and return shiny reactive
}
