% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{transformImage}
\alias{transformImage}
\title{transformImage}
\usage{
transformImage(image, extension, input)
}
\arguments{
\item{image}{magick image}

\item{extension}{name extension for the shiny input parameter}

\item{input}{shiny input}
}
\description{
Apply given transformations to a magick image
}
