% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/processing.R
\name{UMAP}
\alias{UMAP}
\alias{UMAP.VoltRon}
\title{UMAP}
\usage{
UMAP(object, ...)

\method{UMAP}{VoltRon}(object, assay = NULL, dims = 30, seed = 1)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{UMAP}: The name of the default assay
}
\description{
calculate UMAP of the VoltRon objects
}
\concept{data-access}
\concept{embedding}
