% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{updateAssay}
\alias{updateAssay}
\title{Update Assay}
\usage{
updateAssay(object, assay)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
update assays from vrAssay to vrAssayV2
}
