% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateMetadataAssay}
\alias{updateMetadataAssay}
\title{updateMetadataAssay}
\usage{
updateMetadataAssay(object1, object2)
}
\arguments{
\item{object1}{vrMetadata object}

\item{object2}{vrMetadata object}
}
\description{
updating assay names for merge
}
