% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{UpdateSequentialTabPanels}
\alias{UpdateSequentialTabPanels}
\title{UpdateSequentialTabPanels}
\usage{
UpdateSequentialTabPanels(input, output, session, npanels)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{npanels}{the number of panels in both reference and query image sections}
}
\description{
A function for automatized selection of reference/query images
}
