% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{VoltRon-class}
\alias{VoltRon-class}
\alias{VoltRon}
\title{The VoltRon Class}
\description{
The VoltRon Class
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of samples (vrSample)}

\item{\code{metadata}}{A vrMetadata object that includes metadata of ROIs, spots, and cells}

\item{\code{sample.metadata}}{Contains meta-information about each sample, layer and assay}

\item{\code{graph}}{A igraph object}

\item{\code{main.assay}}{The type of the main assay (i.e. Visium, Xenium, GeoMx etc.)}

\item{\code{project}}{Name of the project}
}}

