% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{VoltRon-methods}
\alias{VoltRon-methods}
\alias{$.VoltRon}
\alias{$<-.VoltRon}
\alias{.DollarNames.VoltRon}
\alias{[[,VoltRon,character,missing-method}
\alias{[[<-,VoltRon,character,missing,ANY-method}
\alias{[[,VoltRon,character,character-method}
\alias{[[<-,VoltRon,character,character,ANY-method}
\title{Methods for VoltRon}
\usage{
\method{$}{VoltRon}(x, i, ...)

\method{$}{VoltRon}(x, i, ...) <- value

\method{.DollarNames}{VoltRon}(object, pattern = "")

\S4method{[[}{VoltRon,character,missing}(x, i, j, ...)

\S4method{[[}{VoltRon,character,missing,ANY}(x, i, j, ...) <- value

\S4method{[[}{VoltRon,character,character}(x, i, j, ...)

\S4method{[[}{VoltRon,character,character,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{ An R object for which valid names after \code{"$"}
    are computed and returned.
  }

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\description{
Methods for \code{\link{VoltRon}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{$}: Metadata access for \code{VoltRon} objects

\item \code{`$`(VoltRon) <- value}: Metadata overwrite for \code{VoltRon} objects

\item \code{.DollarNames(VoltRon)}: Autocompletion for \code{$} access for \code{VoltRon} objects

\item \code{x[[i}: Accessing vrAssay or vrSample objects from \code{VoltRon} objects

\item \code{`[[`(x = VoltRon, i = character, j = missing) <- value}: Overwriting vrAssay or vrSample objects from \code{VoltRon} objects

\item \code{x[[i}: Accessing vrLayer objects from \code{VoltRon} objects

\item \code{`[[`(x = VoltRon, i = character, j = character) <- value}: Overwriting vrLayer objects from \code{VoltRon} objects

}}
\concept{voltron}
