% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{vrAssay-class}
\alias{vrAssay-class}
\alias{vrAssay}
\title{The vrAssay (VoltRon Assay) Class}
\description{
The vrAssay (VoltRon Assay) Class
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{raw count table}

\item{\code{normdata}}{normalized count table}

\item{\code{coord}}{spatial coordinates of the assay}

\item{\code{coord_reg}}{spatial coordinates of the registered assay}

\item{\code{segments}}{spatial coordinates of the segments}

\item{\code{segments_reg}}{spatial coordinates of the registered segments}

\item{\code{image}}{image of the spatial assay}

\item{\code{params}}{additional parameters used by different assay types}

\item{\code{type}}{the type of the assay (cell, spot, ROI)}
}}

