% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{vrAssay-class}
\alias{vrAssay-class}
\alias{vrAssay}
\title{The vrAssay (VoltRon Assay) Class}
\description{
The vrAssay (VoltRon Assay) Class
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{raw count table}

\item{\code{normdata}}{normalized count table}

\item{\code{featuredata}}{feature metadata}

\item{\code{embeddings}}{list of embeddings}

\item{\code{coords}}{spatial coordinates of the assay}

\item{\code{coords_reg}}{spatial coordinates of the registered assay}

\item{\code{segments}}{spatial coordinates of the segments, if available}

\item{\code{segments_reg}}{spatial coordinates of the registered segments, if available}

\item{\code{subcellular}}{subcellular data of the assays, if available}

\item{\code{image}}{image of the spatial assay, bitmap class}

\item{\code{params}}{additional parameters used by different assay types}

\item{\code{type}}{the type of the assay (cell, spot, ROI)}
}}

