% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/metadata.R,
%   R/assay.R
\name{vrAssayNames}
\alias{vrAssayNames}
\alias{vrAssayNames.VoltRon}
\alias{vrAssayNames.vrMetadata}
\alias{vrAssayNames.vrAssay}
\alias{vrAssayNames.vrAssayV2}
\alias{vrAssayNames<-.vrAssay}
\alias{vrAssayNames<-.vrAssayV2}
\title{Get Assay names}
\usage{
vrAssayNames(object, assay = NULL)

\method{vrAssayNames}{VoltRon}(object, assay = NULL)

\method{vrAssayNames}{vrMetadata}(object)

\method{vrAssayNames}{vrAssay}(object)

\method{vrAssayNames}{vrAssayV2}(object)

\method{vrAssayNames}{vrAssay}(object) <- value

\method{vrAssayNames}{vrAssayV2}(object) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{value}{assay name}
}
\description{
Given a VoltRon, vrMetadata or vrAssay object, get/set names of assays.
}
