% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/metadata.R,
%   R/assay.R
\name{vrAssayNames}
\alias{vrAssayNames}
\alias{vrAssayNames,VoltRon-method}
\alias{vrAssayNames,vrMetadata-method}
\alias{vrAssayNames,vrAssay-method}
\alias{vrAssayNames,vrAssayV2-method}
\alias{vrAssayNames<-,vrAssay-method}
\alias{vrAssayNames<-,vrAssayV2-method}
\title{Get Assay names}
\usage{
vrAssayNames(object, ...)

\S4method{vrAssayNames}{VoltRon}(object, assay = NULL)

\S4method{vrAssayNames}{vrMetadata}(object)

\S4method{vrAssayNames}{vrAssay}(object)

\S4method{vrAssayNames}{vrAssayV2}(object)

\S4method{vrAssayNames}{vrAssay}(object) <- value

\S4method{vrAssayNames}{vrAssayV2}(object) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{value}{assay name}
}
\description{
Given a VoltRon, vrMetadata or vrAssay object, get/set names of assays.
}
