% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{vrAssayParams}
\alias{vrAssayParams}
\title{Get assay parameters}
\usage{
vrAssayParams(object, param = NULL)
}
\arguments{
\item{object}{a vrAssay object}

\item{param}{the parameter value to return}
}
\description{
Given a vrAssay object, if there are any, get a list of parameters of the assay(s)
}
