% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrAssayTypes.vrAssay}
\alias{vrAssayTypes.vrAssay}
\alias{vrAssayTypes}
\alias{vrAssayTypes.VoltRon}
\title{Get Assay types}
\usage{
\method{vrAssayTypes}{vrAssay}(object)

vrAssayTypes(object, ...)

\method{vrAssayTypes}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Given a VoltRon or vrAssay object, get types of assays. 
Here, an assay type is of either tile, molecule, cell, spot or ROI.
}
