% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrAssayTypes}
\alias{vrAssayTypes}
\alias{vrAssayTypes.VoltRon}
\alias{vrAssayTypes.vrAssay}
\alias{vrAssayTypes.vrAssayV2}
\title{Get assay types}
\usage{
vrAssayTypes(object, assay = NULL)

\method{vrAssayTypes}{VoltRon}(object, assay = NULL)

\method{vrAssayTypes}{vrAssay}(object)

\method{vrAssayTypes}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Given a VoltRon or vrAssay object, get types of assays. 
Here, an assay type is of either tile, molecule, cell, spot or ROI.
}
