% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrAssayTypes.vrAssay}
\alias{vrAssayTypes.vrAssay}
\alias{vrAssayTypes}
\alias{vrAssayTypes.VoltRon}
\title{Get Assay types}
\usage{
\method{vrAssayTypes}{vrAssay}(object, ...)

vrAssayTypes(object, ...)

\method{vrAssayTypes}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{vrAssayTypes}: The name of the default assay
}
\description{
Given a VoltRon object, give types of assays of some type, or of the main assay.
Here, an assay type is of either cell, spot or ROI.
}
\concept{data-access}
