% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrBarPlot}
\alias{vrBarPlot}
\title{vrBarPlot}
\usage{
vrBarPlot(
  object,
  features = NULL,
  assay = NULL,
  x.label = NULL,
  group.by = "Sample",
  split.by = NULL,
  norm = TRUE,
  log = FALSE,
  ncol = 2,
  nrow = NULL
)
}
\arguments{
\item{object}{a VoltRon object}

\item{features}{a set of features to be visualized, either from 
\link{vrFeatures} of raw or normalized data or columns of the 
\link{Metadata}.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{x.label}{labels of the x axis}

\item{group.by}{a column of metadata from \link{Metadata} used as 
grouping label for the spatial entities}

\item{split.by}{the column to split the barplots by}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) 
will be log transformed}

\item{ncol}{column wise number of plots, for \link{ggarrange}}

\item{nrow}{row wise number of plots, for \link{ggarrange}}
}
\description{
vrBarPlot
}
