% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrBarPlot}
\alias{vrBarPlot}
\title{vrBarPlot}
\usage{
vrBarPlot(
  object,
  features = NULL,
  assay = NULL,
  x.label = "ROI.name",
  group.by = "ROI.name",
  norm = TRUE,
  ncol = 2,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{A VoltRon object}

\item{features}{a set of features to be visualized}

\item{assay}{assay name}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{norm}{if TRUE, the normalized data is used}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrBarPlot
}
