% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R,
%   R/image.R
\name{vrCoordinates}
\alias{vrCoordinates}
\alias{vrCoordinates,VoltRon-method}
\alias{vrCoordinates,vrAssay-method}
\alias{vrCoordinates,vrAssayV2-method}
\alias{vrCoordinates,vrImage-method}
\alias{vrCoordinates,vrSpatial-method}
\alias{vrCoordinates<-,VoltRon-method}
\alias{vrCoordinates<-,vrAssay-method}
\alias{vrCoordinates<-,vrAssayV2-method}
\alias{vrCoordinates<-,vrImage-method}
\alias{vrCoordinates<-,vrSpatial-method}
\alias{vrCoordinates<-}
\title{vrCoordinates}
\usage{
vrCoordinates(object, ...)

\S4method{vrCoordinates}{VoltRon}(
  object,
  assay = NULL,
  image_name = NULL,
  spatial = NULL,
  reg = FALSE
)

\S4method{vrCoordinates}{vrAssay}(object, image_name = NULL, spatial = NULL, reg = FALSE)

\S4method{vrCoordinates}{vrAssayV2}(object, image_name = NULL, spatial = NULL, reg = FALSE)

\S4method{vrCoordinates}{vrImage}(object)

\S4method{vrCoordinates}{vrSpatial}(object)

\S4method{vrCoordinates}{VoltRon}(
  object,
  assay = NULL,
  image_name = NULL,
  spatial = NULL,
  reg = FALSE
) <- value

\S4method{vrCoordinates}{vrAssay}(object, image_name = NULL, spatial = NULL, reg = FALSE) <- value

\S4method{vrCoordinates}{vrAssayV2}(object, image_name = NULL, spatial = NULL, reg = FALSE) <- value

\S4method{vrCoordinates}{vrImage}(object) <- value

\S4method{vrCoordinates}{vrSpatial}(object) <- value

vrCoordinates(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{image_name}{(deprecated, use \code{spatial}) the name/key of the image associated with the coordinates}

\item{spatial}{the name of the spatial coordinate system}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainImage}) is requested}

\item{value}{coordinates}
}
\description{
Given a VoltRon, vrAssay or vrSpatial object, get and set the coordinates of assays and coordinate systems
}
