% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/objects.R
\name{vrCoordinates.vrAssay}
\alias{vrCoordinates.vrAssay}
\alias{vrCoordinates<-.vrAssay}
\alias{vrCoordinates}
\alias{vrCoordinates<-}
\alias{vrCoordinates.vrImage}
\alias{vrCoordinates<-.vrImage}
\alias{vrCoordinates.VoltRon}
\alias{vrCoordinates<-.VoltRon}
\title{#' @param reg TRUE if registered segments are being updated
#'
#' @rdname vrCoordinates
#' @method vrCoordinates vrAssay
#'
#' @export
#'
vrCoordinates.vrAssay <- function(object, reg = FALSE, ...) {
  if(reg){
    if(nrow(object@coords_reg) < 1) {
      return(object@coords)
    } else {
      return(object@coords_reg)
    }
  } else {
    return(object@coords)
  }
}}
\usage{
\method{vrCoordinates}{vrAssay}(object, main_image = NULL, reg = FALSE)

\method{vrCoordinates}{vrAssay}(object, main_image = NULL, reg = FALSE, ...) <- value

vrCoordinates(object, ...)

vrCoordinates(object, ...) <- value

\method{vrCoordinates}{vrImage}(object)

\method{vrCoordinates}{vrImage}(object, ...) <- value

\method{vrCoordinates}{VoltRon}(object, reg = FALSE, assay = NULL, ...)

\method{vrCoordinates}{VoltRon}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{main_image}{the key of the image associated with the coordinates}

\item{reg}{TRUE if registered segments are being updated}

\item{...}{Arguments passed to other methods}

\item{value}{the new set of 2D coordinates}

\item{assay}{assay}
}
\description{
Get and set the coordinates of the main assay
}
