% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/objects.R
\name{vrCoordinates.vrAssay}
\alias{vrCoordinates.vrAssay}
\alias{vrCoordinates<-.vrAssay}
\alias{vrCoordinates}
\alias{vrCoordinates<-}
\alias{vrCoordinates.vrImage}
\alias{vrCoordinates<-.vrImage}
\alias{vrCoordinates.VoltRon}
\alias{vrCoordinates<-.VoltRon}
\title{vrCoordinates}
\usage{
\method{vrCoordinates}{vrAssay}(object, image_name = NULL, reg = FALSE)

\method{vrCoordinates}{vrAssay}(object, image_name = NULL, reg = FALSE) <- value

vrCoordinates(object, assay = NULL, image_name = NULL, reg = FALSE)

vrCoordinates(object, image_name = NULL, reg = FALSE) <- value

\method{vrCoordinates}{vrImage}(object)

\method{vrCoordinates}{vrImage}(object) <- value

\method{vrCoordinates}{VoltRon}(object, assay = NULL, image_name = NULL, reg = FALSE)

\method{vrCoordinates}{VoltRon}(object, image_name = NULL, reg = FALSE) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{image_name}{the key of the image associated with the coordinates}

\item{reg}{TRUE if registered coordinates of the main image (\code{vrMainImage}) is requested}

\item{value}{new coordinates of spatial points}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Given a VoltRon, vrAssay or vrImage object, get and set the coordinates of assays and coordinate systems
}
