% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrCoordinates.vrAssay}
\alias{vrCoordinates.vrAssay}
\alias{vrCoordinates<-.vrAssay}
\alias{vrCoordinates}
\alias{vrCoordinates<-}
\alias{vrCoordinates.VoltRon}
\alias{vrCoordinates<-.VoltRon}
\title{vrCoordinates}
\usage{
\method{vrCoordinates}{vrAssay}(object, reg = FALSE, ...)

\method{vrCoordinates}{vrAssay}(object, reg = FALSE, ...) <- value

vrCoordinates(object, ...)

vrCoordinates(object, ...) <- value

\method{vrCoordinates}{VoltRon}(object, reg = FALSE, assay = NULL, ...)

\method{vrCoordinates}{VoltRon}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{reg}{TRUE if registered segments are being updated}

\item{...}{Arguments passed to other methods}

\item{value}{the new set of 2D coordinates}

\item{assay}{assay}
}
\value{
\code{vrCoordinates}: The name of the default assay

\code{vrCoordinates<-}: The coordinates updated
}
\description{
Get and set the coordinates of the main assay
}
\concept{data-access}
