% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrData.vrAssay}
\alias{vrData.vrAssay}
\alias{vrData}
\alias{vrData.VoltRon}
\title{vrData}
\usage{
\method{vrData}{vrAssay}(object, features = NULL, norm = FALSE, ...)

vrData(object, ...)

\method{vrData}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{features}{features}

\item{norm}{TRUE if normalized data should be returned}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{tile_size}{the size of tiles if an assay is of type tile}
}
\description{
Get data from the main.assay
}
