% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{vrData}
\alias{vrData}
\alias{vrData,VoltRon-method}
\alias{vrData,vrAssay-method}
\alias{vrData,vrAssayV2-method}
\title{vrData}
\usage{
vrData(object, ...)

\S4method{vrData}{VoltRon}(
  object,
  assay = NULL,
  features = NULL,
  feat_type = NULL,
  norm = FALSE,
  ...
)

\S4method{vrData}{vrAssay}(object, features = NULL, feat_type = NULL, norm = FALSE, ...)

\S4method{vrData}{vrAssayV2}(object, features = NULL, feat_type = NULL, norm = FALSE, ...)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{additional parameters passed to other methods and \link{vrImages}}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{features}{the set of features}

\item{feat_type}{the feature set type}

\item{norm}{TRUE if normalized data should be returned}
}
\description{
Get data of assays
}
