% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrData.vrAssay}
\alias{vrData.vrAssay}
\alias{vrData}
\alias{vrData.VoltRon}
\title{vrData}
\usage{
\method{vrData}{vrAssay}(object, features = NULL, norm = FALSE, ...)

vrData(object, ...)

\method{vrData}{VoltRon}(object, assay = NULL, norm = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{features}{features}

\item{norm}{TRUE if normalized data should be returned}

\item{...}{additional parameters passed to \code{vrData.vrAssay}}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Get data from the main.assay
}
