% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrData}
\alias{vrData}
\alias{vrData.VoltRon}
\alias{vrData.vrAssay}
\title{vrData}
\usage{
vrData(object, assay = NULL, features = NULL, norm = FALSE)

\method{vrData}{VoltRon}(object, assay = NULL, features = NULL, norm = FALSE, ...)

\method{vrData}{vrAssay}(object, features = NULL, norm = FALSE, ...)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{features}{the set of features}

\item{norm}{TRUE if normalized data should be returned}

\item{...}{additional parameters passed to other methods and \link{vrImages}}
}
\description{
Get data of assays
}
