% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R
\name{vrDistances.vrAssay}
\alias{vrDistances.vrAssay}
\alias{vrDistances}
\title{vrDistances}
\usage{
\method{vrDistances}{vrAssay}(object, reg = FALSE, method = "euclidean", ...)

vrDistances(object, ...)
}
\arguments{
\item{object}{An object}

\item{reg}{TRUE if registered segments are being updated}

\item{method}{the method argument passed to \code{base::dist}}

\item{...}{Arguments passed to other methods}
}
\description{
Get distances between spatial points using coordinates
}
