% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingFeaturePlot}
\alias{vrEmbeddingFeaturePlot}
\title{vrEmbeddingFeaturePlot}
\usage{
vrEmbeddingFeaturePlot(
  object,
  embedding = "pca",
  features = NULL,
  norm = TRUE,
  log = FALSE,
  assay = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 1,
  keep.scale = "feature",
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{features}{a set of features, either from the rows of rawdata, normdata or columns of the metadata}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) will be log transformed}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{keep.scale}{whether unify all scales for all features or not}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting features of spatially resolved cells and spots on embeddings from multiple assays in a VoltRon object.
}
