% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingFeaturePlot}
\alias{vrEmbeddingFeaturePlot}
\title{vrEmbeddingFeaturePlot}
\usage{
vrEmbeddingFeaturePlot(
  object,
  embedding = "pca",
  features = NULL,
  combine.features = FALSE,
  n.tile = 0,
  norm = TRUE,
  log = FALSE,
  assay = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 1,
  keep.scale = "feature",
  common.legend = TRUE,
  collapse.plots = TRUE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{features}{a set of features to be visualized, either from \link{vrFeatures} of raw or normalized data or columns of the \link{Metadata}.}

\item{combine.features}{whether to combine all features in one plot}

\item{n.tile}{should points be aggregated into tiles before visualization (see \link{geom_tile}). Applicable only for cells and molecules}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) will be log transformed}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{ncol}{column wise number of plots, for \link{ggarrange}}

\item{nrow}{row wise number of plots, for \link{ggarrange}}

\item{font.size}{font size}

\item{pt.size}{point size}

\item{keep.scale}{whether unify all scales for all features or not}

\item{common.legend}{whether to use a common legend for all plots, see \link{ggarrange}}

\item{collapse.plots}{whether to combine all ggplots}
}
\description{
Plotting features of spatially resolved cells and spots on embeddings from multiple assays in a VoltRon object.
}
