% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrEmbeddingNames.vrAssay}
\alias{vrEmbeddingNames.vrAssay}
\alias{vrEmbeddingNames}
\alias{vrEmbeddingNames.VoltRon}
\title{Get Embedding names}
\usage{
\method{vrEmbeddingNames}{vrAssay}(object)

vrEmbeddingNames(object, ...)

\method{vrEmbeddingNames}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}
}
\value{
\code{vrEmbeddingNames}: The name of embeddings
}
\description{
Given a VoltRon object, give names of embeddings
}
