% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrEmbeddingNames}
\alias{vrEmbeddingNames}
\alias{vrEmbeddingNames.VoltRon}
\alias{vrEmbeddingNames.vrAssay}
\title{Get Embedding names}
\usage{
vrEmbeddingNames(object, assay = NULL)

\method{vrEmbeddingNames}{VoltRon}(object, assay = NULL)

\method{vrEmbeddingNames}{vrAssay}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Given a VoltRon or vrAssay object, give names of embeddings
}
