% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrEmbeddingNames.vrAssay}
\alias{vrEmbeddingNames.vrAssay}
\alias{vrEmbeddingNames}
\alias{vrEmbeddingNames.VoltRon}
\title{Get Embedding names}
\usage{
\method{vrEmbeddingNames}{vrAssay}(object)

vrEmbeddingNames(object, assay = NULL)

\method{vrEmbeddingNames}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Given a VoltRon or vrAssay object, give names of embeddings
}
