% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingPlot}
\alias{vrEmbeddingPlot}
\title{vrEmbeddingPlot}
\usage{
vrEmbeddingPlot(
  object,
  embedding = "pca",
  group.by = "Sample",
  group.ids = NULL,
  split.by = NULL,
  colors = NULL,
  assay = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 5,
  pt.size = 1,
  label = FALSE,
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{group.ids}{a subset of categories defined in metadata column from \code{group.by}}

\item{split.by}{a column of metadata from \link{Metadata} used as splitting spatial entities into ggplot panels, see \link{facet_wrap}}

\item{colors}{the color set for group.by. Should be of the same size of group.id (if specified) or unique elements in group.by}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{ncol}{column wise number of plots, for \link{facet_wrap}}

\item{nrow}{row wise number of plots, for \link{facet_wrap}}

\item{font.size}{font size of labels, if label is TRUE}

\item{pt.size}{point size}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{common.legend}{whether to use a common legend for all plots, see \link{ggarrange}}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting embeddings of cells and spots on associated images from multiple assays in a VoltRon object.
}
