% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrEmbeddingPlot}
\alias{vrEmbeddingPlot}
\title{vrEmbeddingPlot}
\usage{
vrEmbeddingPlot(
  object,
  embedding = "pca",
  group.by = "Sample",
  group.ids = NULL,
  colors = NULL,
  assay = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 5,
  pt.size = 1,
  label = FALSE,
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{embedding}{the embedding type, i.e. pca, umap etc.}

\item{group.by}{a grouping label for the spatial entities}

\item{group.ids}{a subset of categories defined with in the grouping label \code{group.by}}

\item{colors}{the color set for group.by. Should be of the same size of group.id (if specified) or unique elements in group.by}

\item{assay}{the assay name}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font size of labels, if label is TRUE}

\item{pt.size}{point size}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting embeddings of cells and spots on associated images from multiple assays in a VoltRon object.
}
