% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrEmbeddings}
\alias{vrEmbeddings}
\alias{vrEmbeddings.VoltRon}
\alias{vrEmbeddings.vrAssay}
\alias{vrEmbeddings.vrAssayV2}
\alias{vrEmbeddings<-.vrAssay}
\alias{vrEmbeddings<-.vrAssayV2}
\alias{vrEmbeddings<-.VoltRon}
\title{vrEmbeddings}
\usage{
vrEmbeddings(object, assay = NULL, type = "pca", dims = 1:30)

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", dims = 1:30)

\method{vrEmbeddings}{vrAssay}(object, type = "pca", dims = 1:30)

\method{vrEmbeddings}{vrAssayV2}(object, type = "pca", dims = 1:30)

\method{vrEmbeddings}{vrAssay}(object, type = "pca") <- value

\method{vrEmbeddings}{vrAssayV2}(object, type = "pca") <- value

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", overwrite = FALSE) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{type}{the key name for the embedding}

\item{dims}{the set of dimensions of the embedding data}

\item{value}{new embedding data}

\item{overwrite}{Whether the existing embedding with name 'type' should be overwritten}
}
\description{
Given a VoltRon or vrAssay object, get embeddings of spatial points
}
