% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrEmbeddings.vrAssay}
\alias{vrEmbeddings.vrAssay}
\alias{vrEmbeddings<-.vrAssay}
\alias{vrEmbeddings}
\alias{vrEmbeddings<-}
\alias{vrEmbeddings.VoltRon}
\alias{vrEmbeddings<-.VoltRon}
\title{vrEmbeddings}
\usage{
\method{vrEmbeddings}{vrAssay}(object, type = "pca", dims = 1:30)

\method{vrEmbeddings}{vrAssay}(object, type = "pca", ...) <- value

vrEmbeddings(object, ...)

vrEmbeddings(object, ...) <- value

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", dims = 1:30, ...)

\method{vrEmbeddings}{VoltRon}(object, assay = NULL, type = "pca", overwrite = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{type}{the key name for the embedding}

\item{dims}{the set of dimensions of the embedding data}

\item{...}{Arguments passed to other methods}

\item{value}{new embedding data}

\item{assay}{assay}

\item{overwrite}{Whether the existing embedding with name 'type' should be overwritten}
}
\description{
Get embeddings of spatial points
}
