% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrFeatureData.vrAssay}
\alias{vrFeatureData.vrAssay}
\alias{vrFeatureData<-.vrAssay}
\alias{vrFeatureData}
\alias{vrFeatureData<-}
\alias{vrFeatureData.VoltRon}
\alias{vrFeatureData<-.VoltRon}
\title{Feature Data}
\usage{
\method{vrFeatureData}{vrAssay}(object, ...)

\method{vrFeatureData}{vrAssay}(object, ...) <- value

vrFeatureData(object, ...)

vrFeatureData(object, ...) <- value

\method{vrFeatureData}{VoltRon}(object, assay = NULL, ...)

\method{vrFeatureData}{VoltRon}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{new Feature Data}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Get and set feature data from the main.assay
}
