% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrFeatureData}
\alias{vrFeatureData}
\alias{vrFeatureData.VoltRon}
\alias{vrFeatureData.vrAssay}
\alias{vrFeatureData<-.VoltRon}
\alias{vrFeatureData<-.vrAssay}
\title{Feature Data}
\usage{
vrFeatureData(object, assay = NULL)

\method{vrFeatureData}{VoltRon}(object, assay = NULL)

\method{vrFeatureData}{vrAssay}(object)

\method{vrFeatureData}{VoltRon}(object, assay = NULL) <- value

\method{vrFeatureData}{vrAssay}(object) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Get and set feature data from the main.assay
}
