% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrFeatureData.vrAssay}
\alias{vrFeatureData.vrAssay}
\alias{vrFeatureData<-.vrAssay}
\alias{vrFeatureData}
\alias{vrFeatureData<-}
\alias{vrFeatureData.VoltRon}
\title{Feature Data}
\usage{
\method{vrFeatureData}{vrAssay}(object, ...)

\method{vrFeatureData}{vrAssay}(object, ...) <- value

vrFeatureData(object, ...)

vrFeatureData(object, ...) <- value

\method{vrFeatureData}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{new feature data}

\item{assay}{assay}
}
\value{
\code{vrFeatureData}: The name of the default assay

\code{vrFeatureData<-}: The coordinates updated
}
\description{
Get and set feature data from the main.assay
}
\concept{data-access}
