% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R
\name{vrFeatures}
\alias{vrFeatures}
\alias{vrFeatures.VoltRon}
\alias{vrFeatures.vrAssay}
\alias{vrFeatures.vrAssayV2}
\title{vrFeatures}
\usage{
vrFeatures(object, assay = NULL)

\method{vrFeatures}{VoltRon}(object, assay = NULL)

\method{vrFeatures}{vrAssay}(object)

\method{vrFeatures}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Get names of the features.
}
