% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrFeatures.vrAssay}
\alias{vrFeatures.vrAssay}
\alias{vrFeatures}
\alias{vrFeatures.VoltRon}
\title{vrFeatures}
\usage{
\method{vrFeatures}{vrAssay}(object, ...)

vrFeatures(object, ...)

\method{vrFeatures}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{vrFeatures}: The name of the default assay
}
\description{
Get features from the main.assay
}
\concept{data-access}
