% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{vrFeatures}
\alias{vrFeatures}
\alias{vrFeatures,VoltRon-method}
\alias{vrFeatures,vrAssay-method}
\alias{vrFeatures,vrAssayV2-method}
\title{vrFeatures}
\usage{
vrFeatures(object, ...)

\S4method{vrFeatures}{VoltRon}(object, assay = NULL)

\S4method{vrFeatures}{vrAssay}(object)

\S4method{vrFeatures}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Get names of the features.
}
