% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/assay.R, R/objects.R
\name{vrFeatureTypeNames}
\alias{vrFeatureTypeNames}
\alias{vrFeatureTypeNames,vrAssayV2-method}
\alias{vrFeatureTypeNames,vrAssay-method}
\alias{vrFeatureTypeNames,VoltRon-method}
\title{vrFeatureTypeNames}
\usage{
vrFeatureTypeNames(object, ...)

\S4method{vrFeatureTypeNames}{vrAssayV2}(object)

\S4method{vrFeatureTypeNames}{vrAssay}(object)

\S4method{vrFeatureTypeNames}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
If NULL, the default assay will be used, see \link{vrMainAssay}. If given as "all", then provides a summary of spatial systems across all assays}
}
\description{
Get the names of the feature types
}
