% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{vrGraph}
\alias{vrGraph}
\alias{vrGraph.VoltRon}
\title{vrGraph}
\usage{
vrGraph(object, ...)

\method{vrGraph}{VoltRon}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}
}
\value{
\code{vrGraph}: The name of the default assay
}
\description{
Get graph from the main.assay
}
