% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{vrGraph}
\alias{vrGraph}
\alias{vrGraph<-}
\title{vrGraph}
\usage{
vrGraph(object, assay = NULL, graph.type = "kNN")

vrGraph(object, graph.type = "kNN") <- value
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{graph.type}{the type of the graph, either custom or given by \link{getProfileNeighbors} or \link{getSpatialNeighbors} functions}

\item{value}{new graph}
}
\description{
Get graph of a VoltRon object
}
