% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{vrGraph}
\alias{vrGraph}
\alias{vrGraph<-}
\alias{vrGraph.VoltRon}
\alias{vrGraph<-.VoltRon}
\title{vrGraph}
\usage{
vrGraph(object, assay = NULL, graph.type = "kNN")

vrGraph(object, graph.type = "kNN") <- value

\method{vrGraph}{VoltRon}(object, assay = NULL, graph.type = "kNN")

\method{vrGraph}{VoltRon}(object, graph.type = "kNN") <- value
}
\arguments{
\item{object}{A VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{graph.type}{the type of the graph, either custom or given by \code{getProfileNeighbors} or \code{getSpatialNeighbors} functions}

\item{value}{new graph}
}
\description{
Get graph of a VoltRon object
}
