% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{vrGraphNames}
\alias{vrGraphNames}
\alias{vrGraphNames.VoltRon}
\title{vrGraphNames}
\usage{
vrGraphNames(object, ...)

\method{vrGraphNames}{VoltRon}(object, assay = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}
}
\description{
Get names of all graphs
}
