% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{vrGraphNames}
\alias{vrGraphNames}
\title{vrGraphNames}
\usage{
vrGraphNames(object, assay = NULL)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}
}
\description{
Get names of all graphs
}
