% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrHeatmapPlot}
\alias{vrHeatmapPlot}
\title{HeatmapPlot}
\usage{
vrHeatmapPlot(
  object,
  assay = NULL,
  assay.type = NULL,
  group.by = "clusters",
  norm = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{assay}{assay name}

\item{assay.type}{assay type}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{norm}{if TRUE, the normalized data is used}
}
\description{
HeatmapPlot
}
