% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrHeatmapPlot}
\alias{vrHeatmapPlot}
\title{HeatmapPlot}
\usage{
vrHeatmapPlot(
  object,
  assay = NULL,
  assay.type = NULL,
  features = NULL,
  group.by = "clusters",
  norm = TRUE,
  scaled = TRUE,
  show_row_names = NULL,
  outlier.quantile = 0.99,
  ...
)
}
\arguments{
\item{object}{VoltRon object}

\item{assay}{assay name}

\item{assay.type}{assay type}

\item{features}{a set of features to be visualized}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{norm}{if TRUE, the normalized data is used}

\item{scaled}{if TRUE, the data will be scaled before visualization}

\item{show_row_names}{if TRUE, row names of the heatmap will be shown}

\item{outlier.quantile}{quantile for detecting outliers whose values are set to the quantile, change to lower values to adjust large number of outliers, default: 0.99}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
HeatmapPlot
}
