% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrHeatmapPlot}
\alias{vrHeatmapPlot}
\title{HeatmapPlot}
\usage{
vrHeatmapPlot(
  object,
  assay = NULL,
  features = NULL,
  group.by = "clusters",
  norm = TRUE,
  scaled = TRUE,
  show_row_names = NULL,
  cluster_rows = TRUE,
  show_heatmap_legend = FALSE,
  outlier.quantile = 0.99,
  highlight.some = FALSE,
  n_highlight = 30,
  font.size = 13.2,
  ...
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{features}{a set of features to be visualized from \link{vrFeatures} of raw or normalized data.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{norm}{if TRUE, the normalized data is used}

\item{scaled}{if TRUE, the data will be scaled before visualization}

\item{show_row_names}{if TRUE, row names of the heatmap will be shown}

\item{cluster_rows}{if TRUE, the rows of the heatmap will be clustered}

\item{show_heatmap_legend}{if TRUE, the heatmap legend is shown}

\item{outlier.quantile}{quantile for detecting outliers whose values are set to the quantile, change to lower values to adjust large number of outliers, default: 0.99}

\item{highlight.some}{if TRUE, some rows will be showed at random, reproducible by \code{seed} argument}

\item{n_highlight}{the number of row labels shown, if \code{show_row_names} is TRUE}

\item{font.size}{font size}

\item{...}{additional parameters passed to \link{getVariableFeatures}}
}
\description{
HeatmapPlot
}
