% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{class}
\name{vrImage-class}
\alias{vrImage-class}
\alias{vrImage}
\title{The vrImage (VoltRon Image) Class}
\description{
The vrImage (VoltRon Image) Class
}
\section{Slots}{

\describe{
\item{\code{coords}}{spatial coordinates of the assay}

\item{\code{segments}}{spatial coordinates of the segments, if available}

\item{\code{image}}{image of the spatial assay, bitmap class}

\item{\code{main_channel}}{the key of the main channel of vrImage object}
}}

