% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrImageChannelNames}
\alias{vrImageChannelNames}
\alias{vrImageChannelNames,VoltRon-method}
\alias{vrImageChannelNames,vrAssay-method}
\alias{vrImageChannelNames,vrAssayV2-method}
\alias{vrImageChannelNames,vrImage-method}
\alias{vrImageChannelNames,vrSpatial-method}
\title{vrImageChannelNames}
\usage{
vrImageChannelNames(object, ...)

\S4method{vrImageChannelNames}{VoltRon}(object, assay = NULL, ...)

\S4method{vrImageChannelNames}{vrAssay}(object, name = NULL, ...)

\S4method{vrImageChannelNames}{vrAssayV2}(object, name = NULL, ...)

\S4method{vrImageChannelNames}{vrImage}(object, return.report = TRUE)

\S4method{vrImageChannelNames}{vrSpatial}(object, return.report = TRUE)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the key of the image}

\item{return.report}{if TRUE and no image is present, return a character 
stating that there is no image}
}
\description{
Get names of all image channels
}
