% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImageChannelNames}
\alias{vrImageChannelNames}
\alias{vrImageChannelNames.VoltRon}
\alias{vrImageChannelNames.vrAssay}
\alias{vrImageChannelNames.vrImage}
\title{vrImageChannelNames}
\usage{
vrImageChannelNames(object, ...)

\method{vrImageChannelNames}{VoltRon}(object, assay = NULL)

\method{vrImageChannelNames}{vrAssay}(object, main_image = NULL)

\method{vrImageChannelNames}{vrImage}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}

\item{main_image}{the key of the main image}
}
\description{
Get names of all image channels
}
