% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImageChannelNames}
\alias{vrImageChannelNames}
\alias{vrImageChannelNames.VoltRon}
\alias{vrImageChannelNames.vrAssay}
\alias{vrImageChannelNames.vrImage}
\title{vrImageChannelNames}
\usage{
vrImageChannelNames(object, assay = NULL, name = NULL)

\method{vrImageChannelNames}{VoltRon}(object, assay = NULL)

\method{vrImageChannelNames}{vrAssay}(object, name = NULL)

\method{vrImageChannelNames}{vrImage}(object)
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{name}{the key of the image}
}
\description{
Get names of all image channels
}
