% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImageNames}
\alias{vrImageNames}
\alias{vrImageNames.VoltRon}
\alias{vrImageNames.vrAssay}
\title{vrImageNames}
\usage{
vrImageNames(object, assay = NULL)

\method{vrImageNames}{VoltRon}(object, assay = NULL)

\method{vrImageNames}{vrAssay}(object)
}
\arguments{
\item{object}{A VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Get names of all images
}
