% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImageNames}
\alias{vrImageNames}
\alias{vrImageNames.VoltRon}
\alias{vrImageNames.vrAssay}
\title{vrImageNames}
\usage{
vrImageNames(object, ...)

\method{vrImageNames}{VoltRon}(object, assay = NULL)

\method{vrImageNames}{vrAssay}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{assay}
}
\description{
Get names of all images
}
