% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrImageNames}
\alias{vrImageNames}
\alias{vrImageNames.VoltRon}
\alias{vrImageNames.vrAssay}
\alias{vrImageNames.vrAssayV2}
\title{vrImageNames}
\usage{
vrImageNames(object, assay = NULL)

\method{vrImageNames}{VoltRon}(object, assay = NULL)

\method{vrImageNames}{vrAssay}(object)

\method{vrImageNames}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
If NULL, the default assay will be used, see \link{vrMainAssay}. If equals to "all", then provides a summary of spatial systems across all assays}
}
\description{
Get names of all images
}
