% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrImages}
\alias{vrImages}
\alias{vrImages,VoltRon-method}
\alias{vrImages,vrAssay-method}
\alias{vrImages,vrAssayV2-method}
\alias{vrImages,vrImage-method}
\alias{vrImages,vrSpatial-method}
\alias{vrImages<-,vrAssay-method}
\alias{vrImages<-,vrAssayV2-method}
\alias{vrImages<-,vrImage-method}
\alias{vrImages<-,vrSpatial-method}
\alias{vrImages<-}
\title{vrImages}
\usage{
vrImages(object, ...)

\S4method{vrImages}{VoltRon}(
  object,
  assay = NULL,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  as.raster = FALSE,
  scale.perc = 100
)

\S4method{vrImages}{vrAssay}(
  object,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  as.raster = FALSE,
  scale.perc = 100
)

\S4method{vrImages}{vrAssayV2}(
  object,
  spatial = NULL,
  reg = FALSE,
  channel = NULL,
  as.raster = FALSE,
  scale.perc = 100
)

\S4method{vrImages}{vrImage}(object, channel = NULL, as.raster = FALSE, scale.perc = 100)

\S4method{vrImages}{vrSpatial}(object, channel = NULL, as.raster = FALSE, scale.perc = 100)

\S4method{vrImages}{vrAssay}(object, spatial = NULL, channel = NULL, reg = FALSE) <- value

\S4method{vrImages}{vrAssayV2}(object, spatial = NULL, channel = NULL, reg = FALSE) <- value

\S4method{vrImages}{vrImage}(object, channel = NULL) <- value

\S4method{vrImages}{vrSpatial}(object, channel = NULL) <- value

vrImages(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{spatial}{the name of the main spatial system}

\item{reg}{TRUE if registered coordinates of the main image 
(\link{vrMainSpatial}) is requested}

\item{channel}{the name of the channel associated with the image}

\item{as.raster}{return as raster}

\item{scale.perc}{scale percentage if lower resolution image needed}

\item{value}{a raster image or an image of \code{magick-image} object}
}
\description{
Get images of VoltRon objects
}
