% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImages}
\alias{vrImages}
\alias{vrImages<-}
\alias{vrImages.VoltRon}
\alias{vrImages.vrAssay}
\alias{vrImages<-.vrAssay}
\alias{vrImages.vrImage}
\alias{vrImages<-.vrImage}
\title{vrImages}
\usage{
vrImages(object, ...)

vrImages(object, ...) <- value

\method{vrImages}{VoltRon}(object, assay = NULL, ...)

\method{vrImages}{vrAssay}(object, name = NULL, reg = FALSE, channel = NULL, ...)

\method{vrImages}{vrAssay}(object, name = NULL, channel = NULL, reg = FALSE, ...) <- value

\method{vrImages}{vrImage}(object, channel = NULL, as.raster = FALSE)

\method{vrImages}{vrImage}(object, channel = NULL) <- value
}
\arguments{
\item{object}{A vrAssay object}

\item{...}{arguements passed to \code{vrImages.vrImage}}

\item{value}{new image}

\item{assay}{assay}

\item{name}{the name of the main image}

\item{reg}{TRUE if registered images are assigned}

\item{channel}{the name of the channel associated with the main image}

\item{as.raster}{return as a raster}
}
\description{
Get images of a VoltRon assay

#' @param object A vrLayer object
#' @param ... arguements passed to \code{vrImages.vrAssay}
#'
#' @rdname vrImages
#' @method vrImages vrLayer
#'
#' @export
#'
vrImages.vrLayer <- function(object, ...){
  sapply(object@assay, function(assy) vrImages(assy, ...), USE.NAMES = TRUE)
}
}
