% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImages}
\alias{vrImages}
\alias{vrImages<-}
\alias{vrImages.VoltRon}
\alias{vrImages.vrSample}
\alias{vrImages.vrLayer}
\alias{vrImages.vrAssay}
\alias{vrImages<-.vrAssay}
\alias{vrImages.vrImage}
\title{vrImages}
\usage{
vrImages(object, ...)

vrImages(object, ...) <- value

\method{vrImages}{VoltRon}(object, ...)

\method{vrImages}{vrSample}(object, ...)

\method{vrImages}{vrLayer}(object, ...)

\method{vrImages}{vrAssay}(object, main_image = NULL, reg = FALSE, as.raster = FALSE)

\method{vrImages}{vrAssay}(object, main_image = NULL, reg = FALSE, ...) <- value

\method{vrImages}{vrImage}(object, main_channel = NULL, as.raster = FALSE)
}
\arguments{
\item{object}{A vrImage object}

\item{...}{Arguments passed to other methods}

\item{value}{new image}

\item{main_image}{the name of the main image}

\item{reg}{TRUE if registered images are assigned}

\item{as.raster}{return as a raster}

\item{main_channel}{the name of the main channel}
}
\description{
Get images of a VoltRon assay

# get main image is main_image is null
    if(is.null(main_image)) {
      main_image <- object@main_image
    }
}
\details{

}
