% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImages}
\alias{vrImages}
\alias{vrImages.Seurat}
\alias{vrImages.VoltRon}
\alias{vrImages.vrSample}
\alias{vrImages.vrLayer}
\alias{vrImages.vrAssay}
\title{vrImages}
\usage{
vrImages(object, ...)

\method{vrImages}{Seurat}(seu, ...)

\method{vrImages}{VoltRon}(sr, ...)

\method{vrImages}{vrSample}(vr, ...)

\method{vrImages}{vrLayer}(vr, ...)

\method{vrImages}{vrAssay}(vr, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{vrImages}: The name of the default assay
}
\description{
Get images of a VoltRon assay
}
\concept{data-access}
