% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrImages}
\alias{vrImages}
\alias{vrImages<-}
\alias{vrImages.Seurat}
\alias{vrImages.VoltRon}
\alias{vrImages.vrSample}
\alias{vrImages.vrLayer}
\alias{vrImages.vrAssay}
\alias{vrImages<-.vrAssay}
\title{vrImages}
\usage{
vrImages(object, ...)

vrImages(object, ...) <- value

\method{vrImages}{Seurat}(seu, ...)

\method{vrImages}{VoltRon}(object, ...)

\method{vrImages}{vrSample}(object, ...)

\method{vrImages}{vrLayer}(object, ...)

\method{vrImages}{vrAssay}(object)

\method{vrImages}{vrAssay}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{A vrAssay object}

\item{...}{Arguments passed to other methods}

\item{value}{new image}

\item{seu}{A Seurat Object}

\item{reg}{TRUE if registered coordinates are assigned}
}
\value{
\code{vrImages}: The name of the default assay
}
\description{
Get images of a VoltRon assay
}
