% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{vrLayer-methods}
\alias{vrLayer-methods}
\alias{[[,vrLayer,character,ANY-method}
\alias{[[<-,vrLayer,character,ANY-method}
\title{Methods for vrLayer objects}
\usage{
\S4method{[[}{vrLayer,character,ANY}(x, i)

\S4method{[[}{vrLayer,character,ANY}(x, i) <- value
}
\arguments{
\item{x}{A vrLayer object}

\item{i}{the name of assay associated with the layer, see \link{SampleMetadata}}

\item{value}{a vrAssayV2 object, see \link{vrAssayV2}}
}
\description{
Methods for \code{\link{vrLayer}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{x[[i}: Accessing vrAssay objects from \code{vrLayer} objects

\item \code{`[[`(x = vrLayer, i = character, j = ANY) <- value}: Overwriting vrAssay objects from \code{vrLayer} objects

}}
\concept{vrlayer}
