% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R
\name{vrMainAssay}
\alias{vrMainAssay}
\alias{vrMainAssay,VoltRon-method}
\alias{vrMainAssay<-,VoltRon-method}
\title{Main Assay}
\usage{
vrMainAssay(object, ...)

\S4method{vrMainAssay}{VoltRon}(object)

\S4method{vrMainAssay}{VoltRon}(object) <- value
}
\arguments{
\item{object}{a VoltRon object}

\item{...}{arguments passed to other methods.}
}
\description{
Get and set the main assay of a VoltRon object

Get and set the main assay of a VoltRon object
}
