% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{vrMainAssay}
\alias{vrMainAssay}
\alias{vrMainAssay<-}
\title{Main Assay}
\usage{
vrMainAssay(object)

vrMainAssay(object) <- value
}
\arguments{
\item{object}{a VoltRon object}

\item{value}{new assay name}
}
\description{
Get and set the main assay of a VoltRon object
}
