% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrMainChannel}
\alias{vrMainChannel}
\alias{vrMainChannel<-}
\alias{vrMainChannel.vrAssay}
\alias{vrMainChannel<-.vrAssay}
\alias{vrMainChannel.vrImage}
\alias{vrMainChannel<-.vrImage}
\title{vrMainChannel}
\usage{
vrMainChannel(object, name = NULL)

vrMainChannel(object, name = NULL) <- value

\method{vrMainChannel}{vrAssay}(object, name = NULL)

\method{vrMainChannel}{vrAssay}(object, name = NULL) <- value

\method{vrMainChannel}{vrImage}(object)

\method{vrMainChannel}{vrImage}(object) <- value
}
\arguments{
\item{object}{a vrAssay or vrImage object}

\item{name}{the name of the image}

\item{value}{the name of main channel}
}
\description{
Get the main channel name of the image

Set the main channel name of the image
}
