% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrMainChannel}
\alias{vrMainChannel}
\alias{vrMainChannel,vrAssay-method}
\alias{vrMainChannel,vrAssayV2-method}
\alias{vrMainChannel,vrImage-method}
\alias{vrMainChannel,vrSpatial-method}
\alias{vrMainChannel<-,vrAssay-method}
\alias{vrMainChannel<-,vrAssayV2-method}
\alias{vrMainChannel<-,vrImage-method}
\alias{vrMainChannel<-,vrSpatial-method}
\alias{vrMainChannel<-}
\title{vrMainChannel}
\usage{
vrMainChannel(object, ...)

\S4method{vrMainChannel}{vrAssay}(object, name = NULL)

\S4method{vrMainChannel}{vrAssayV2}(object, name = NULL)

\S4method{vrMainChannel}{vrImage}(object)

\S4method{vrMainChannel}{vrSpatial}(object)

\S4method{vrMainChannel}{vrAssay}(object, name = NULL) <- value

\S4method{vrMainChannel}{vrAssayV2}(object, name = NULL) <- value

\S4method{vrMainChannel}{vrImage}(object) <- value

\S4method{vrMainChannel}{vrSpatial}(object) <- value

vrMainChannel(object, ...) <- value
}
\arguments{
\item{object}{a vrAssay or vrSpatial object}

\item{...}{arguments passed to other methods.}

\item{name}{the name of the image}

\item{value}{the name of main channel of the spatial system}
}
\description{
Get and set the main channel name of the spatial system.
}
