% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrMainChannel}
\alias{vrMainChannel}
\alias{vrMainChannel<-}
\alias{vrMainChannel.vrImage}
\alias{vrMainChannel<-.vrImage}
\title{vrMainChannel}
\usage{
vrMainChannel(object, ...)

vrMainChannel(object, ...) <- value

\method{vrMainChannel}{vrImage}(object)

\method{vrMainChannel}{vrImage}(object) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{the name of main channel}
}
\description{
Get the main channel name of the image

Set the main channel name of the image
}
