% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/assay.R
\name{vrMainFeatureType}
\alias{vrMainFeatureType}
\alias{vrMainFeatureType,VoltRon-method}
\alias{vrMainFeatureType,vrAssayV2-method}
\alias{vrMainFeatureType,vrAssay-method}
\alias{vrMainFeatureType<-,VoltRon-method}
\alias{vrMainFeatureType<-,vrAssayV2-method}
\alias{vrMainFeatureType<-,vrAssay-method}
\alias{vrMainFeatureType<-}
\title{vrMainFeatureType}
\usage{
vrMainFeatureType(object, ...)

\S4method{vrMainFeatureType}{VoltRon}(object, assay = NULL)

\S4method{vrMainFeatureType}{vrAssayV2}(object)

\S4method{vrMainFeatureType}{vrAssay}(object)

\S4method{vrMainFeatureType}{VoltRon}(object, assay = NULL) <- value

\S4method{vrMainFeatureType}{vrAssayV2}(object, ignore = FALSE) <- value

\S4method{vrMainFeatureType}{vrAssay}(object, ignore = FALSE) <- value

vrMainFeatureType(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
If NULL, the default assay will be used, see \link{vrMainAssay}. If given as "all", then provides a summary of spatial systems across all assays.}

\item{value}{the name of main feature set.}

\item{ignore}{ignore if some assays dont have the feature set name}
}
\description{
Get the main feature type

Set the main image
}
