% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrMainImage}
\alias{vrMainImage}
\alias{vrMainImage.VoltRon}
\alias{vrMainImage.vrAssay}
\alias{vrMainImage.vrAssayV2}
\alias{vrMainImage<-.VoltRon}
\alias{vrMainImage<-.vrAssay}
\alias{vrMainImage<-.vrAssayV2}
\title{vrMainImage}
\usage{
vrMainImage(object, assay = NULL)

\method{vrMainImage}{VoltRon}(object, assay = NULL)

\method{vrMainImage}{vrAssay}(object)

\method{vrMainImage}{vrAssayV2}(object)

\method{vrMainImage}{VoltRon}(object, assay = NULL) <- value

\method{vrMainImage}{vrAssay}(object, ignore = FALSE) <- value

\method{vrMainImage}{vrAssayV2}(object, ignore = FALSE) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
If NULL, the default assay will be used, see \link{vrMainAssay}. If given as "all", then provides a summary of spatial systems across all assays.}

\item{value}{the name of main image}

\item{ignore}{if TRUE, the non-existing spatial coordinate system will be ignored.}
}
\description{
Get the main image
}
