% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/image.R
\name{vrMainSpatial}
\alias{vrMainSpatial}
\alias{vrMainSpatial.VoltRon}
\alias{vrMainSpatial.vrAssay}
\alias{vrMainSpatial.vrAssayV2}
\alias{vrMainSpatial<-.VoltRon}
\alias{vrMainSpatial<-.vrAssay}
\alias{vrMainSpatial<-.vrAssayV2}
\title{vrMainSpatial}
\usage{
vrMainSpatial(object, assay = NULL)

\method{vrMainSpatial}{VoltRon}(object, assay = NULL)

\method{vrMainSpatial}{vrAssay}(object)

\method{vrMainSpatial}{vrAssayV2}(object)

\method{vrMainSpatial}{VoltRon}(object, assay = NULL) <- value

\method{vrMainSpatial}{vrAssay}(object) <- value

\method{vrMainSpatial}{vrAssayV2}(object) <- value
}
\arguments{
\item{object}{a VoltRon or vrAssay object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{value}{the name of main image}
}
\description{
Get the main spatial system name
}
