% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{class}
\name{vrMetadata-class}
\alias{vrMetadata-class}
\alias{vrMetadata}
\title{The vrMetadata (VoltRon Metadata) Class}
\description{
The vrMetadata (VoltRon Metadata) Class
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of layers for the this vrSample object}
}}

