% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{vrNeighbourhoodEnrichment}
\alias{vrNeighbourhoodEnrichment}
\title{vrNeighbourhoodEnrichment}
\usage{
vrNeighbourhoodEnrichment(
  object,
  assay = NULL,
  group.by = NULL,
  graph.type = "delaunay",
  num.sim = 1000,
  seed = 1
)
}
\arguments{
\item{object}{a VoltRon object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}

\item{group.by}{a column from metadata to seperate spatial points}

\item{graph.type}{the type of graph to determine spatial neighborhood}

\item{num.sim}{the number of simulations}

\item{seed}{seed}
}
\description{
Conduct Neighborhood enrichment test for pairs of clusters for all assays
}
