% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrNeighbourhoodEnrichmentPlot}
\alias{vrNeighbourhoodEnrichmentPlot}
\title{vrNeighbourhoodEnrichmentPlot}
\usage{
vrNeighbourhoodEnrichmentPlot(results, type = c("assoc", "segreg"))
}
\arguments{
\item{results}{The results from \link{vrNeighbourhoodEnrichment}}

\item{type}{the type of spatial test. Either "assoc" for association test or "segreg" for segregation test}
}
\description{
Plotting results of \link{vrNeighbourhoodEnrichment}
}
