% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrProportionPlot}
\alias{vrProportionPlot}
\title{vrPercentagePlot}
\usage{
vrProportionPlot(object, assay = NULL, x.label = NULL, ...)
}
\arguments{
\item{object}{A VoltRon object}

\item{assay}{assay name}

\item{x.label}{labels of the x axis}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrPercentagePlot
}
