% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrProportionPlot}
\alias{vrProportionPlot}
\title{vrPercentagePlot}
\usage{
vrProportionPlot(
  object,
  assay = NULL,
  x.label = NULL,
  group.by = "Sample",
  ncol = 2,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{A VoltRon object}

\item{assay}{assay name}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrPercentagePlot
}
