% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{vrSample-class}
\alias{vrSample-class}
\alias{vrSample}
\title{The vrSample (VoltRon Sample) Class}
\description{
The vrSample (VoltRon Sample) Class
}
\section{Slots}{

\describe{
\item{\code{layer}}{A list of layers (vrLayer)}

\item{\code{zlocation}}{a vector of z coordinates of layers}

\item{\code{adjacency}}{an adjacency matrix of connected layers within a block}
}}

