% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{vrSample-methods}
\alias{vrSample-methods}
\alias{[[,vrSample,character,ANY-method}
\alias{[[<-,vrSample,character,ANY-method}
\alias{[[,vrBlock,character,ANY-method}
\alias{[[<-,vrBlock,character,ANY-method}
\title{Methods for vrSample objects}
\usage{
\S4method{[[}{vrSample,character,ANY}(x, i)

\S4method{[[}{vrSample,character,ANY}(x, i) <- value

\S4method{[[}{vrBlock,character,ANY}(x, i)

\S4method{[[}{vrBlock,character,ANY}(x, i) <- value
}
\arguments{
\item{x}{A vrSample object}

\item{i}{the name of layer associated with the sample, see \link{SampleMetadata}}

\item{value}{a vrLayer object, see \link{vrLayer}}
}
\description{
Methods for \code{\link{vrSample}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{x = vrSample[[i = character, j = ANY]]}: Accessing vrLayer objects from \code{vrSample} objects

\item \code{x = vrSample[[i = character, j = ANY]] <- value}: Accessing vrLayer objects from \code{vrSample} objects

\item \code{x = vrBlock[[i = character, j = ANY]]}: (deprecated) Accessing vrLayer objects from \code{vrBlock} objects

\item \code{x = vrBlock[[i = character, j = ANY]] <- value}: (deprecated) Overwriting vrLayer objects from \code{vrBlock} objects

}}
\concept{vrsample}
