% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metadata.R
\name{vrSampleNames}
\alias{vrSampleNames}
\alias{vrSampleNames.vrMetadata}
\title{Get Sample names}
\usage{
vrSampleNames(object, ...)

\method{vrSampleNames}{vrMetadata}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Given a vrMetadata object, give names of samples
}
