% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/objects.R
\name{vrSegments.vrAssay}
\alias{vrSegments.vrAssay}
\alias{vrSegments<-.vrAssay}
\alias{vrSegments}
\alias{vrSegments<-}
\alias{vrSegments.vrImage}
\alias{vrSegments<-.vrImage}
\alias{vrSegments.VoltRon}
\alias{vrSegments<-.VoltRon}
\title{vrSegments}
\usage{
\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE)

\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE) <- value

vrSegments(object, assay = NULL, image_name = NULL, reg = FALSE)

vrSegments(object, image_name = NULL, reg = FALSE) <- value

\method{vrSegments}{vrImage}(object)

\method{vrSegments}{vrImage}(object) <- value

\method{vrSegments}{VoltRon}(object, assay = NULL, image_name = NULL, reg = FALSE)

\method{vrSegments}{VoltRon}(object, image_name = NULL, reg = FALSE) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{image_name}{the key of the image associated with the coordinates}

\item{reg}{TRUE if registered coordinates of the main image (\code{vrMainImage}) is requested}

\item{value}{new segment coordinates of spatial points}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \code{SampleMetadata(object)}}
}
\description{
Given a VoltRon, vrAssay or vrImage object, get and set the segment coordinates of assays and coordinate systems
}
