% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/objects.R
\name{vrSegments.vrAssay}
\alias{vrSegments.vrAssay}
\alias{vrSegments<-.vrAssay}
\alias{vrSegments}
\alias{vrSegments<-}
\alias{vrSegments.vrImage}
\alias{vrSegments<-.vrImage}
\alias{vrSegments.VoltRon}
\alias{vrSegments<-.VoltRon}
\title{vrSegments}
\usage{
\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE)

\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE, ...) <- value

vrSegments(object, ...)

vrSegments(object, ...) <- value

\method{vrSegments}{vrImage}(object)

\method{vrSegments}{vrImage}(object, ...) <- value

\method{vrSegments}{VoltRon}(object, assay = NULL, image_name = NULL, reg = FALSE, ...)

\method{vrSegments}{VoltRon}(object, image_name = NULL, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{image_name}{the key of the image associated with the coordinates}

\item{reg}{TRUE if registered segments are being updated}

\item{...}{Arguments passed to other methods}

\item{value}{the new set of 2D segments for each spatial point}

\item{assay}{assay}
}
\description{
Get and set the segments of spatian points of the main assay
}
