% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/assay.R, R/image.R
\name{vrSegments}
\alias{vrSegments}
\alias{vrSegments.VoltRon}
\alias{vrSegments.vrAssay}
\alias{vrSegments.vrImage}
\alias{vrSegments.vrSpatial}
\alias{vrSegments<-.VoltRon}
\alias{vrSegments<-.vrAssay}
\alias{vrSegments<-.vrImage}
\alias{vrSegments<-.vrSpatial}
\title{vrSegments}
\usage{
vrSegments(object, assay = NULL, image_name = NULL, reg = FALSE)

\method{vrSegments}{VoltRon}(object, assay = NULL, image_name = NULL, reg = FALSE)

\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE)

\method{vrSegments}{vrImage}(object)

\method{vrSegments}{vrSpatial}(object)

\method{vrSegments}{VoltRon}(object, image_name = NULL, reg = FALSE) <- value

\method{vrSegments}{vrAssay}(object, image_name = NULL, reg = FALSE) <- value

\method{vrSegments}{vrImage}(object) <- value

\method{vrSegments}{vrSpatial}(object) <- value
}
\arguments{
\item{object}{a VoltRon, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{image_name}{the name/key of the image associated with the coordinates}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainImage}) is requested}

\item{value}{new segment coordinates of spatial points}
}
\description{
Given a VoltRon, vrAssay or vrImage object, get and set the segment coordinates of assays and coordinate systems
}
