% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/objects.R
\name{vrSegments.vrAssay}
\alias{vrSegments.vrAssay}
\alias{vrSegments<-.vrAssay}
\alias{vrSegments}
\alias{vrSegments<-}
\alias{vrSegments.VoltRon}
\alias{vrSegments<-.VoltRon}
\title{vrSegments}
\usage{
\method{vrSegments}{vrAssay}(object, reg = FALSE, ...)

\method{vrSegments}{vrAssay}(object, reg = FALSE, ...) <- value

vrSegments(object, ...)

vrSegments(object, ...) <- value

\method{vrSegments}{VoltRon}(object, reg = FALSE, assay = NULL, ...)

\method{vrSegments}{VoltRon}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{vrSegments}: The name of the default assay

\code{vrSegments<-}: The coordinates updated
}
\description{
Get and set the segments of spatian points of the main assay
}
\concept{data-access}
