% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/objects.R
\name{vrSegments.vrAssay}
\alias{vrSegments.vrAssay}
\alias{vrSegments<-.vrAssay}
\alias{vrSegments}
\alias{vrSegments<-}
\alias{vrSegments.vrImage}
\alias{vrSegments.VoltRon}
\alias{vrSegments<-.VoltRon}
\title{#' @param reg TRUE if registered segments are being updated
#'
#' @rdname vrSegments
#' @method vrSegments vrAssay
#'
#' @export
#'
vrSegments.vrAssay <- function(object, reg = FALSE, ...) {
  if(reg){
    if(length(object@segments_reg) < 1) {
      return(object@segments)
    } else {
      return(object@segments_reg)
    }
  } else {
    return(object@segments)
  }
}}
\usage{
\method{vrSegments}{vrAssay}(object, main_image = NULL, reg = FALSE)

\method{vrSegments}{vrAssay}(object, reg = FALSE, ...) <- value

vrSegments(object, ...)

vrSegments(object, ...) <- value

\method{vrSegments}{vrImage}(object)

\method{vrSegments}{VoltRon}(object, reg = FALSE, assay = NULL, ...)

\method{vrSegments}{VoltRon}(object, reg = FALSE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{main_image}{the key of the image associated with the segments}

\item{reg}{TRUE if registered segments are being updated}

\item{...}{Arguments passed to other methods}

\item{value}{the new set of 2D segments for each spatial point}

\item{assay}{assay}
}
\description{
Get and set the segments of spatian points of the main assay
}
