% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialFeaturePlot}
\alias{vrSpatialFeaturePlot}
\title{vrSpatialFeaturePlot}
\usage{
vrSpatialFeaturePlot(
  object,
  features,
  group.by = "label",
  plot.segments = FALSE,
  n.tile = 0,
  norm = TRUE,
  log = FALSE,
  assay = NULL,
  graph.name = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  cell.shape = 16,
  title.size = 10,
  alpha = 0.6,
  keep.scale = "feature",
  label = FALSE,
  background = NULL,
  reg = FALSE,
  crop = FALSE,
  scale.image = TRUE,
  common.legend = FALSE,
  legend.loc = "right",
  collapse.plots = TRUE
)
}
\arguments{
\item{object}{a VoltRon object}

\item{features}{a set of features to be visualized, either from \link{vrFeatures} of raw or normalized data or columns of the \link{Metadata}.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{plot.segments}{plot segments from \link{vrSegments} instead of points}

\item{n.tile}{should points be aggregated into tiles before visualization (see \link{geom_tile}). Applicable only for cells and molecules}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) will be log transformed}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{graph.name}{if not NULL, the spatial graph is with name \code{graph.name} is visualized as well, see \link{vrGraphNames}}

\item{ncol}{column wise number of plots, for \link{ggarrange}}

\item{nrow}{row wise number of plots, for \link{ggarrange}}

\item{font.size}{font size}

\item{pt.size}{point size}

\item{cell.shape}{the shape of the points representing cells, see \link{geom_point}}

\item{title.size}{title size of legend and plot}

\item{alpha}{alpha level of colors of visualized points and segments}

\item{keep.scale}{whether unify all scales for all features or not}

\item{label}{if TRUE, labels of ROIs will be visualized too}

\item{background}{the background of the plot. Either an image name, see \link{vrImageNames} or a vector of length two with image name 
and a channel name, see \link{vrImageChannelNames}. Type "black" or "white" for black or white backgrounds. if NULL, the main image (\link{vrMainSpatial}) 
and main channel (\link{vrMainChannel}) will be in the background. Otherwise the background will be grey.}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainSpatial}) is requested}

\item{crop}{whether to crop an image of a spot assay to the extend of spots}

\item{scale.image}{if TRUE, background image will be scaled down to a low resolution (width: 1000px)}

\item{common.legend}{whether to use a common legend for all plots, see \link{ggarrange}}

\item{legend.loc}{the location of the legend, default is "right"}

\item{collapse.plots}{whether to combine all ggplots}
}
\description{
Plotting single/multiple features of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
