% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialFeaturePlot}
\alias{vrSpatialFeaturePlot}
\title{vrSpatialFeaturePlot}
\usage{
vrSpatialFeaturePlot(
  object,
  features,
  group.by = "label",
  plot.segments = FALSE,
  norm = TRUE,
  log = FALSE,
  assay = NULL,
  graph.name = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  title.size = 10,
  alpha = 0.6,
  keep.scale = "feature",
  label = FALSE,
  background = NULL,
  reg = FALSE,
  crop = FALSE,
  common.legend = FALSE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{features}{a set of features, either from the rows of rawdata, normdata or columns of the metadata}

\item{group.by}{a grouping label for the spatial entities}

\item{plot.segments}{plot segments instead of points}

\item{norm}{if TRUE, the normalized data is used}

\item{log}{if TRUE, data features (excluding metadata features) will be log transformed}

\item{assay}{the assay name}

\item{graph.name}{if not NULL, the spatial graph is with name \code{graph.name} is visualized as well}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{title.size}{title size of legend and plot}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{keep.scale}{whether unify all scales for all features or not}

\item{label}{if TRUE, labels of ROIs will be visualized too}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{reg}{if TRUE, the registered coordinates will be used}

\item{crop}{whether to crop an image of a spot assay}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting single/multiple features of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
