% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/image.R
\name{vrSpatialNames}
\alias{vrSpatialNames}
\alias{vrSpatialNames,VoltRon-method}
\alias{vrSpatialNames,vrAssay-method}
\alias{vrSpatialNames,vrAssayV2-method}
\title{vrSpatialNames}
\usage{
vrSpatialNames(object, ...)

\S4method{vrSpatialNames}{VoltRon}(object, assay = NULL)

\S4method{vrSpatialNames}{vrAssay}(object)

\S4method{vrSpatialNames}{vrAssayV2}(object)
}
\arguments{
\item{object}{a VoltRon or vrAssay object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), 
see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}. If equals 
to "all", then provides a summary of spatial systems across all assays}
}
\description{
Get names of all spatial systems
}
