% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialPlot}
\alias{vrSpatialPlot}
\title{vrSpatialPlot}
\usage{
vrSpatialPlot(
  object,
  group.by = "Sample",
  plot.segments = FALSE,
  group.ids = NULL,
  colors = NULL,
  n.tile = NULL,
  assay = NULL,
  graph.name = NULL,
  graph.edge.color = "orange",
  reduction = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  cell.shape = 21,
  alpha = 1,
  label = FALSE,
  spatial = NULL,
  channel = NULL,
  background.color = NULL,
  background = NULL,
  reg = FALSE,
  crop = FALSE,
  combine.groups = FALSE,
  legend.pt.size = 2,
  legend.text.size = 14,
  scale.image = TRUE,
  legend.loc = "right",
  common.legend = TRUE,
  collapse.plots = TRUE,
  interactive = FALSE,
  shiny.options = list()
)
}
\arguments{
\item{object}{a VoltRon object}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{plot.segments}{plot segments from \link{vrSegments} instead of points}

\item{group.ids}{a subset of categories defined in metadata column from \code{group.by}}

\item{colors}{the color set for group.by. Should be of the same size of group.id (if specified) or unique elements in group.by}

\item{n.tile}{The number of tiles on x-and y-axis for rasterizing points (see \link[ggplot2]{geom_tile}). The rasterization is performed automatically for large number of points 
Only applicable to spots, cells and molecules. If \code{n.tile = 0} will turn of automated rasterization.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}.
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{graph.name}{if not NULL, the spatial graph is with name \code{graph.name} is visualized as well, see \link{vrGraphNames}}

\item{graph.edge.color}{the colors of the graph edges, if \code{graph.name} is not NULL.}

\item{reduction}{used by \code{vrSpatialPlotVitessce} to visualize an embedding alongside with the spatial plot.}

\item{ncol}{column wise number of plots, for \link[ggpubr]{ggarrange}}

\item{nrow}{row wise number of plots, for \link[ggpubr]{ggarrange}}

\item{font.size}{font size}

\item{pt.size}{point size}

\item{cell.shape}{the shape of the points representing cells, see \link[ggplot2]{geom_point}}

\item{alpha}{alpha level of colors of visualized points and segments}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{spatial}{the name of the spatial coordinate system}

\item{channel}{the name of the channel associated with the image}

\item{background.color}{the color of plot background if a channel is not specified, or the spatial coord system doesnt have an image.}

\item{background}{(DEPRECATED) the background of the plot. Either an image name, see \link{vrImageNames} or a vector of length two with image name
and a channel name, see \link{vrImageChannelNames}. Type "black" or "white" for black or white backgrounds. if NULL, the main image (\link{vrMainSpatial})
and main channel (\link{vrMainChannel}) will be in the background. Otherwise the background will be grey.}

\item{reg}{TRUE if registered coordinates of the main image (\link{vrMainSpatial}) is requested}

\item{crop}{whether to crop an image of a spot assay to the extend of spots}

\item{combine.groups}{if TRUE, tile colors will reflect relative abundance of either of two groups, strictly for visualizing two groups when assay is a molecule typed and tiled (see \code{n.tile}).}

\item{legend.pt.size}{the size of points at the legend}

\item{legend.text.size}{the size of the text at the legend}

\item{scale.image}{if TRUE, background image will be scaled down to a low resolution (width: 1000px)}

\item{legend.loc}{the location of the legend, default is "right"}

\item{common.legend}{whether to use a common legend for all plots, see \link[ggpubr]{ggarrange}}

\item{collapse.plots}{whether to combine all ggplots}

\item{interactive}{if TRUE, run interactive plot}

\item{shiny.options}{a list of shiny options (browser, host, port etc.) passed \code{options} argument of \link[shiny]{shinyApp}. For more information, see \link[shiny]{runApp}}
}
\description{
Plotting identification of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
