% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialPlot}
\alias{vrSpatialPlot}
\title{vrSpatialPlot}
\usage{
vrSpatialPlot(
  object,
  group.by = "Sample",
  transcripts = NULL,
  assay = NULL,
  assay.type = NULL,
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  alpha = 1,
  label = FALSE,
  background = NULL,
  crop = FALSE,
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{group.by}{a grouping label for the spatial entities}

\item{transcripts}{Only for spatial data with existing subcellular information, visualize the location of given transcripts}

\item{assay}{the assay name}

\item{assay.type}{the assay type name: 'cell', 'spot' or 'ROI'}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{crop}{whether to crop an image of a spot assay}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting identification of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
