% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialPlot}
\alias{vrSpatialPlot}
\title{vrSpatialPlot}
\usage{
vrSpatialPlot(
  object,
  group.by = "Sample",
  plot.segments = FALSE,
  group.ids = NULL,
  assay = NULL,
  graph.name = NULL,
  reduction = "umap",
  ncol = 2,
  nrow = NULL,
  font.size = 2,
  pt.size = 2,
  cell.shape = 21,
  alpha = 1,
  label = FALSE,
  background = NULL,
  reg = FALSE,
  crop = FALSE,
  legend.pt.size = 2,
  legend.loc = "right",
  common.legend = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{object}{VoltRon object}

\item{group.by}{a grouping label for the spatial entities}

\item{plot.segments}{plot segments instead of points}

\item{group.ids}{a subset of categories defined with in the grouping label \code{group.by}}

\item{assay}{the assay name}

\item{graph.name}{if not NULL, the spatial graph is with name \code{graph.name} is visualized as well}

\item{reduction}{Used by \code{vrSpatialPlotInteractive}, to visualize an embedding alongside with the spatial plot.}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{cell.shape}{the shape of the points representing cells, see \code{help(geom_point)}}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{label}{if TRUE, the labels of the ROI assays will be visualized}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{reg}{if TRUE, the registered coordinates will be used}

\item{crop}{whether to crop an image of a spot assay}

\item{legend.pt.size}{the size of points at the legend}

\item{legend.loc}{the location of the legend, default is "right"}

\item{common.legend}{whether to use a common legend for all plots}

\item{collapse}{whether to combine all ggplots}
}
\description{
Plotting identification of spatially resolved cells, spots, and ROI on associated images from multiple assays in a VoltRon object.
}
