% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrSpatialPlotSingle}
\alias{vrSpatialPlotSingle}
\title{vrSpatialPlotSingle}
\usage{
vrSpatialPlotSingle(
  assay,
  metadata,
  group.by = "Sample",
  font.size = 2,
  pt.size = 2,
  alpha = 0.6,
  plot_title = NULL,
  background = "image",
  crop = FALSE
)
}
\arguments{
\item{assay}{vrAssay object}

\item{metadata}{the metadata associated with the assay}

\item{group.by}{a grouping label for the spatial entities}

\item{font.size}{font sizes}

\item{pt.size}{point size}

\item{alpha}{alpha level for cells/spots/ROIs}

\item{plot_title}{the title of the single plot}

\item{background}{the background of the plot, either "image" for overlaying the image of the assay, or "black" or "white" background (suitable for IF based assays)}

\item{crop}{whether to crop an image of a spot assay}
}
\description{
Plotting a single assay from a VoltRon object. We plot the identification of spatially resolved cells, spots, and ROI on associated images.
}
