% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R, R/objects.R, R/metadata.R,
%   R/assay.R, R/image.R, R/sample.R
\name{vrSpatialPoints}
\alias{vrSpatialPoints}
\alias{vrSpatialPoints,VoltRon-method}
\alias{vrSpatialPoints,vrMetadata-method}
\alias{vrSpatialPoints,vrAssay-method}
\alias{vrSpatialPoints,vrAssayV2-method}
\alias{vrSpatialPoints,vrImage-method}
\alias{vrSpatialPoints,vrSpatial-method}
\alias{vrSpatialPoints,vrSample-method}
\alias{vrSpatialPoints,vrBlock-method}
\alias{vrSpatialPoints,vrLayer-method}
\alias{vrSpatialPoints<-,vrAssay-method}
\alias{vrSpatialPoints<-,vrAssayV2-method}
\alias{vrSpatialPoints<-,vrImage-method}
\alias{vrSpatialPoints<-,vrSpatial-method}
\alias{vrSpatialPoints<-}
\title{vrSpatialPoints}
\usage{
vrSpatialPoints(object, ...)

\S4method{vrSpatialPoints}{VoltRon}(object, assay = NULL)

\S4method{vrSpatialPoints}{vrMetadata}(object, assay = NULL)

\S4method{vrSpatialPoints}{vrAssay}(object)

\S4method{vrSpatialPoints}{vrAssayV2}(object)

\S4method{vrSpatialPoints}{vrImage}(object)

\S4method{vrSpatialPoints}{vrSpatial}(object)

\S4method{vrSpatialPoints}{vrSample}(object)

\S4method{vrSpatialPoints}{vrBlock}(object)

\S4method{vrSpatialPoints}{vrLayer}(object)

\S4method{vrSpatialPoints}{vrAssay}(object) <- value

\S4method{vrSpatialPoints}{vrAssayV2}(object) <- value

\S4method{vrSpatialPoints}{vrImage}(object) <- value

\S4method{vrSpatialPoints}{vrSpatial}(object) <- value

vrSpatialPoints(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon, vrSample, vrLayer, vrAssay or vrSpatial object.}

\item{...}{arguments passed to other methods.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{value}{names for spatial points}
}
\description{
Get and set spatial entities.
}
\examples{
vrSpatialPoints(visium_data)
vrSpatialPoints(visium_data, assay = "Visium")
vrSpatialPoints(visium_data, assay = "Assay1")
}
