% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R, R/metadata.R,
%   R/assay.R, R/image.R, R/sample.R
\name{vrSpatialPoints}
\alias{vrSpatialPoints}
\alias{vrSpatialPoints.VoltRon}
\alias{vrSpatialPoints.vrMetadata}
\alias{vrSpatialPoints.vrAssay}
\alias{vrSpatialPoints.vrImage}
\alias{vrSpatialPoints.vrSpatial}
\alias{vrSpatialPoints.vrSample}
\alias{vrSpatialPoints.vrBlock}
\alias{vrSpatialPoints.vrLayer}
\alias{vrSpatialPoints<-.vrAssay}
\alias{vrSpatialPoints<-.vrImage}
\alias{vrSpatialPoints<-.vrSpatial}
\title{vrSpatialPoints}
\usage{
vrSpatialPoints(object, assay = NULL)

\method{vrSpatialPoints}{VoltRon}(object, assay = NULL)

\method{vrSpatialPoints}{vrMetadata}(object, assay = NULL)

\method{vrSpatialPoints}{vrAssay}(object)

\method{vrSpatialPoints}{vrImage}(object)

\method{vrSpatialPoints}{vrSpatial}(object)

\method{vrSpatialPoints}{vrSample}(object, ...)

\method{vrSpatialPoints}{vrBlock}(object, ...)

\method{vrSpatialPoints}{vrLayer}(object, ...)

\method{vrSpatialPoints}{vrAssay}(object) <- value

\method{vrSpatialPoints}{vrImage}(object, ...) <- value

\method{vrSpatialPoints}{vrSpatial}(object, ...) <- value
}
\arguments{
\item{object}{a VoltRon, vrSample, vrLayer, vrAssay or vrImage object}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{...}{arguments passed to other methods}

\item{value}{new spatial points}
}
\description{
Get and set spatial entities.
}
\examples{
vrSpatialPoints(visium_data)
vrSpatialPoints(visium_data, assay = "Visium")
vrSpatialPoints(visium_data, assay = "Assay1")
}
