% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/generics.R, R/image.R, R/metadata.R,
%   R/objects.R, R/sample.R
\name{vrSpatialPoints.vrAssay}
\alias{vrSpatialPoints.vrAssay}
\alias{vrSpatialPoints<-.vrAssay}
\alias{vrSpatialPoints}
\alias{vrSpatialPoints<-}
\alias{vrSpatialPoints.vrImage}
\alias{vrSpatialPoints<-.vrImage}
\alias{vrSpatialPoints.vrMetadata}
\alias{vrSpatialPoints.VoltRon}
\alias{vrSpatialPoints.vrSample}
\alias{vrSpatialPoints.vrLayer}
\title{vrSpatialPoints}
\usage{
\method{vrSpatialPoints}{vrAssay}(object, ...)

\method{vrSpatialPoints}{vrAssay}(object, ...) <- value

vrSpatialPoints(object, ...)

vrSpatialPoints(object, ...) <- value

\method{vrSpatialPoints}{vrImage}(object, ...)

\method{vrSpatialPoints}{vrImage}(object, ...) <- value

\method{vrSpatialPoints}{vrMetadata}(object, assay = NULL)

\method{vrSpatialPoints}{VoltRon}(object, ...)

\method{vrSpatialPoints}{vrSample}(object, ...)

\method{vrSpatialPoints}{vrLayer}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{new spatial points}

\item{assay}{the assay name or type}
}
\description{
Get and set spatial entities (cells, spots, ROI)
}
