% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrViolinPlot}
\alias{vrViolinPlot}
\title{vrViolinPlot}
\usage{
vrViolinPlot(
  object,
  assay = NULL,
  assay.type = NULL,
  features = NULL,
  group.by = "clusters",
  norm = TRUE,
  ...
)
}
\arguments{
\item{object}{A VoltRon object}

\item{assay}{assay name}

\item{assay.type}{assay type}

\item{features}{a set of features to be visualized}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{norm}{if TRUE, the normalized data is used}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrViolinPlot
}
