% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrViolinPlot}
\alias{vrViolinPlot}
\title{vrViolinPlot}
\usage{
vrViolinPlot(
  object,
  features = NULL,
  assay = NULL,
  group.by = "Sample",
  norm = TRUE,
  points = TRUE,
  ncol = 2,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{A VoltRon object}

\item{features}{a set of features to be visualized}

\item{assay}{assay name}

\item{group.by}{a column from metadata to seperate columns of the heatmap}

\item{norm}{if TRUE, the normalized data is used}

\item{points}{if TRUE, measures are visualized as points as well.}

\item{ncol}{column wise number of plots, for \code{ggarrange}}

\item{nrow}{row wise number of plots, for \code{ggarrange}}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrViolinPlot
}
