% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vrViolinPlot}
\alias{vrViolinPlot}
\title{vrViolinPlot}
\usage{
vrViolinPlot(
  object,
  features = NULL,
  assay = NULL,
  group.by = "Sample",
  norm = TRUE,
  points = TRUE,
  ncol = 2,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{A VoltRon object}

\item{features}{a set of features to be visualized, either from \link{vrFeatures} of raw or normalized data or columns of the \link{Metadata}.}

\item{assay}{assay name (exp: Assay1) or assay class (exp: Visium, Xenium), see \link{SampleMetadata}. 
if NULL, the default assay will be used, see \link{vrMainAssay}.}

\item{group.by}{a column of metadata from \link{Metadata} used as grouping label for the spatial entities}

\item{norm}{if TRUE, the normalized data is used}

\item{points}{if TRUE, measures are visualized as points as well.}

\item{ncol}{column wise number of plots, for \link{ggarrange}}

\item{nrow}{row wise number of plots, for \link{ggarrange}}

\item{...}{additional parameters passed to \code{getVariableFeatures}}
}
\description{
vrViolinPlot
}
