% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{warpImage}
\alias{warpImage}
\title{warpImage}
\usage{
warpImage(ref_image, query_image, mapping)
}
\arguments{
\item{ref_image}{reference image}

\item{query_image}{query image}

\item{mapping}{a list of the homography matrices and TPS keypoints}
}
\description{
Warping a query image given a homography image
}
